### 2020-09-01 #2617 @pibion
@kthyng I'm not super experienced with R, but this is a topic I'm comfortable with.  I'll go ahead and accept.
