### 2020-09-02 #2617 @wesleyburr
> @wesleyburr I've invited your suggested reviewers and I'll post in this space as they accept or decline.
>
> It's taking a little while longer than usual both to find reviewers and for them to review, but we'll try to keep the process moving along :)

No worries - I read the announcement about the delays. The fact that you folks are working as fast as you are in this "time of COVID" is remarkable, and we're thankful for any review at any pace folks can handle.
