### 2020-09-13 #2617 @yguitton
Hi,

 I'm not a MALDI expert, si i will wait for an other opportunity to review
an other paper.

Regards

Le jeu. 10 sept. 2020 20:02, pibion <notifications@github.com> a écrit :

> 👋 @yguitton <https://github.com/yguitton> & @biswapriyamisra
> <https://github.com/biswapriyamisra>, would any of you be willing to
> review this submission for JOSS? We carry out our checklist-driven reviews
> here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
