### 2020-09-22 #2617 @wesleyburr
> @wesleyburr apologies, it's taking some time to find a third reviewer. I'm working on it, but in the meantime, how do you feel about two reviewers for this project?
>
> I prefer having three reviewers, but our current reviewers have good expertise with this subject. I'd be okay with moving ahead with only two in this case. How do you feel?

I don't mind at all - 3 is certainly better for perspectives, but if the 2 reviewers who agreed are ok with it, I'm happy to work with them to make the process run as smoothly as we can!
