### 2020-09-24 #2694 @whedon
[REVIEW]: subMALDI: an open framework R package for processing irregularly-spaced mass spectrometry data
**Submitting author:** @wesleyburr (<a href="http://orcid.org/0000-0002-2058-1899">Wesley Burr</a>)
**Repository:** <a href="https://github.com/wesleyburr/subMaldi" target ="_blank">https://github.com/wesleyburr/subMaldi</a>
**Version:** v1.0-5

- [x] Test that I can make edits
