### 2020-10-02 #2694 @pibion
@jspaezp, @sigven, could you confirm that you can check and uncheck the boxes at the top of this issue?
