### 2020-10-05 d1af868
Update DESCRIPTION
Some platforms don't read the line breaks right, and don't install the suggests as needed. Found while recreating vignettes.


Former-commit-id: 83bdc0671bfa74a8ff3479dc5ab901c29d2963d4
Former-commit-id: c04966c3fab7d7acfe462d051b6e1caa73097d89
