### 2020-10-05 8de19dc
Update DESCRIPTION
Some platforms don't read the line breaks right, and don't install the suggests as needed. Found while recreating vignettes.
