+++ DESCRIPTION
@@ -173,6 +173,7 @@ LazyData: true
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 Suggests: knitr, rmarkdown, testthat
 VignetteBuilder: knitr
 RoxygenNote: 7.1.1
@@ -283,6 +284,9 @@ VignetteBuilder: rmarkdown
     rmarkdown,
     testthat
 >>>>>>> 40a45a4... Added tests
+=======
+Suggests: knitr, rmarkdown, testthat
+>>>>>>> 558f3f0... Update DESCRIPTION
 VignetteBuilder: knitr
 >>>>>>> de15a92... Fixed packaged structure
 RoxygenNote: 7.1.1
