+++ doc/subMALDIprocessing.R
@@ -1,0 +1,156 @@
+## ---- include = FALSE---------------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+
+## -----------------------------------------------------------------------------
+library("subMALDI")
+
+
+data("bsline")
+
+## ----warning = FALSE, fig.width = 6, fig.height = 3, fig.align = "center"-----
+plotSpectrum(dat = bsline, mass_dat = bsline$mass, intensity_dat = bsline$raw,
+             colour = "black", min_mz = 200, max_mz = 5000, x_ticks = 500)
+
+## -----------------------------------------------------------------------------
+sgolay <- smoothSpectrum(dat = bsline, mass_dat = "mass", intensity_dat = "raw",
+                       method = "sgolay", p = 4, n = 7, m = 0, ts = 1)
+
+## ----warning = FALSE, fig.width = 6, fig.height = 4, fig.align = "center"-----
+test <- data.frame(bsline$mass, bsline$raw, sgolay$sg)
+names(test) <- c("mass", "Raw Spectrum", "Savitzky-Golay Smooth")
+
+plotSpectra(dat = test, mass_dat = "mass", spec1 = "Raw Spectrum", 
+            spec2 = "Savitzky-Golay Smooth", min_mz = 200, max_mz = 5000,
+            x_ticks = 500)
+
+## -----------------------------------------------------------------------------
+mov_avg <- smoothSpectrum(dat = bsline, mass_dat = bsline$mass, intensity_dat = bsline$raw,
+                       method = "mov_avg", n = 7)
+
+## -----------------------------------------------------------------------------
+test <- createSpecDF(min_mz = 200, max_mz = 5000)
+test <- select(test, full_mz)
+test <- transform(test, "RawSpectrum" = 0, "MovingAverage" = 0)
+test <- mapSpectrum(bsline, massCol = "mass", intenseCol = "raw", 
+                 colName = "RawSpectrum", spec_df = test)
+test <- mapSpectrum(mov_avg, massCol = "mass", intenseCol = "mov_avg", 
+                 colName = "MovingAverage", spec_df = test)
+test <- rmveEmpty(test)
+
+## ----warning = FALSE, fig.width = 6, fig.height = 4, fig.align = "center"-----
+plotSpectra(dat = test, mass_dat = "full_mz", spec1 = "RawSpectrum", 
+            spec2 = "MovingAverage", min_mz = 200, max_mz = 5000,
+            x_ticks = 500)
+
+## -----------------------------------------------------------------------------
+monotone_min <- baselineCorr(dat = sgolay, mass_dat = "mass", 
+                             intensity_dat = "sg", method = "monotone_min")
+
+## ----warning = FALSE, fig.width = 6, fig.height = 4, fig.align = "center"-----
+test <- data.frame(sgolay$mass, sgolay$sg, monotone_min$baseline)
+names(test) <- c("mz", "Smoothed", "BaselineCorrected")
+
+plotSpectra(dat = test, mass_dat = "mz", spec1 = "Smoothed", 
+            spec2 = "BaselineCorrected", min_mz = 200, max_mz = 5000,
+            x_ticks = 500)
+
+## -----------------------------------------------------------------------------
+linear_int <- baselineCorr(dat = mov_avg, mass_dat = "mass", 
+                           intensity_dat = "mov_avg", method = "linear", n = 7)
+
+## ----warning = FALSE, fig.width = 6, fig.height = 4, fig.align = "center"-----
+test <- data.frame(mov_avg$mass, mov_avg$mov_avg, linear_int$baseline)
+names(test) <- c("mz", "Smoothed", "BaselineCorrected")
+
+plotSpectra(dat = test, mass_dat = "mz", spec1 = "Smoothed", 
[…]
