+++ doc/subMALDIprocessing.Rmd
@@ -1,0 +1,474 @@
+---
+title: "Processing with subMALDI"
+author: Kristen Yeh^1^ and Wesley Burr ^2^
+output: 
+  rmarkdown::html_vignette:
+    toc: true
+    toc_depth: 1
+    number_sections: true
+vignette: >
+  %\VignetteIndexEntry{Processing with subMALDI}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+
+^1^ Forensic Science Program, Trent University, Peterborough, ON, Canada
+
+^2^ Faculty of Science, Mathematics, Trent University, Peterborough, ON, Canada
+
+# Introduction
+
+In this vignette we demonstrate a pre-processing workflow for smoothing, baseline correction, and peak detection of raw, irregularly-spaced mass spectrometry data using `subMALDI`. This package is freely available from [GitHub](https://github.com/wesleyburr/subMaldi) and was created using MALDI FT-ICR mass spectrometry data. 
+
+It is recommended that users first read through the *"`subMALDI` Workflow"* vignette before proceeding with pre-processing functions. This will assist users in understanding the basic functions included in `subMALDI`, including import, plotting, and mapping functions for comparison of multiple spectra.
+
+## Loading Sample Data
+
+We'll begin this vignette by installing the package and loading a sample data set included with the package, called `"bsline.rda"`.
+
+
+```{r}
+library("subMALDI")
+
+
+data("bsline")
+```
+
+
+Although most modern-day high resolution mass spectrometry instruments come with their own spectral processing software which are capable of automatic baseline correction, this software can fail, producing a spectrum with an uneven baseline. 
+
+Our data set `bsline` is an example of this scenario. A plot of the raw spectrum is shown below.
+
+
+```{r warning = FALSE, fig.width = 6, fig.height = 3, fig.align = "center"}
+plotSpectrum(dat = bsline, mass_dat = bsline$mass, intensity_dat = bsline$raw,
+             colour = "black", min_mz = 200, max_mz = 5000, x_ticks = 500)
+```
+
+
+Observing this plot, we can clearly see the baseline of the spectrum is increasing as the x-axis increases, with the exception of a nearly Gaussian-shaped curve from the *m/z* range 700 to 1700. Before we can extract meaningful data from this raw spectrum, this irregular baseline must be corrected.
+
+We will work through the pre-processing and baseline correction below, demonstrating several different methods for each processing step.
+
+---
+
+# Smoothing Filters
+
+As the sensitivity of high-resolution mass spectrometry instruments increases, so does the amount of detectable noise. This produces spectra which appear thicker around the baseline than at high intensity peaks, without proper spectral processing. We can see this phenomenon in our sample data set `bsline` from the *m/z* range 200 to 1700. Here, the baseline of the spectrum is very thick relative to that at higher *m/z* values and high intensity peaks. 
+
+In order to minimize the noise in our spectrum and create a more uniform signal throughout the analyzed *m/z* range, we will use the `subMALDI` function `smoothSpectrum`. This function offers two different filters for spectral smoothing: the Savitzky-Golay and moving average filters.
+
+## Savitzky-Golay
+
+The Savitzy-Golay filter is a least-squares smoothing method [1]. It performs a least squares fit of a small set of consecutive data points to a polynomial and takes the central point of the fitted polynomial curve as the output [1,2]. 
+
[…]
