+++ doc/subMALDIworkflow.Rmd
@@ -1,0 +1,535 @@
+---
+title: "subMALDI Workflow"
+author: Kristen Yeh^1^ and Wesley Burr^2^
+output: 
+  rmarkdown::html_vignette:
+    toc: true
+    toc_depth: 1
+    df_print: kable
+    number_sections: true
+vignette: >
+  %\VignetteIndexEntry{subMALDI Workflow}
+  %\VignetteEngine{knitr::rmarkdown}
+  \usepackage[utf8]{inputenc}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+^1^ Forensic Science Program, Trent University, Peterborough, ON, Canada
+
+^2^ Faculty of Science, Mathematics, Trent University, Peterborough, ON, Canada
+
+# Introduction
+
+In this vignette we demonstrate a processing workflow for irregularly-spaced mass spectrometry data using `subMALDI`. This package is freely available from [GitHub](https://github.com/wesleyburr/subMaldi) and was created using MALDI FT-ICR mass spectrometry data. 
+
+---
+
+
+# Installation and Import
+
+## Installation
+
+To download the package, select the '.tar.gz' file from the git repository. This file can then used to install the package in R. Below, we load the package and its dependencies.
+
+```{r setup}
+library(subMALDI)
+```
+
+## Import
+Two import functions are included in the package. Both read .CSV files which contain spectral data from a single spectrum per file. While `readcsvSpec()` imports one spectrum at a time and turns them into a pairwise data frame of *m/z* and intensity data, `readcsvDir()` is capable of reading all the .CSV files in a folder. The pairwise data frames generated for all of these spectra are then output as .rda files into a designated output folder.
+
+For example, let's say we have a set of 5 spectral replicates in our working directory. These are stored as .CSV files in a folder called "my_spec". In the original file, mass data is stored in a column called "m.z", while intensity data is stored in a column called "I". If we wanted to import only one of the spectra, specifically `"Sample1.csv"`, then we wound run the following code. 
+
+```{r eval = FALSE}
+my_spec1 <- readcsvSpec(spec_file = "~/my_spec/Sample1.csv", massCol = "m.z", intenseCol = "I")
+```
+
+Conversely, if we wanted to import all spectra in the folder "my_spec" into a new folder called "out", we would run the following.
+
+```{r eval = FALSE}
+readcsvDir(direct = "~/my_spec/", massCol = "m.z", intenseCol = "I", output = "~/my_spec/out/")
+
+load("~/my_spec/out/Sample1.rda")
+```
+
+This will create a pairwise data frame of *m/z* and intensity data, much like the one shown below.
+
+```{r }
+data("Blank1")
+head(Blank1)
+```
+
+
+---
+
+
+# Creating a Mapped Data Frame
+
+In order to compare unevenly spaced spectral data, it is essential that the spectra be standardized by some means. When performing R analysis, replicates must be aligned against the same data structure, requiring that each spectrum has a consistent number of data points. 
+
+## Create an Empty Data Frame
+
+`subMALDI` permits the comparison of unevenly spaced mass spectrometry data by mapping each peak in a spectrum to all of the theoretically possible *m/z* values in the analyzed mass range.
+
[…]
