+++ doc/subMALDIworkflow.R
@@ -1,0 +1,222 @@
+## ---- include = FALSE---------------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+
+## ----setup--------------------------------------------------------------------
+library(subMALDI)
+
+## ----eval = FALSE-------------------------------------------------------------
+#  my_spec1 <- readcsvSpec(spec_file = "~/my_spec/Sample1.csv", massCol = "m.z", intenseCol = "I")
+
+## ----eval = FALSE-------------------------------------------------------------
+#  readcsvDir(direct = "~/my_spec/", massCol = "m.z", intenseCol = "I", output = "~/my_spec/out/")
+#  
+#  load("~/my_spec/out/Sample1.rda")
+
+## -----------------------------------------------------------------------------
+data("Blank1")
+head(Blank1)
+
+## -----------------------------------------------------------------------------
+spec_df <- createSpecDF(min_mz = 53.76, max_mz = 1100, res = 1e-04, dig = 4)
+
+head(spec_df)
+
+## -----------------------------------------------------------------------------
+spec_df <- select(spec_df, "full_mz")
+spec_df <- transform(spec_df, "Blank1" = 0, "Blank2" = 0, "Before1" = 0, "Before2" = 0)
+
+head(spec_df)
+
+## -----------------------------------------------------------------------------
+data("Blank1")
+data("Blank2")
+data("Before1")
+data("Before2")
+
+## -----------------------------------------------------------------------------
+spec_df <- mapSpectrum(dat = Blank1, massCol = "mass", intenseCol = "Intensity", 
+            spec_df = spec_df, colName = "Blank1", thresh = 1e-04, dig = 4)
+spec_df <- mapSpectrum(dat = Blank2, massCol = "mass", intenseCol = "Intensity", 
+            spec_df = spec_df, colName = "Blank2", thresh = 1e-04, dig = 4)
+spec_df <- mapSpectrum(dat = Before1, massCol = "mass", intenseCol = "Intensity", 
+            spec_df = spec_df, colName = "Before1", thresh = 1e-04, dig = 4)
+spec_df <- mapSpectrum(dat = Before2, massCol = "mass", intenseCol = "Intensity", 
+            spec_df = spec_df, colName = "Before2", thresh = 1e-04, dig = 4)
+
+head(spec_df)
+
+## -----------------------------------------------------------------------------
+spec <- rmveEmpty(spec_df)
+
+head(spec)
+nrow(spec)
+
+## -----------------------------------------------------------------------------
+avg_spec <- avgSpectra(spec, method = "mean", spec1 = "Blank1", spec2 = "Blank2")
+
+head(avg_spec)
+
+## -----------------------------------------------------------------------------
+spec <- select(avg_spec, "full_mz", "Before1", "Before2", "Average")
+sub_spec <- transform(spec, "Sub_Before1" = 0, "Sub_Before2" = 0)
+
+## -----------------------------------------------------------------------------
+sub_spec <- subSpectra(dat = sub_spec, Blank_Var = "Average", 
+                       Sample = "Before1", Sub_Sample = "Sub_Before1")
+sub_spec <- subSpectra(dat = sub_spec, Blank_Var = "Average", 
+                       Sample = "Before2", Sub_Sample = "Sub_Before2")
+
+head(sub_spec)
+
[…]
