### 2020-10-06 wesleyburr/subMaldi#5 @jspaezp
[JOSS REVIEW]: subMALDI - Jspaezp
https://github.com/openjournals/joss-reviews/issues/2694#issuecomment-703271864
Sorry for the late review but here are some of the notes I have on the package.

  - Yes, GPL-V3
  - Kristen Yeh seems to be the main code contributor to the repository.

  - See **Installation instructions:** under **Documentation**

  - Despite the fact that I understand why this would predominantly
    favor MALDI data over any other type of ionization method, It seems to be
    equally applicable to any other of "single spectra" acquisition method.
    Therefore I believe it should be explicit in the paper the assumptions
    in the data and not only the ionization method.
    (centroided/profile "single spectra" and column order/name specifications).
  - In addition, an example on how to get the text-format spectra files from
    mass spectrometer raw format would be useful (sample acquisition/export settings).
  - It is found in the vignettes and a `tar` is contained in the github
    repository. There is no mention on how to install it in the `README`.
  - In addition, having instructions on how to install from source code
    could be useful (`remotes::install_github`, `devtools::install` and
    `BiocManager::install` are easy alternatives). **EDIT 20210317: DONE**
  - I believe "bundled" packages (.tar packages) should be moved to the releases
    section of the repository. Since these files should not be subject to
    version control (which would also greatly reduce the size of the repository
    when cloning it).
    https://docs.github.com/en/free-pro-team@latest/github/administering-a-repository/managing-releases-in-a-repository **EDIT 20210317: Please update the latest release, latest one was built on Aug 30**
  - Adding the `biocViews:` line in the `DESCRIPTION` file makes dependencies
    be searched in bioconductor. (installation fails without it because the
    package `MassSpecWavelet` is not in CRAN)
    https://github.com/r-lib/devtools/issues/700#issuecomment-235127291 **EDIT 20210317: DONE**
  - Contained in the vignettes, further specification about where in the full workflow this software fits would be useful (after peak-picking and centroiding, before differential "expression" analysis)
  - I notice a `man` directory with build documentation. Which seems to be
    complete and well written. Nonetheless, I highly encourage the authors to
    migrate to automatically generated documentation using `roxygen2` (not
    subject to paper approval). This migration would make it really
    straightforward to make documentation more accessible with an automatically
    generated documentation website (https://pkgdown.r-lib.org/) **EDIT 20210317: DONE, I would suggest organizing the data and functions separately in the reference menu but looks good to me**
  - Consider continuous integration method to run tests automatically on commits.
  - Some functions that are key to the features of the package are not tested.
    Reported Coverage produced by `covr::package_coverage(quiet = FALSE)`

```
subMALDI Coverage: 9.90%
R/find_max.R: 0.00%
R/normMethod_quantile.R: 0.00%
R/peak_sub.R: 0.00%
R/plot_lbl.R: 0.00%
R/plot_spec.R: 0.00%
R/smooth_sub.R: 0.00%
R/smoothSpectrum.R: 0.00%
R/normMethod_custimp.R: 4.90%
R/normMethod_TIC.R: 5.23%
R/normMethod_median.R: 6.49%
R/normMethod_custom.R: 6.79%
R/normMethod_stdev.R: 10.16%
R/normMethod_max.R: 11.92%
R/normMethod_RMS.R: 15.09%
R/avgSpectra.R: 27.10%
R/peakDet.R: 66.67%
R/subMALDI.R: 68.85%
R/normSpectra.R: 87.50%
R/baseline_loess.R: 100.00%
R/baseline_sub.R: 100.00%
R/baselineCorr.R: 100.00%
```

- [x] **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.
  - Please add contributions guidelines, and ideally a link to it in the `README`. **EDIT 20210317: DONE**

  - A schematic showing some of the processing tools would be useful here to
    depict the processing functions provided. (the same plots shown in
    the vignettes could be used).
  - When mentioning `It has been utilized in a scientific article accepted into the *Journal of Forensic Chemistry* (FORC-D-20-00095)` a reference to the paper would be required. **EDIT 20210317: DONE (really well)**

**Final remarks**
- I believe that some clenaup of the git repository is required, since right
  now the size of a `git clone` is 780mb (the probable cause is having data
  or compressed files version controlled). Not critical but highly encouraged. **EDIT 20210317: DONE**

- The code is easily readable and easy to browse, nonetheless the test code base
  does not share the same quality standard, please consider looking into this.
  An automated code formatting tool could help in this regard
  (for instance: https://github.com/r-lib/styler). **EDIT 20210317: DONE, there is still some formatting that is not consistent with the rest of the code-base but nothing that raises a red flag**

- Package size can be reduced by compressing the data files,
  output from package check:

```
    Warning: large data files saved inefficiently:
                  size ASCII compress
    Master.rda  30.5Mb FALSE     none
    Master2.rda  4.2Mb FALSE     none
    bsline.rda   1.2Mb FALSE     none

    Note: significantly better compression could be obtained
          by using R CMD build --resave-data
                old_size new_size compress
    After1.rda     1.9Mb    682Kb       xz
    After2.rda     2.5Mb    897Kb       xz
    Before1.rda    6.3Mb    2.1Mb       xz
    Before2.rda    3.0Mb   1016Kb       xz
    Blank1.rda     1.6Mb    521Kb       xz
    Blank2.rda     4.0Mb    1.3Mb       xz
    Master.rda    30.5Mb    4.1Mb       xz
    Master2.rda    4.2Mb    892Kb       xz
    bsline.rda     1.2Mb    135Kb       xz
```
**EDIT 20210317: DONE**

- There are common patterns in the code base that can be generalized easily. **EDIT 20210317: DONE (really well)**

```r
...
if(is.null(spec3)){

  if(is.null(spec2)){

  # Single spectrum
  mass <- dat[[mass_dat]]
  intense <- dat[[spec1]]

  max_Intensity <- max(intense)

  dat <- select(dat, all_of(mass_dat))
  dat <- transform(dat, "Normalized" = 0)
  dat["Normalized"] <- .normalize(intense)
  return(dat)
  }

  # Two spectra
  else{
    mass <- dat[[mass_dat]]
    intense1 <- dat[[spec1]]
    intense2 <- dat[[spec2]]

    max_Intensity1 <- max(intense1)
    max_Intensity2 <- max(intense2)

    dat <- select(dat, all_of(mass_dat))
    dat[spec1] <- .normalize(intense1)
    dat[spec2] <- .normalize(intense2)
    return(dat)
  }}

  # Three spectra
  else{
    mass <- dat[[mass_dat]]
    intense1 <- dat[[spec1]]
    intense2 <- dat[[spec2]]
    intense3 <- dat[[spec3]]

    max_Intensity1 <- max(intense1)
    max_Intensity2 <- max(intense2)
    max_Intensity3 <- max(intense3)

    dat <- select(dat, all_of(mass_dat))
    dat[spec1] <- .normalize(intense1)
    dat[spec2] <- .normalize(intense2)
    dat[spec3] <- .normalize(intense3)
    return(dat)
  }}
  ...
```

can be replaced with something like so:

```r
spectra_cols <- c(spec1, spec2, spec3)

xxx <- function(..., mass_dat, spectra_cols) {
    mass <- dat[[mass_dat]]
    intensities <- lapply(spectra_cols, function(x){ dat[[x]] })
    intensities <- lapply(intensities, .normalize)

    dat <- do.call(what = data.frame, args = c(list(mass), intensities))
    colnames(dat) <- c(mass_dat, spectra_cols)
    return(dat)
}
```

which would provide for free the possibility to use the function with an
arbitrary number of spectra (given memory restrictions). In addition to ease in
extensibility and "maintainability".

Let me know if any clarification is needed for any of the mentioned points.
Kindest wishes,
Sebastian
