### 2020-10-07 #2694 @sigven
@pibion Thanks! Checkboxes work finally now.
