### 2020-10-23 #2694 @jspaezp
Hello @pibion, I referenced the issue I created in the host repository  a couple of weeks ago.
I also checked some of the boxes that I believe have been fully complied with the standards of the journal.

Let me know if anything additional is required and feel free to reach me to my email if I have not replied to an issue in 24h.

link to the issue:
https://github.com/wesleyburr/subMaldi/issues/5
