### 2020-12-01 #2694 @pibion
@sigven if you'd like to follow @jspaezp's style, you can open an issue over at https://github.com/wesleyburr/subMaldi and copy and paste the checklist into the message for your issue.

If you mention this issue (https://github.com/openjournals/joss-reviews/issues/2694) in your issue over at https://github.com/wesleyburr/subMaldi then it'll show up in our timeline over here.  This is what @jspaezp did near the top of his issue.
