### 2021-01-18 #2694 @wesleyburr
@pibion - can you point me to JOSS policy on authorship changes? Specifically, we'd like to add an author, as we begin the work on the revisions. The maintainer status of this package is changing over to a new software author, and we'd like her to assist with the revisions as she will be the one maintaining the changes and adding to them into the future.

Also, sincere apologies for the delay in beginning revisions - COVID is kicking our butts over here, and finding time to do science in this doom world has been hard.
