### 2021-01-20 #2694 @jspaezp
@wesleyburr I dont't think there is a clear guideline on it ... I feel like it is OK to change them during the review process as long as the paper is modified and the final version contains information regarding what additions to the software/paper made that contribution authorship-worthy.

Especially since the reviewers (us...) have to check the following checkpoint before the paper is actually accepted.

>>> Contribution and authorship: Has the submitting author (@wesleyburr) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
