### 2021-01-25 #2694 @wesleyburr
Great, thanks. I'll make sure to make it clear where my contributions lie, personally - the first round of repo commits were all done by our lead author (and student supervisee), so all of my code writing and editing and such was funneled through her, and she integrated it into her personal repo, which eventually migrated to GitHub. Which isn't great for credit tracking! So for all the edits, we'll make sure that everyone contributes in their own way to the central repo so it's easy for you folks to tick your box.
