### 2021-01-27 97e1f33
Issue (JOSS 5): Added biocViews: as a field in DESCRIPTION to make dependencies searched in BioConductor database
Former-commit-id: e99bf5d016e94c288b25ca2512b731e909238c7a
Former-commit-id: aef01d09f9bb77ff2c8324b943ce14c4dfdd58c0
