+++ DESCRIPTION
@@ -18,3 +18,4 @@ LazyData: true
 Suggests: knitr, rmarkdown, testthat
 VignetteBuilder: knitr
 RoxygenNote: 7.1.1
+biocViews:
