### 2021-01-27 1664fa7
Issue (JOSS 5): Added biocViews: as a field in DESCRIPTION to make dependencies searched in BioConductor database
Former-commit-id: 3d37e9b8758b0f39d04e171a64402928f549c69d
Former-commit-id: 865dd8dcac699b9bcfe99dc031d646ca12a061de
