### 2021-01-27 0561616
Issue (JOSS 5): Added biocViews: as a field in DESCRIPTION to make dependencies searched in BioConductor database
Former-commit-id: 0d7374923f689d7a6070c4f00c2a5d3a649d51a0
Former-commit-id: c3dbc3ae04d2afb81853fa9d48a056f169042c7b
