### 2021-01-27 488eb75
JOSS 5 Issue: Generalized functions for any number of spectra; removed redundancies
Former-commit-id: cb040f3d5e418c7088fb33b9c1a87ad3364f2b33
Former-commit-id: 03ce683e02e8e0ebf81e53f437b0993bcde78fa1
