+++ R/avgSpectra.R
@@ -1,6 +1,6 @@
 # --------------------------------------------------------------------------------------------
-# Date: June 23, 2020
-# Author: Kristen Yeh
+# Date: January 27, 2021
+# Author: Kristen Yeh, Sophie Castel
 # Title: subMALDI Average Spectra
 # --------------------------------------------------------------------------------------------
 
@@ -8,221 +8,56 @@
 # First user should organize standardized data frame so all samples are together
 # First column should always be full_mz, the rest should be spectra, each scan per column
 
-# UPDATE JUNE 2020:
-# Editing the function so that you select which spectra are averaged
-# New column is made in original df beside the selected spectra
 
-avgSpectra <- function(dat, method = "mean", spec1, spec2, 
-                       spec3 = NULL, spec4 = NULL, spec5 = NULL, spec6 = NULL){
+avgSpectra <- function(dat, method = "mean", spectra_cols){
+  
+  # --------------
+  # LOGICAL CHECKS
+  # --------------
+  
+  if(length(spectra_cols) < 2){
+    stop("Only one spectrum input. Please enter two spectra for averaging.")
+  }
+  
+  if(!all(spectra_cols %in% colnames(dat))){
+    logic <- which(!(spectra_cols %in% colnames(dat)))
+    stop(c("Columns '",paste0(as.character(spectra_cols[logic]), sep = "', "), " not found in specified dataframe."))
+  }
   
   if(method == "sum"){
-    .avg_sum(dat = dat, spec1 = spec1, spec2 = spec2, spec3 = spec3, spec4 = spec4,
-             spec5 = spec5, spec6 = spec6) } 
-  else{
-    .avg_mean(dat = dat, spec1 = spec1, spec2 = spec2, spec3 = spec3, spec4 = spec4,
-             spec5 = spec5, spec6 = spec6) }
+    .avg_sum(dat = dat, spectra_cols) } 
+  else if(method == "mean"){
+    .avg_mean(dat = dat, spectra_cols) }
 }
 
 # --------------
 # METHOD = SUM
 # --------------
 
-
-.avg_sum <- function(dat, spec1, spec2, spec3 = NULL, spec4 = NULL,
-                     spec5 = NULL, spec6 = NULL){
+.avg_sum <- function(dat, spectra_cols){
   
-  if(is.null(spec6)){
+  mz <- dat$full_mz
   
-  if(is.null(spec5)){
+  spectra <- lapply(spectra_cols, function(x){dat[x]})
+  i <- do.call(what = data.frame, args = c(spectra))
   
-  if(is.null(spec4)){
+  dat <- cbind(mz, dat, Sum = apply(i, 1, sum, na.rm = TRUE))
   
-  if(is.null(spec3)){
-  
-  if(is.null(spec2)){
-  
-  # One spectrum
-  stop("Only one spectrum input. Please enter two spectra for averaging.")}
-
-  # ---------------------------------------------------------------------
-  # Two spectra
-  else{
-    i1 <- dat[[spec1]]
-    i2 <- dat[[spec2]]
-    i <- data.frame(i1,i2)
-    i <- transform(i, avg = apply(i,1, sum, na.rm = TRUE))
-    
-    dat <- transform(dat, "Sum" = 0)
-    dat["Sum"] <- i$avg
-    return(dat) }}
-  
-  # ---------------------------------------------------------------------
-  # Three spectra
-  else{
-    i1 <- dat[[spec1]]
-    i2 <- dat[[spec2]]
-    i3 <- dat[[spec3]]
-    i <- data.frame(i1,i2,i3)
[…]
