### 2021-01-27 a40fc1f
JOSS 5 Issue: Generalized functions for any number of spectra; removed redundancies
