+++ R/normMethod_max.R
@@ -1,159 +1,48 @@
 # -----------------------------------------------------------------------
-# Last Updated: June 1, 2020
-# Author: Kristen Yeh
+# Last Updated: January 27, 2021
+# Author: Kristen Yeh, Sophie Castel
 # Title: subMALDI: Normalization Method - Maximum Intensity
 # -----------------------------------------------------------------------
-
 
 # -----------------------
 # NORMALIZATION FUNCTION
 # -----------------------
-
 
 # Function rescales intensity to 0,1
 .normalize <- function(x) {
   return((x - min(x)) / (max(x) - min(x)))
 }
 
-
 # ------------------------------
 # METHOD: MAX. OF EACH SPECTRUM
 # ------------------------------
 
-
-.normMethod_max <- function(dat, mass_dat, spec1, spec2 = NULL, 
-                           spec3 = NULL, spec4 = NULL, spec5 = NULL, 
-                           spec6 = NULL){
+.normMethod_max <- function(dat, mass_dat, spectra_cols){
   
-  if(is.null(spec6)){
+  # ---------------------
+  # LOGICAL CHECKS
+  # ---------------------
   
-  if(is.null(spec5)){
-  
-  if(is.null(spec4)){
-  
-  if(is.null(spec3)){
-  
-  if(is.null(spec2)){
-    
-  # Single spectrum
-  mass <- dat[[mass_dat]]
-  intense <- dat[[spec1]]
-  
-  max_Intensity <- max(intense)
+  stopifnot(
+    is.character(mass_dat),
+    is.character(spectra_cols),
+    mass_dat %in% colnames(dat),
+    all(spectra_cols %in% colnames(dat))
+  )
   
   
-  dat <- select(dat, all_of(mass_dat))
-  dat <- transform(dat, "Normalized" = 0)
-  dat["Normalized"] <- .normalize(intense)
+  
+  mz <- dat[[mass_dat]]
+  
+  spectra <- lapply(spectra_cols, function(x){dat[x]})
+  i <- do.call(what = data.frame, args = c(spectra))
+  
+  i_n <- apply(i, 2, FUN = .normalize)
+  dat <- cbind(mz, i_n)
+  colnames(dat) <- c("full_mz", spectra_cols)
+  
   return(dat)
-  }
-  
-  # Two spectra
-  else{
-    mass <- dat[[mass_dat]]
-    intense1 <- dat[[spec1]]
-    intense2 <- dat[[spec2]]
-    
-    max_Intensity1 <- max(intense1)
-    max_Intensity2 <- max(intense2)
-    
-    dat <- select(dat, all_of(mass_dat))
-    dat[spec1] <- .normalize(intense1)
-    dat[spec2] <- .normalize(intense2)
-    return(dat)
-  }}
-  
-  # Three spectra
-  else{
-    mass <- dat[[mass_dat]]
-    intense1 <- dat[[spec1]]
-    intense2 <- dat[[spec2]]
-    intense3 <- dat[[spec3]]
-    
-    max_Intensity1 <- max(intense1)
-    max_Intensity2 <- max(intense2)
-    max_Intensity3 <- max(intense3)
-    
-    dat <- select(dat, all_of(mass_dat))
-    dat[spec1] <- .normalize(intense1)
-    dat[spec2] <- .normalize(intense2)
-    dat[spec3] <- .normalize(intense3)
-    return(dat)
-  }}
-  
-  # Four spectra
-  else{
-    mass <- dat[[mass_dat]]
-    intense1 <- dat[[spec1]]
-    intense2 <- dat[[spec2]]
-    intense3 <- dat[[spec3]]
[…]
