### 2021-01-27 27e5fea
JOSS 5 Issue: Generalized functions for any number of spectra; removed redundancies
Former-commit-id: f13ea02e733751c7eba12d111be0d6092bd28ea6
Former-commit-id: 0aee1c6859c9eb98ba885a71dc1e9a447702df23
