+++ R/find_max.R
@@ -38,6 +38,7 @@
 # -----------------------------------------------------------------------
 # Last Updated: January 27, 2021
 # Author: Kristen Yeh, Sophie Castel
+<<<<<<< HEAD
 # Title: subMALDI: Find Peak Maxima
 # -----------------------------------------------------------------------
 
@@ -64,16 +65,54 @@ find_max <- function (dat, mass_dat, spectra_cols){
   i <- do.call(what = data.frame, args = c(spectra))
   rownames(i) <- mz
   
+=======
+# Title: subMALDI Peak Maximum Functions
+# -----------------------------------------------------------------------
+
+# -----------------
+# FIND PEAK MAXIMA
+# -----------------
+
+# Find peak maximum and associated m/z of spectra
+find_max <- function (dat, mass_dat, spectra_cols){
+  
+  # --------------
+  # LOGICAL CHECKS
+  # --------------
+  
+  stopifnot(
+    is.character(mass_dat),
+    is.character(spectra_cols),
+    mass_dat %in% colnames(dat),
+    all(spectra_cols %in% colnames(dat))
+  )
+  
+  
+  
+  mz <- dat[[mass_dat]]
+  
+  spectra <- lapply(spectra_cols, function(x){dat[x]})
+  i <- do.call(what = data.frame, args = c(spectra))
+  rownames(i) <- mz
+  
+>>>>>>> 29ad768... JOSS 5 Issue: Generalized functions for any number of spectra; removed redundancies
   max_i <- apply(i, 2, max)
   which_max <- apply(i, 2, which.max)
   max_mz <- mz[which_max]
   
+<<<<<<< HEAD
   max_spec <- data.frame(cbind(max_i, max_mz))
+=======
+  max_spec <- cbind(max_i, max_mz)
+>>>>>>> 29ad768... JOSS 5 Issue: Generalized functions for any number of spectra; removed redundancies
   colnames(max_spec) <- c("Intensity (Max)", "Mass")
   
   return(max_spec)
 }
+<<<<<<< HEAD
 >>>>>>> 5cce370... JOSS 5 Issue: Generalized functions for any number of spectra; removed redundancies
+=======
+>>>>>>> 29ad768... JOSS 5 Issue: Generalized functions for any number of spectra; removed redundancies
 
 # -----------------------------------------------------------------------
 <<<<<<< HEAD
@@ -104,6 +143,7 @@ find_max <- function (dat, mass_dat, spectra_cols){
 # ------------------------
 
 find_max_set <- function(dat, mass_dat, spectra_cols){
+<<<<<<< HEAD
   
   # --------------
   # LOGICAL CHECKS
@@ -148,6 +188,36 @@ find_max_set <- function(dat, mass_dat, spectra_cols){
   return(max_spec)
   
 >>>>>>> 5cce370... JOSS 5 Issue: Generalized functions for any number of spectra; removed redundancies
+=======
+  
+  # --------------
+  # LOGICAL CHECKS
+  # --------------
+  
+  stopifnot(
+    is.character(mass_dat),
+    is.character(spectra_cols),
+    mass_dat %in% colnames(dat),
+    all(spectra_cols %in% colnames(dat))
+  )
+  
+  
+  mz <- dat[[mass_dat]]
+  
+  spectra <- lapply(spectra_cols, function(x){dat[x]})
+  i <- do.call(what = data.frame, args = c(spectra))
+  rownames(i) <- mz
+  
+  which_max <- which(i == max(i), arr.ind = TRUE)
+  
+  max_spec <- data.frame(Intensity = as.numeric(i[which_max[1], which_max[2]]), Mass = as.numeric(rownames(which_max)))
+  
[…]
