+++ R/normMethod_median.R
@@ -1,396 +1,65 @@
 # -----------------------------------------------------------------------
-# Last Updated: June 12, 2020
-# Author: Kristen Yeh
+# Last Updated: January 27, 2021
+# Author: Kristen Yeh, Sophie Castel
 # Title: subMALDI: Normalization Method - Median
 # -----------------------------------------------------------------------
-
 
 # ---------------
 # METHOD: MEDIAN
 # ---------------
-
-.normMethod_median <- function(dat, mass_dat, spec1, spec2 = NULL, spec3 = NULL, 
-                              spec4 = NULL, spec5 = NULL, spec6 = NULL){
   
-  if(is.null(spec6)){
-  
-  if(is.null(spec5)){
-  
-  if(is.null(spec4)){
-  
-  if(is.null(spec3)){
-  
-  if(is.null(spec2)){
+.normMethod_median <- function(dat, mass_dat, spectra_cols){
     
-    # One spectrum
-    stop("Only one spectrum input. Please enter two spectra for comparison.")}
-  
-  # ---------------------------------------------------------------------
-  # Two spectra
-  else{
-    intense1 <- dat[[spec1]]
-    intense2 <- dat[[spec2]]
+    # ---------------------
+    # LOGICAL CHECKS
+    # ---------------------
     
-    int <- gather(dat, key = "Spectra", value = "Intensity",
-                  all_of(spec1), all_of(spec2), factor_key = TRUE)
-    int <- dplyr::filter(int, Intensity != 0)
+    stopifnot(
+      is.character(mass_dat),
+      is.character(spectra_cols),
+      mass_dat %in% colnames(dat),
+      all(spectra_cols %in% colnames(dat))
+    )
     
-    s1 <- int[int$Spectra == spec1,] %>% select("full_mz", "Intensity")
-    m1 <- median(s1$Intensity)
+    if(length(spectra_cols) < 2){
+      stop("Only one spectrum input. Please enter two spectra.")
+    }
     
-    s2 <- int[int$Spectra == spec2,] %>% select("full_mz", "Intensity")
-    m2 <- median(s2$Intensity)
     
-    if(m1 > m2){
-      m <- m1/m2
-      dat <- select(dat, all_of(mass_dat))
-      dat[spec1] <- intense1
-      dat[spec2] <- intense2*m
-      return(dat)} 
-    else{
-      m <- m2/m1
-      dat <- select(dat, all_of(mass_dat))
-      dat[spec1] <- intense1*m
-      dat[spec2] <- intense2
-      return(dat)}}}
-  
-  # ---------------------------------------------------------------------
-  # Three spectra
-  else{
-    intense1 <- dat[[spec1]]
-    intense2 <- dat[[spec2]]
-    intense3 <- dat[[spec3]]
+    mz <- dat[[mass_dat]]
     
-    int <- gather(dat, key = "Spectra", value = "Intensity",
-                  all_of(spec1), all_of(spec2), all_of(spec3), factor_key = TRUE)
-    int <- dplyr::filter(int, Intensity != 0)
+    spectra <- lapply(spectra_cols, function(x){dat[x]})
+    i <- do.call(what = data.frame, args = c(spectra))
+    i[i == 0 ] <- NA
     
-    s1 <- int[int$Spectra == spec1,] %>% select("full_mz", "Intensity")
-    s2 <- int[int$Spectra == spec2,] %>% select("full_mz", "Intensity")
[…]

[…]
