### 2021-01-27 29ad768
JOSS 5 Issue: Generalized functions for any number of spectra; removed redundancies
