+++ R/find_max.R
@@ -1,201 +1,44 @@
 # -----------------------------------------------------------------------
-# Last Updated: June 2, 2020
-# Author: Kristen Yeh
+# Last Updated: January 27, 2021
+# Author: Kristen Yeh, Sophie Castel
 # Title: subMALDI Peak Maximum Functions
 # -----------------------------------------------------------------------
-
 
 # -----------------
 # FIND PEAK MAXIMA
 # -----------------
 
-
-# Find peak maximum and associated m/z of a single spectrum
-find_max <- function (dat, mass_dat, spec1, spec2 = NULL, 
-                      spec3 = NULL, spec4 = NULL, spec5 = NULL, spec6 = NULL){
+# Find peak maximum and associated m/z of spectra
+find_max <- function (dat, mass_dat, spectra_cols){
   
-  if(is.null(spec6)) {
-    
-    if(is.null(spec5)) {
-      
-      if(is.null(spec4)){
-        
-        if(is.null(spec3)){
-          
-          if(is.null(spec2)){
-            # Single spectra
-            mass <- dat[[mass_dat]]
-            intense <- dat[[spec1]]
-            
-            max_Intensity <- max(intense)
-            mz <- mass[which(intense == max_Intensity)]
-            max_spec <- c(spec1, mz, max_Intensity)
-            
-            result <- data.frame("Spectrum" = 0, "Mass" = 0, "Max_Intensity" = 0)
-            result [ 1,] <- max_spec
-            return(result)} 
-          
-          # Two spectra
-          else{
-            mass <- dat[[mass_dat]]
-            intense1 <- dat[[spec1]]
-            intense2 <- dat[[spec2]]
-            
-            max_Intensity1 <- max(intense1)
-            mz1 <- mass[which(intense1 == max_Intensity1)]
-            max_spec1 <- c(spec1, mz1, max_Intensity1)
-            
-            max_Intensity2 <- max(intense2)
-            mz2 <- mass[which(intense2 == max_Intensity2)]
-            max_spec2 <- c(spec2, mz2, max_Intensity2)
-            
-            result <- data.frame("Spectrum" = 0, "Mass" = 0, "Max_Intensity" = 0)
-            result[1, ] <- max_spec1
-            result[2, ] <- max_spec2
-            return(result)
-          }} 
-        
-        # Three spectra
-        else{
-          mass <- dat[[mass_dat]]
-          intense1 <- dat[[spec1]]
-          intense2 <- dat[[spec2]]
-          intense3 <- dat[[spec3]]
-          
-          max_Intensity1 <- max(intense1)
-          mz1 <- mass[which(intense1 == max_Intensity1)]
-          max_spec1 <- c(spec1, mz1, max_Intensity1)
-          
-          max_Intensity2 <- max(intense2)
-          mz2 <- mass[which(intense2 == max_Intensity2)]
-          max_spec2 <- c(spec2, mz2, max_Intensity2)
-          
-          max_Intensity3 <- max(intense3)
-          mz3 <- mass[which(intense3 == max_Intensity3)]
-          max_spec3 <- c(spec3, mz3, max_Intensity3)
-          
-          result <- data.frame("Spectrum" = 0, "Mass" = 0, "Max_Intensity" = 0)
-          result[1, ] <- max_spec1
-          result[2, ] <- max_spec2
-          result[3, ] <- max_spec3
-          return(result)
-        }}
-      
-      # Four spectra
[…]
