### 2021-01-27 9a55da9
JOSS 5 Issue: Generalized functions for any number of spectra; removed redundancies
Former-commit-id: a40fc1fd373566956c977a68431cf7c8822a809d
Former-commit-id: 5040d8865ea9b74c4d5fa2d1404a5805aefa87d1
