### 2021-01-27 d7c4b05
JOSS 5 Issue: Generalized functions for any number of spectra; removed redundancies
Former-commit-id: cdd8f5dd786101b47dc27dc5b74fe0c7ca57a9f2
Former-commit-id: 92a4b97bc1cc608d883c5d6578c63790ca973826
