### 2021-01-27 cdd8f5d
JOSS 5 Issue: Generalized functions for any number of spectra; removed redundancies
