+++ dev.Rmd
@@ -13,6 +13,7 @@ knitr::opts_chunk$set(echo = TRUE)
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 =======
 >>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 =======
@@ -33,6 +34,9 @@ knitr::opts_chunk$set(echo = TRUE)
 >>>>>>> 46dcb76... Updated issue log
 =======
 >>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
+=======
+## JOSS 11
+>>>>>>> f435d1f... Updated issue log
 
 ```{r}
 # loading uncompressed data files
@@ -70,6 +74,7 @@ save(Master2, file = "/home/sophie/subMaldi/data/compressed/Master2.rda", compre
 
 ```
 
+<<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
@@ -163,6 +168,14 @@ Wrote biocViews: into DESCRIPTION install_github() can search BioConductor for M
 =======
 ## avgSpectra.R (done)
 >>>>>>> 3753a1b... JOSS 10 (Issue): Fixed formatting of test files
+=======
+## JOSS 5 
+Wrote biocViews: into DESCRIPTION install_github() can search BioConductor for MassSpecWavelet
+
+## JOSS 12 - fix redundancies in code.. this is a doozie
+
+# avgSpectra.R
+>>>>>>> f435d1f... Updated issue log
 
 ```{r}
 avgSpectra <- function(dat, method = "mean", spectra_cols){
@@ -226,6 +239,7 @@ avgSpectra <- function(dat, method = "mean", spectra_cols){
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 ## find_max.R (done)
 =======
 # find_max.R
@@ -245,6 +259,9 @@ avgSpectra <- function(dat, method = "mean", spectra_cols){
 =======
 ## find_max.R (done)
 >>>>>>> 3753a1b... JOSS 10 (Issue): Fixed formatting of test files
+=======
+# find_max.R
+>>>>>>> f435d1f... Updated issue log
 
 ```{r}
 # -----------------
@@ -308,6 +325,7 @@ find_max_set <- function(dat, mass_dat, spectra_cols){
 }
 ```
 
+<<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
@@ -460,6 +478,19 @@ norm_custimp <- function(dat, mass_dat, norm_mz, spectra_cols, showHI = FALSE){
 # METHOD: CUSTOM M/Z VALUE, NOT PRECISE
 # --------------------------------------
 =======
+=======
+# normMethod_custimp.R
+
+```{r}
+```
+
+# normMethod_custom.R
+
+```{r}
+```
+
+# normMethod_max.R
+>>>>>>> f435d1f... Updated issue log
 
 ```{r}
 # -----------------------
@@ -475,11 +506,16 @@ norm_custimp <- function(dat, mass_dat, norm_mz, spectra_cols, showHI = FALSE){
 # METHOD: MAX. OF EACH SPECTRUM
 # ------------------------------
 
+<<<<<<< HEAD
 norm_max <- function(dat, mass_dat, spectra_cols){
 >>>>>>> 3753a1b... JOSS 10 (Issue): Fixed formatting of test files
 
 norm_custimp <- function(dat, mass_dat, norm_mz, spectra_cols, showHI = FALSE){
   
+=======
[…]
