+++ dev.Rmd
@@ -1,0 +1,360 @@
+---
+title: "dev"
+author: "Sophie Castel"
+date: "1/27/2021"
+output: pdf_document
+---
+
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(echo = TRUE)
+```
+
+## JOSS 11
+
+```{r}
+# loading uncompressed data files
+
+load(file = "/home/sophie/subMaldi/data/After1.rda")
+load(file = "/home/sophie/subMaldi/data/After2.rda")
+load(file = "/home/sophie/subMaldi/data/Before1.rda")
+load(file = "/home/sophie/subMaldi/data/Before2.rda")
+load(file = "/home/sophie/subMaldi/data/Blank1.rda")
+load(file = "/home/sophie/subMaldi/data/Blank2.rda")
+load(file = "/home/sophie/subMaldi/data/bsline.rda")
+load(file = "/home/sophie/subMaldi/data/Master.rda")
+load(file = "/home/sophie/subMaldi/data/Master2.rda")
+
+# test compress, comparing methods
+save(Master, file = "/home/sophie/subMaldi/data/compressed/Master_T.rda", compress = TRUE)
+save(Master, file = "/home/sophie/subMaldi/data/compressed/Master_xz.rda", compress = "xz")
+
+file.info("/home/sophie/subMaldi/data/compressed/Master_T.rda")
+file.info("/home/sophie/subMaldi/data/compressed/Master_xz.rda")
+
+# opening test compress
+load("/home/sophie/subMaldi/data/compressed/Master_xz.rda")
+
+# saving else compressed files
+save(After1, file = "/home/sophie/subMaldi/data/compressed/After1.rda", compress = "xz")
+save(After2, file = "/home/sophie/subMaldi/data/compressed/After2.rda", compress = "xz")
+save(Before1, file = "/home/sophie/subMaldi/data/compressed/Before1.rda", compress = "xz")
+save(Before2, file = "/home/sophie/subMaldi/data/compressed/Before2.rda", compress = "xz")
+save(Blank1, file = "/home/sophie/subMaldi/data/compressed/Blank1.rda", compress = "xz")
+save(Blank2, file = "/home/sophie/subMaldi/data/compressed/Blank2.rda", compress = "xz")
+save(bsline, file = "/home/sophie/subMaldi/data/compressed/bsline.rda", compress = "xz")
+save(Master, file = "/home/sophie/subMaldi/data/compressed/Master.rda", compress = "xz")
+save(Master2, file = "/home/sophie/subMaldi/data/compressed/Master2.rda", compress = "xz")
+
+```
+
+## JOSS 5 
+Wrote biocViews: into DESCRIPTION install_github() can search BioConductor for MassSpecWavelet
+
+## JOSS 12 - fix redundancies in code.. this is a doozie
+
+# avgSpectra.R
+
+```{r}
+avgSpectra <- function(dat, method = "mean", spectra_cols){
+  
+  # --------------
+  # LOGICAL CHECKS
+  # --------------
+  
+  if(length(spectra_cols) < 2){
+    stop("Only one spectrum input. Please enter two spectra for averaging.")
+  }
+  
+  if(!all(spectra_cols %in% colnames(dat))){
+    logic <- which(!(spectra_cols %in% colnames(dat)))
+    stop(c("Columns '",paste0(as.character(spectra_cols[logic]), sep = "', "), " not found in specified dataframe."))
+  }
+  
+  if(method == "sum"){
+    .avg_sum(dat = dat, spectra_cols) } 
+  else if(method == "mean"){
[…]
