### 2021-01-28 292c64a
JOSS 12 (Issue):  Generalized functions for any number of spectra; removed redundancies
Former-commit-id: 2c4d112edf76e5ece9aab384d10bc16bee383e04
Former-commit-id: 100a1165efbf23dba18b1b6b44d320fe021cb4a6
