### 2021-01-28 0445685
JOSS 12 (Issue):  Generalized functions for any number of spectra; removed redundancies
Former-commit-id: 75b37414caa075a87314891bd056abc23ca0a1f7
Former-commit-id: 503dd73158b97a526550e5e90a438afb3cd80234
