### 2021-01-28 c733ae7
JOSS 12 (Issue):  Generalized functions for any number of spectra; removed redundancies
