### 2021-01-28 c523b70
JOSS 12 (Issue):  Generalized functions for any number of spectra; removed redundancies
Former-commit-id: 9e867049b57e62b325dd9a8b53296bab5baa3d2e
Former-commit-id: 0f5f34ce5152d430124819028be2ad548f9ae339
