### 2021-01-28 f5d7b41
JOSS 12 (Issue):  Generalized functions for any number of spectra; removed redundancies
Former-commit-id: bede9d4490f44fe55bca7818bd89b79345531ac8
Former-commit-id: 2f895d2896c16d0ba16ae52294390d045b2d6441
