### 2021-01-28 75b3741
JOSS 12 (Issue):  Generalized functions for any number of spectra; removed redundancies
