+++ R/normMethod_RMS.R
@@ -4,11 +4,9 @@
 # Title: subMALDI: Normalization Method - RMS
 # -----------------------------------------------------------------------
 
-
 # -------------------------
 # METHOD: ROOT MEAN SQUARE
 # -------------------------
-
 
 .RMS <- function(y, n){
   right <- 1 / (n - 1)
@@ -16,171 +14,42 @@
   prod <- prod(left, right)
   sqrt(prod)
 }
-  
+
 .norm_RMS <- function(y, RMS){
-  return(y/RMS)
+  return(y/RMS) # single value
 }
 
-.normMethod_RMS <- function(dat, mass_dat, spec1, spec2 = NULL, 
-                            spec3 = NULL, spec4 = NULL, spec5 = NULL, spec6 = NULL){
+.nMethod_RMS <- function(dat, mass_dat, spectra_cols){
   
-  if(is.null(spec6)){
+  # ---------------------
+  # LOGICAL CHECKS
+  # ---------------------
   
-  if(is.null(spec5)){
+  stopifnot(
+    is.character(mass_dat),
+    is.character(spectra_cols),
+    mass_dat %in% colnames(dat),
+    all(spectra_cols %in% colnames(dat))
+  )
   
-  if(is.null(spec4)){
+  mz <- dat[[mass_dat]]
   
-  if(is.null(spec3)){
+  spectra <- lapply(spectra_cols, function(x){dat[x]})
+  i <- do.call(what = data.frame, args = c(spectra))
   
-  if(is.null(spec2)){
+  n <- nrow(dat)
   
-  # One spectra
-  intense <- dat[[spec1]]
-  dat <- select(dat, all_of(mass_dat))
+  rms <- apply(i, 2, FUN = function(x) { .RMS(y = x, n = n)})
   
-  n <- length(intense)
-  rms <- .RMS(y = intense, n = n)
-  dat[spec1] <- .norm_RMS(y = intense, RMS = rms)
-  return(dat)}
+  i_rms <- t(t(i)*rms^-1)  # i / rms matrix multiplication
   
-  # Two spectra
-  else{
-    intense1 <- dat[[spec1]]
-    intense2 <- dat[[spec2]]
-    dat <- select(dat, all_of(mass_dat))
-    
-    n1 <- length(intense1)
-    n2 <- length(intense2)
-    
-    rms1 <- .RMS(y = intense1, n = n1)
-    rms2 <- .RMS(y = intense2, n = n2)
-    
-    dat[spec1] <- .norm_RMS(y = intense1, RMS = rms1)
-    dat[spec2] <- .norm_RMS(y = intense2, RMS = rms2)
-    return(dat)}}
+  dat <- data.frame(mz, i_rms)
   
-  # Three spectra
-  else{
-    intense1 <- dat[[spec1]]
-    intense2 <- dat[[spec2]]
-    intense3 <- dat[[spec3]]
-    dat <- select(dat, all_of(mass_dat))
-    
-    n1 <- length(intense1)
-    n2 <- length(intense2)
-    n3 <- length(intense3)
-    
-    rms1 <- .RMS(y = intense1, n = n1)
-    rms2 <- .RMS(y = intense2, n = n2)
-    rms3 <- .RMS(y = intense3, n = n3)
-    
-    dat[spec1] <- .norm_RMS(y = intense1, RMS = rms1)
-    dat[spec2] <- .norm_RMS(y = intense2, RMS = rms2)
-    dat[spec3] <- .norm_RMS(y = intense3, RMS = rms3)
-    return(dat)}}
[…]
