### 2021-01-28 b8d55fc
JOSS 12 (Issue):  Generalized functions for any number of spectra; removed redundancies
Former-commit-id: f78c2c200f5303a2e691e5d93697227b8a5c1bc2
Former-commit-id: 1fb66fc72f31458476fd8e856399a0372e904f9a
