### 2021-01-28 d1e499f
JOSS 12 (Issue):  Generalized functions for any number of spectra; removed redundancies
