### 2021-01-28 f6538f7
JOSS 12 (Issue):  Generalized functions for any number of spectra; removed redundancies
