### 2021-01-28 6c904e1
JOSS 12 (Issue):  Generalized functions for any number of spectra; removed redundancies
