+++ R/normMethod_quantile.R
@@ -1,16 +1,16 @@
 # --------------------------------------------------------------------------------------------
-# Date: June 16, 2020
-# Author: Kristen Yeh
+# Date: January 28, 2021
+# Author: Kristen Yeh, Sophie Castel
 # Title: Normalization Method - Quantile
 # --------------------------------------------------------------------------------------------
 
 
 # Quantile normalization consists of two steps:
-    # 1. Create a mapping between ranks and values. For rank 1, find the n values,
-       # one per spectrum, that are the smallest value on the spectrum, and save their
-       # averages. Similarly for rank 2 and the second smallest values, and on up to
-       # the n largest values, one per spectrum.
-    # 2. For each spectrum, replace the actual values with these averages.
+# 1. Create a mapping between ranks and values. For rank 1, find the n values,
+# one per spectrum, that are the smallest value on the spectrum, and save their
+# averages. Similarly for rank 2 and the second smallest values, and on up to
+# the n largest values, one per spectrum.
+# 2. For each spectrum, replace the actual values with these averages.
 
 # Basically, index each intensity value in each spectrum then sort
 # by intensity (value). Probably need 2 truncate. Average spectrum for the sorted spectra
@@ -18,520 +18,73 @@
 # The spectra are then reverted to their unsorted order using the index.
 
 
-.normMethod_quantile <- function(dat, mass_dat, spec1, spec2 = NULL, spec3 = NULL, 
-                               spec4 = NULL, spec5 = NULL, spec6 = NULL){
+.nMethod_quantile <- function(dat, mass_dat, spectra_cols){
   
-  if(is.null(spec6)){
+  # ---------------------
+  # LOGICAL CHECKS
+  # ---------------------
   
-  if(is.null(spec5)){
+  stopifnot(
+    is.character(mass_dat),
+    is.character(spectra_cols),
+    mass_dat %in% colnames(dat),
+    all(spectra_cols %in% colnames(dat)),
+    is.numeric(lower),
+    is.numeric(upper)
+  )    
   
-  if(is.null(spec4)){
-  
-  if(is.null(spec3)){
-  
-  if(is.null(spec2)){
-    
-    # One spectrum
-    stop("Only one spectrum input. Please enter two spectra for comparison.")}
-  
-  # ---------------------------------------------------------------------
-  # Two spectra
-  else{
-    int <- select(dat, all_of(mass_dat), all_of(spec1), all_of(spec2))
-    int <- gather(int, key = "Spectra", value = "Intensity",
-                  all_of(spec1), all_of(spec2), factor_key = TRUE)
-    int <- dplyr::filter(int, Intensity != 0)
-    int <- int %>% group_by(Spectra)  %>% mutate(id = row_number())
-    ordr <- int[order(int$Intensity, decreasing = TRUE),]
-    
-    r1 <- int$id[which(int$Spectra == spec1)] %>% max() %>% as.numeric()
-    r2 <- int$id[which(int$Spectra == spec2)] %>% max() %>% as.numeric()
-    
-    if(r1 > r2){
[…]
