### 2021-01-28 2df3f8d
JOSS 12 (Issue):  Generalized functions for any number of spectra; removed redundancies
Former-commit-id: 480ae1b82783e4a262a6fd36e078c13c2624350b
Former-commit-id: ca64c5c0cfdb7a65ceb18279b19be791c586978c
