+++ R/normMethod_stdev.R
@@ -1,208 +1,52 @@
 # -----------------------------------------------------------------------
-# Last Updated: June 12, 2020
-# Author: Kristen Yeh
+# Last Updated: January 28, 2021
+# Author: Kristen Yeh, Sophie Castel
 # Title: subMALDI: Normalization Method - Standard Deviation of Noise
 # -----------------------------------------------------------------------
-
 
 # ------------------------------------
 # METHOD: STANDARD DEVIATION OF NOISE
 # ------------------------------------
 
 # Find a region of the spectrum with only noise and minimal peaks
-    # (Should be same region for all spectra)
+# (Should be same region for all spectra)
 # Evaluate standard deviation of intensity in that region for each spec
 # Divide intensity of EACH PEAK IN SPEC by its st. dev.
 
 
-.normMethod_stdev <- function(dat, mass_dat, lower = 900, upper= 1100 , spec1, spec2 = NULL, spec3 = NULL,
-                              spec4 = NULL, spec5 = NULL, spec6 = NULL){
+.normMethod_stdev <- function(dat, mass_dat, lower = 900, upper= 1100 , spectra_cols){
   
-  if(is.null(spec6)){
+  # ---------------------
+  # LOGICAL CHECKS
+  # ---------------------
   
-  if(is.null(spec5)){
+  stopifnot(
+    is.character(mass_dat),
+    is.character(spectra_cols),
+    mass_dat %in% colnames(dat),
+    all(spectra_cols %in% colnames(dat)),
+    is.numeric(lower),
+    is.numeric(upper)
+  )
   
-  if(is.null(spec4)){
+  mz <- dat[[mass_dat]]
   
-  if(is.null(spec3)){
+  spectra <- lapply(spectra_cols, function(x){dat[x]})
+  i <- do.call(what = data.frame, args = c(spectra))
   
-  if(is.null(spec2)){
-    
-    # One spectrum
-    mass <- dat[[mass_dat]]
-    intense <- dat[[spec1]]
-    
-    noise <- intense[which(mass > lower & upper > mass)]
-    noise <- noise[ noise != 0 ]
-    st_dev <- sd(noise)
-    
-    dat <- select(dat, all_of(mass_dat))
-    dat[spec1] <- .norm_TIC(y = intense, all_y = st_dev)
-    return(dat) }
+  noise <- i[which(mz > lower & upper > mz),]
+  noise[noise == 0] <- NA
   
-  # ---------------------------------------------------------------------
-  # Two spectra
-  else{
-    mass <- dat[[mass_dat]]
-    intense1 <- dat[[spec1]]
-    intense2 <- dat[[spec2]]
-    
-    noise1 <- intense1[which(mass > lower & upper > mass)]
-    noise2 <- intense2[which(mass > lower & upper > mass)]
-    
-    noise1 <- noise1[ noise1 != 0 ]
-    noise2 <- noise2[ noise2 != 0 ]
-    
-    st_dev1 <- sd(noise1)
-    st_dev2 <- sd(noise2)
-    
-    dat <- select(dat, all_of(mass_dat))
-    dat[spec1] <- .norm_TIC(y = intense1, all_y = st_dev1)
-    dat[spec2] <- .norm_TIC(y = intense2, all_y = st_dev2)
-    return(dat) } }
+  std_dev <- apply(noise, 2, FUN = sd, na.rm = TRUE)
   
[…]
