### 2021-01-28 332c7c0
JOSS 12 (Issue):  Generalized functions for any number of spectra; removed redundancies
