### 2021-01-28 16bb7a4
JOSS 12 (Issue):  Generalized functions for any number of spectra; removed redundancies
Former-commit-id: 6c904e1f253e2fd457a19f3a17c708e3b29a0dcf
Former-commit-id: 5b0f6b9ef502d2c6c7928cc8ba6e3893ba6e2488
