+++ R/normMethod_TIC.R
@@ -1,9 +1,8 @@
 # -----------------------------------------------------------------------
-# Last Updated: June 12, 2020
-# Author: Kristen Yeh
+# Last Updated: January 28, 2021
+# Author: Kristen Yeh, Sophie Castel
 # Title: subMALDI: Normalization Method - TIC
 # -----------------------------------------------------------------------
-
 
 # ----------------------
 # METHOD: NORMALIZE TIC
@@ -14,344 +13,59 @@
 # normalization factor.
 
 
-.normMethod_TIC <- function(dat, mass_dat, spec1, spec2 = NULL, 
-                            spec3 = NULL, spec4 = NULL, spec5 = NULL, spec6 = NULL){
+.normMethod_TIC <- function(dat, mass_dat, spectra_cols){
   
-  if(is.null(spec6)){
+  # ---------------------
+  # LOGICAL CHECKS
+  # ---------------------
   
-  if(is.null(spec5)){
+  stopifnot(
+    is.character(mass_dat),
+    is.character(spectra_cols),
+    mass_dat %in% colnames(dat),
+    all(spectra_cols %in% colnames(dat)),
+    is.numeric(lower),
+    is.numeric(upper)
+  )    
   
-  if(is.null(spec4)){
+  if(length(spectra_cols) < 2){
+    stop("Only one spectrum input. Please enter two spectra for comparison.")
+  }
   
-  if(is.null(spec3)){
-    
-    if(is.null(spec2)){
-      
-      # One spectrum
-      stop("Only one spectrum input. Please enter two spectra for comparison.")}
-    
-    # ---------------------------------------------------------------------
-    # Two spectra
-    else{
-      intense1 <- dat[[spec1]]
-      intense2 <- dat[[spec2]]
-      
-      i1 <- sum(intense1)
-      i2 <- sum(intense2)
-      
-      if(i1 > i2){
-        i <- i1/i2
-        dat <- select(dat, all_of(mass_dat))
-        dat[spec1] <- intense1
-        dat[spec2] <- intense2*i
-        return(dat)} 
-      else{
-        i <- i2/i1
-        dat <- select(dat, all_of(mass_dat))
-        dat[spec1] <- intense1*i
-        dat[spec2] <- intense2
-        return(dat)}}}
   
-  # ---------------------------------------------------------------------
-  # Three spectra
-  else{
-    intense1 <- dat[[spec1]]
-    intense2 <- dat[[spec2]]
-    intense3 <- dat[[spec3]]
-    
-    i1 <- sum(intense1)
-    i2 <- sum(intense2)
-    i3 <- sum(intense3)
-    t <- c(i1,i2,i3)
-    t_val <- max(t)
-    
-    if(i1 == t_val){
-      a <- i1/i2
-      b <- i1/i3
-      
-      dat <- select(dat, all_of(mass_dat))
-      dat[spec1] <- intense1
-      dat[spec2] <- intense2*a
-      dat[spec3] <- intense3*b
-      return(dat)}
-    else if(i2 == t_val){
-      a <- i2/i1
-      b <- i2/i3
-      
-      dat <- select(dat, all_of(mass_dat))
-      dat[spec1] <- intense1*a
-      dat[spec2] <- intense2
-      dat[spec3] <- intense3*b
-      return(dat)}
-    else{
-      a <- i3/i1
-      b <- i3/i2
-      
-      dat <- select(dat, all_of(mass_dat))
-      dat[spec1] <- intense1*a
-      dat[spec2] <- intense2*b
-      dat[spec3] <- intense3
[…]
