### 2021-01-28 722f3ec
JOSS 12 (Issue):  Generalized functions for any number of spectra; removed redundancies
Former-commit-id: 1fc39064d2dff5f0614117466cbb9176e79dff9a
Former-commit-id: fae94f70b6edce39d0105fa1bfe6af2f7d54c39b
