+++ DESCRIPTION
@@ -98,6 +98,7 @@ Description: subMALDI is an open framework tool that permits organization,
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 Depends: R (>= 3.5.0), dplyr (>= 0.8.0), tidyr (>= 1.1.0), ggplot2 (>= 3.3.0), ggpmisc (>= 0.3.3)
 <<<<<<< HEAD
 <<<<<<< HEAD
@@ -143,6 +144,9 @@ Depends: R (>= 3.5.0), dplyr (>= 0.8.0), tidyr (>= 1.1.0), ggplot2 (>= 3.3.0), g
 =======
 Depends: R (>= 3.5.0), dplyr (>= 0.8.0), tidyr (>= 1.1.0), ggplot2 (>= 3.3.0), ggpmisc (>= 0.3.3), reshape2 (>= 1.4.4), RColorBrewer (>= 1.1-2)
 >>>>>>> 24787bc... Added RColorBrewer dependency
+=======
+Depends: R (>= 3.5.0), dplyr (>= 0.8.0), tidyr (>= 1.1.0), ggplot2 (>= 3.3.0), ggpmisc (>= 0.3.3), reshape2 (>= 1.4.4)
+>>>>>>> 5914622... added reshape2 dependency
 Imports: signal (>= 0.7-6), MassSpecWavelet (>= 1.48.0)
 >>>>>>> 27b29da... Updated docs
 License: GPL (>= 3) + file LICENSE
