+++ DESCRIPTION
@@ -10,7 +10,7 @@ Description: subMALDI is an open framework tool that permits organization,
     of spectral data sets. As a result of the package's open framework, subMALDI data 
     sets are compatible with functions from a wide variety of other R packages, and 
     user-defined functions are easier to implement and test.
-Depends: R (>= 3.5.0), dplyr (>= 0.8.0), tidyr (>= 1.1.0), ggplot2 (>= 3.3.0), ggpmisc (>= 0.3.3)
+Depends: R (>= 3.5.0), dplyr (>= 0.8.0), tidyr (>= 1.1.0), ggplot2 (>= 3.3.0), ggpmisc (>= 0.3.3), reshape2 (>= 1.4.4)
 Imports: signal (>= 0.7-6), MassSpecWavelet (>= 1.48.0)
 License: GPL (>= 3) + file LICENSE
 Encoding: UTF-8
