+++ dev.Rmd
@@ -80,10 +80,14 @@ Wrote biocViews: into DESCRIPTION install_github() can search BioConductor for M
 ## JOSS 5 
 Wrote biocViews: into DESCRIPTION install_github() can search BioConductor for MassSpecWavelet
 
-## JOSS 12 - fix redundancies in code.. this is a doozie
+## JOSS 12 - generalize and fix redundancies in code
 
+<<<<<<< HEAD
 # avgSpectra.R
 >>>>>>> f435d1f... Updated issue log
+=======
+# avgSpectra.R (done)
+>>>>>>> d7c9654... Updated issue log
 
 ```{r}
 avgSpectra <- function(dat, method = "mean", spectra_cols){
@@ -142,10 +146,14 @@ avgSpectra <- function(dat, method = "mean", spectra_cols){
 ```
 
 <<<<<<< HEAD
+<<<<<<< HEAD
 ## find_max.R (done)
 =======
 # find_max.R
 >>>>>>> f435d1f... Updated issue log
+=======
+# find_max.R (done)
+>>>>>>> d7c9654... Updated issue log
 
 ```{r}
 # -----------------
@@ -350,8 +358,12 @@ norm_custimp <- function(dat, mass_dat, norm_mz, spectra_cols, showHI = FALSE){
 ```{r}
 ```
 
+<<<<<<< HEAD
 # normMethod_max.R
 >>>>>>> f435d1f... Updated issue log
+=======
+# normMethod_max.R (done)
+>>>>>>> d7c9654... Updated issue log
 
 ```{r}
 # -----------------------
@@ -443,10 +455,14 @@ norm_max <- function(dat, mass_dat, spectra_cols){
 ```
 
 <<<<<<< HEAD
+<<<<<<< HEAD
 ## norm_median.R (done)
 =======
 # normMethod_median.R
 >>>>>>> f435d1f... Updated issue log
+=======
+# normMethod_median.R (done)
+>>>>>>> d7c9654... Updated issue log
 
 ```{r}
 
@@ -514,6 +530,7 @@ norm_med<- function(dat, mass_dat, spectra_cols){
 }
 ```
 
+<<<<<<< HEAD
 <<<<<<< HEAD
 ## norm_quantile.R (done)
 
@@ -1142,26 +1159,329 @@ library(testthat)
 ```
 =======
 # normMethod_quantile.R
+=======
+# normMethod_quantile.R (done)
+>>>>>>> d7c9654... Updated issue log
 
 ```{r}
+# --------------------------------------------------------------------------------------------
+# Date: January 28, 2021
+# Author: Kristen Yeh, Sophie Castel
+# Title: Normalization Method - Quantile
+# --------------------------------------------------------------------------------------------
+
+
+# Quantile normalization consists of two steps:
+    # 1. Create a mapping between ranks and values. For rank 1, find the n values,
+       # one per spectrum, that are the smallest value on the spectrum, and save their
+       # averages. Similarly for rank 2 and the second smallest values, and on up to
+       # the n largest values, one per spectrum.
+    # 2. For each spectrum, replace the actual values with these averages.
+
+# Basically, index each intensity value in each spectrum then sort
+# by intensity (value). Probably need 2 truncate. Average spectrum for the sorted spectra
+# is calculated. The averages are then inserted into the sorted spectra.
+# The spectra are then reverted to their unsorted order using the index.
+
+
[…]
