+++ dev.Rmd
@@ -126,10 +126,14 @@ Wrote biocViews: into DESCRIPTION install_github() can search BioConductor for M
 ## JOSS 5 
 Wrote biocViews: into DESCRIPTION install_github() can search BioConductor for MassSpecWavelet
 
-## JOSS 12 - fix redundancies in code.. this is a doozie
+## JOSS 12 - generalize and fix redundancies in code
 
+<<<<<<< HEAD
 # avgSpectra.R
 >>>>>>> 46dcb76... Updated issue log
+=======
+# avgSpectra.R (done)
+>>>>>>> 1b35d86... Updated issue log
 
 ```{r}
 avgSpectra <- function(dat, method = "mean", spectra_cols){
@@ -191,6 +195,7 @@ avgSpectra <- function(dat, method = "mean", spectra_cols){
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 ## find_max.R (done)
 =======
 # find_max.R
@@ -204,6 +209,9 @@ avgSpectra <- function(dat, method = "mean", spectra_cols){
 =======
 # find_max.R
 >>>>>>> 46dcb76... Updated issue log
+=======
+# find_max.R (done)
+>>>>>>> 1b35d86... Updated issue log
 
 ```{r}
 # -----------------
@@ -556,8 +564,12 @@ norm_custimp <- function(dat, mass_dat, norm_mz, spectra_cols, showHI = FALSE){
 ```{r}
 ```
 
+<<<<<<< HEAD
 # normMethod_max.R
 >>>>>>> 46dcb76... Updated issue log
+=======
+# normMethod_max.R (done)
+>>>>>>> 1b35d86... Updated issue log
 
 ```{r}
 # -----------------------
@@ -660,6 +672,7 @@ norm_max <- function(dat, mass_dat, spectra_cols){
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 ## norm_median.R (done)
 =======
 # normMethod_median.R
@@ -673,6 +686,9 @@ norm_max <- function(dat, mass_dat, spectra_cols){
 =======
 # normMethod_median.R
 >>>>>>> 46dcb76... Updated issue log
+=======
+# normMethod_median.R (done)
+>>>>>>> 1b35d86... Updated issue log
 
 ```{r}
 
@@ -748,6 +764,7 @@ norm_med<- function(dat, mass_dat, spectra_cols){
 }
 ```
 
+<<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
@@ -2020,26 +2037,329 @@ library(testthat)
 ```
 =======
 # normMethod_quantile.R
+=======
+# normMethod_quantile.R (done)
+>>>>>>> 1b35d86... Updated issue log
 
 ```{r}
+# --------------------------------------------------------------------------------------------
+# Date: January 28, 2021
+# Author: Kristen Yeh, Sophie Castel
+# Title: Normalization Method - Quantile
+# --------------------------------------------------------------------------------------------
+
+
+# Quantile normalization consists of two steps:
+    # 1. Create a mapping between ranks and values. For rank 1, find the n values,
+       # one per spectrum, that are the smallest value on the spectrum, and save their
+       # averages. Similarly for rank 2 and the second smallest values, and on up to
+       # the n largest values, one per spectrum.
+    # 2. For each spectrum, replace the actual values with these averages.
+
+# Basically, index each intensity value in each spectrum then sort
[…]
