+++ dev.Rmd
@@ -50,9 +50,9 @@ save(Master2, file = "/home/sophie/subMaldi/data/compressed/Master2.rda", compre
 ## JOSS 5 
 Wrote biocViews: into DESCRIPTION install_github() can search BioConductor for MassSpecWavelet
 
-## JOSS 12 - fix redundancies in code.. this is a doozie
+## JOSS 12 - generalize and fix redundancies in code
 
-# avgSpectra.R
+# avgSpectra.R (done)
 
 ```{r}
 avgSpectra <- function(dat, method = "mean", spectra_cols){
@@ -110,7 +110,7 @@ avgSpectra <- function(dat, method = "mean", spectra_cols){
 
 ```
 
-# find_max.R
+# find_max.R (done)
 
 ```{r}
 # -----------------
@@ -184,7 +184,7 @@ find_max_set <- function(dat, mass_dat, spectra_cols){
 ```{r}
 ```
 
-# normMethod_max.R
+# normMethod_max.R (done)
 
 ```{r}
 # -----------------------
@@ -271,7 +271,7 @@ find_max_set <- function(dat, mass_dat, spectra_cols){
 
 ```
 
-# normMethod_median.R
+# normMethod_median.R (done)
 
 ```{r}
 
@@ -335,26 +335,323 @@ find_max_set <- function(dat, mass_dat, spectra_cols){
 }
 ```
 
-# normMethod_quantile.R
+# normMethod_quantile.R (done)
 
 ```{r}
+# --------------------------------------------------------------------------------------------
+# Date: January 28, 2021
+# Author: Kristen Yeh, Sophie Castel
+# Title: Normalization Method - Quantile
+# --------------------------------------------------------------------------------------------
+
+
+# Quantile normalization consists of two steps:
+    # 1. Create a mapping between ranks and values. For rank 1, find the n values,
+       # one per spectrum, that are the smallest value on the spectrum, and save their
+       # averages. Similarly for rank 2 and the second smallest values, and on up to
+       # the n largest values, one per spectrum.
+    # 2. For each spectrum, replace the actual values with these averages.
+
+# Basically, index each intensity value in each spectrum then sort
+# by intensity (value). Probably need 2 truncate. Average spectrum for the sorted spectra
+# is calculated. The averages are then inserted into the sorted spectra.
+# The spectra are then reverted to their unsorted order using the index.
+
+
+.nMethod_quantile <- function(dat, mass_dat, spectra_cols){
+  
+  # ---------------------
+  # LOGICAL CHECKS
+  # ---------------------
+  
+  stopifnot(
+    is.character(mass_dat),
+    is.character(spectra_cols),
+    mass_dat %in% colnames(dat),
+    all(spectra_cols %in% colnames(dat)),
+    is.numeric(lower),
+    is.numeric(upper)
+  )    
+  
+  if(length(spectra_cols) < 2){
+    stop("Only one spectrum input. Please enter two spectra for comparison.")
+    }
+ 
+  mz <- dat[[mass_dat]]
+  
+  spectra <- lapply(spectra_cols, function(x){dat[x]})
+  i <- do.call(what = data.frame, args = c(spectra))
+  
+  i <- data.frame(mz, i)
+  colnames(i) <- c("full_mz", spectra_cols)
+  
+  i_melt <- melt(i, id.vars = "full_mz")
[…]
