### 2021-01-28 e72b289
JOSS 12 (Issue): Generalized functions for any number of spectra; removed redundancies
Former-commit-id: 73c5a0b1f326f88b797b9703d054d3850d3947e1
Former-commit-id: 3f582f452f6ad7d7efb51d8d5a6ad95d13682b24
