+++ R/normMethod_custom.R
@@ -1,403 +1,65 @@
 # -----------------------------------------------------------------------
-# Last Updated: June 29, 2020
-# Author: Kristen Yeh
+# Last Updated: January 28, 2020
+# Author: Kristen Yeh, Sophie Castel
 # Title: subMALDI: Normalization Method - Custom m/z
 # -----------------------------------------------------------------------
-
 
 # -------------------------
 # METHOD: CUSTOM M/Z VALUE
 # -------------------------
 
-
 .norm_custom <- function(y, custom_y){
-  return((y - min(y)) / (custom_y - min(y)))
+  return((y - min(y)) / (custom_y - min(y))) # single value
 }
 
-.normMethod_custom <- function(dat, mass_dat, norm_mz, spec1, 
-                              spec2 = NULL, spec3 = NULL, 
-                              spec4 = NULL, spec5 = NULL, 
-                              spec6 = NULL, showHI = FALSE){
-  if(showHI == FALSE){
+.normMethod_custom <- function(dat, mass_dat, norm_mz, spectra_cols, showHI = FALSE){
+  
+  # ---------------------
+  # LOGICAL CHECKS
+  # ---------------------
+  
+  stopifnot(
+    is.character(mass_dat),
+    is.character(spectra_cols),
+    mass_dat %in% colnames(dat),
+    all(spectra_cols %in% colnames(dat))
+  )
+  
+  
+  
   if(is.null(norm_mz)){
     stop('Please select a m/z value. norm_mz is NULL.')
-  } else{
-  
-  if(is.null(spec6)){
-  
-  if(is.null(spec5)){
-  
-  if(is.null(spec4)){
-  
-  if(is.null(spec3)){
-  
-  if(is.null(spec2)){
-    
-          # One spectrum
-          mass <- dat[[mass_dat]]
-          intense <- dat[[spec1]]
-          
-          custom_int <- intense[which(mass == norm_mz)]
-          
-          if(custom_int > 0){
-            dat <- select(dat, all_of(mass_dat))
-            norm <- .norm_custom(y = intense, custom_y = custom_int)
-            norm[norm > 1] <- 1 
-            dat[spec1] <- norm
-            return(dat)}
-          else{stop('The selected maximum intensity is 0.')}
   } 
-    
-    # Two spectra
-    else {
-    mass <- dat[[mass_dat]]
-    intense1 <- dat[[spec1]]
-    intense2 <- dat[[spec2]]
-    
-    custom_int1 <- intense1[which(mass == norm_mz)]
-    custom_int2 <- intense2[which(mass == norm_mz)]
-    
-    custom <- c(custom_int1,custom_int2)
-    if(any(custom == 0)){stop('The selected maximum intensity is 0 in some spectra.')}
-    
-    else{
-    dat <- select(dat, all_of(mass_dat))
-    norm1 <- .norm_custom(y = intense1, custom_y = custom_int1)
-    norm2 <- .norm_custom(y = intense2, custom_y = custom_int2)
-    
-    norm1[norm1 > 1] <- 1 
-    norm2[norm2 > 1] <- 1 
-    dat[spec1] <- norm1
-    dat[spec2] <- norm2
-      return(dat)} 
-    }}
   
-  # Three spectra
-  else{
-    mass <- dat[[mass_dat]]
-    intense1 <- dat[[spec1]]
-    intense2 <- dat[[spec2]]
-    intense3 <- dat[[spec3]]
-    
-    custom_int1 <- intense1[which(mass == norm_mz)]
-    custom_int2 <- intense2[which(mass == norm_mz)]
-    custom_int3 <- intense3[which(mass == norm_mz)]
-    
[…]
