### 2021-01-28 6d4a2e8
JOSS 12 (Issue): Generalized functions for any number of spectra; removed redundancies
