+++ R/normMethod_custimp.R
@@ -1,538 +1,62 @@
 # -----------------------------------------------------------------------
-# Last Updated: July 30, 2020
-# Author: Kristen Yeh
+# Last Updated: January 28, 2021
+# Author: Kristen Yeh, Sophie Castel
 # Title: subMALDI: Normalization Method - Custom m/z, imprecise
 # -----------------------------------------------------------------------
 
-
-# --------------------------------------
-# METHOD: CUSTOM M/Z VALUE, NOT PRECISE
-# --------------------------------------
-
-
-.normMethod_custimp <- function(dat, mass_dat, norm_mz, spec1, 
-                                spec2 = NULL, spec3 = NULL, 
-                                spec4 = NULL, spec5 = NULL, 
-                                spec6 = NULL, showHI = FALSE){
-  if(showHI == FALSE){
-    if(is.null(norm_mz)){
-      stop('Please select a m/z value. norm_mz is NULL.')
-    } else{
-      
-      if(is.null(spec6)){
-        
-        if(is.null(spec5)){
-          
-          if(is.null(spec4)){
-            
-            if(is.null(spec3)){
-              
-              if(is.null(spec2)){
-                
-                # One spectrum
-                mass <- dat[[mass_dat]]
-                intense <- dat[[spec1]]
-                
-                i <- which(startsWith(as.character(mass), norm_mz))
-                custom_int <- intense[i]
-                
-                if(length(custom_int) > 1){
-                  stop("More than one peak per spectra for the given m/z value. 
-                     Please be more precise.")
-                } else{
-                  
-                  if(custom_int > 0){
-                    dat <- select(dat, all_of(mass_dat))
-                    norm <- .norm_custom(y = intense, custom_y = custom_int)
-                    norm[norm > 1] <- 1 
-                    dat[spec1] <- norm
-                    return(dat)}
-                  else{stop('The selected maximum intensity is 0.')}
-                }}
-              # Two spectra
-              else {
-                mass <- dat[[mass_dat]]
-                intense1 <- dat[[spec1]]
-                intense2 <- dat[[spec2]]
-                
-                i <- which(startsWith(as.character(mass), norm_mz))
-                custom_int1 <- intense1[i]
-                custom_int1 <- custom_int1[which(custom_int1 != 0)]
-                custom_int2 <- intense2[i]
-                custom_int2 <- custom_int2[which(custom_int2 != 0)]
-                
-                if(length(custom_int1) > 1 | length(custom_int2) > 1){
-                  stop('More than one peak per spectra for the given m/z value. 
-                     Please be more precise.')
-                } else if(length(custom_int1) < 1 | length(custom_int2) < 1){  
-                  stop('The selected maximum intensity is 0 in some spectra.')
-                } else{
-                  dat <- select(dat, all_of(mass_dat))
-                  norm1 <- .norm_custom(y = intense1, custom_y = custom_int1)
-                  norm2 <- .norm_custom(y = intense2, custom_y = custom_int2)
-                  
-                  norm1[norm1 > 1] <- 1 
-                  norm2[norm2 > 1] <- 1 
-                  dat[spec1] <- norm1
-                  dat[spec2] <- norm2
-                  return(dat)} 
-              }}
-            
-            # Three spectra
-            else{
-              mass <- dat[[mass_dat]]
-              intense1 <- dat[[spec1]]
-              intense2 <- dat[[spec2]]
-              intense3 <- dat[[spec3]]
-              
[…]
