### 2021-01-28 35cdf1b
JOSS 12 (Issue): Generalized functions for any number of spectra; removed redundancies
