### 2021-01-29 a1a9b9b
Updated documentation to reflect recent changes.
