### 2021-01-29 390a640
Updated documentation to reflect recent changes.
Former-commit-id: a1a9b9b9259965db773a3b28350fb1f17676b735
Former-commit-id: 77444408ec29aeb6f00e154ff54d4d77ffe7e482
