+++ man/subMALDI-package.Rd
@@ -22,6 +22,9 @@ Authors:
   \item Kristen Yeh (\href{https://orcid.org/0000-0002-3411-6816}{ORCID})
   \item Wesley Burr (\href{https://orcid.org/0000-0002-2058-1899}{ORCID})
 <<<<<<< HEAD
+<<<<<<< HEAD
+=======
+>>>>>>> ae90a09... Updated documentation to reflect recent changes.
   \item Sophie Castel (\href{https://orcid.org/0000-0001-9086-0917}{ORCID})
 }
 }
