### 2021-01-29 c777ab3
Updated documentation to reflect recent changes.
