### 2021-01-29 8c07c03
Updated documentation to reflect recent changes.
