### 2021-01-29 4faa2ce
Updated documentation to reflect recent changes.
