### 2021-01-29 21f90cc
JOSS 12 (Issue): Generalized functions for any number of spectra; removed redundancies
Former-commit-id: 3c16258d5eaf0497266a3807b60cd7c305fe0a68
Former-commit-id: ce953d4db4d90b100608e9167c841f41833d296e
