### 2021-01-29 c33d9d1
JOSS 12 (Issue): Generalized functions for any number of spectra; removed redundancies
