### 2021-01-29 eb93a69
JOSS 12 (Issue): Generalized functions for any number of spectra; removed redundancies
Former-commit-id: c33d9d13910cd485f9e42be35c630a053d662506
Former-commit-id: 41a6b792c3b12b5597b7b691de49cd89cc92138b
