### 2021-01-29 44575b7
JOSS 12 (Issue): Generalized functions for any number of spectra; removed redundancies
