### 2021-01-29 3c16258
JOSS 12 (Issue): Generalized functions for any number of spectra; removed redundancies
