+++ R/normSpectra.R
@@ -159,41 +159,60 @@ normSpectra <- function(dat, mass_dat, method = NULL, norm_mz = NULL, upper = NU
 
 
 normSpectra <- function(dat, mass_dat, method = NULL, norm_mz = NULL, upper = NULL, lower = NULL,
-                        spectra_cols = spectra_cols, showHI = FALSE){
+                        spectra_cols = NULL, showHI = FALSE){
+  
+  methods <- c("max", "max_set", "custom", "custom_imprecise", "TIC", "rel_TIC", "RMS", "median", "stdev", "quantile")
+    
+  # ---------------------
+  # LOGICAL CHECKS
+  # ---------------------
   
   if(is.null(method)){ stop('Please select a valid normalization method. See ?normSpectra for list of methods.') } 
   
-  else { 
-    if(method == "max"){ .normMethod_max(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) }
+  if(!(method %in% methods)){
+    stop(c("Invalid selection: method = '", method, "'.  Please make your selection from the following options: ", paste0(methods, sep = ", "),"."))
+  }
   
-    else if(method == "custom"){ .normMethod_custom(dat = dat, mass_dat = mass_dat, norm_mz = norm_mz, spectra_cols = spectra_cols, showHI = showHI) } 
+  
+  
+  if(method == "max"){ 
+    .normMethod_max(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) 
+  }
+  
+  if(method == "custom"){ 
+    .normMethod_custom(dat = dat, mass_dat = mass_dat, norm_mz = norm_mz, spectra_cols = spectra_cols, showHI = showHI) 
+  } 
     
-    else if(method == "custom_imprecise"){ .normMethod_custimp(dat = dat, mass_dat = mass_dat, norm_mz = norm_mz, spectra_cols = spectra_cols, showHI = showHI) } 
+  if(method == "custom_imprecise"){ 
+    .normMethod_custimp(dat = dat, mass_dat = mass_dat, norm_mz = norm_mz, spectra_cols = spectra_cols, showHI = showHI) 
+  } 
     
-    else if(method == "max_set"){ 
-       if(is.null(spec2)){ stop('Only one spectrum input. Please enter two spectra for comparison.') } 
-        else{ .normMethod_max_set(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) } }
+  if(method == "max_set"){ 
+    .normMethod_max_set(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) 
+  }
     
-    else if( method == "TIC"){
-      if(is.null(spec2)){ stop('Only one spectrum input. Please enter two spectra for comparison.') } 
-      else{ .normMethod_TIC(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) } }
+  if( method == "TIC"){
+    .normMethod_TIC(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) 
+  } 
     
-    else if(method == "RMS"){
-      .normMethod_RMS(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) }
+  if(method == "RMS"){
+    .normMethod_RMS(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) 
+  }
     
-    else if(method == "rel_TIC"){
-      .normMethod_rel_TIC(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) }
+  if(method == "rel_TIC"){
+    .normMethod_rel_TIC(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) 
+  }
     
[…]
