### 2021-01-29 ddbcc2b
JOSS 12 (Issue): Generalized functions for any number of spectra; removed redundancies
Former-commit-id: 44575b7b3d97bb86a7c95fdbc239923dd54601df
Former-commit-id: f08bdd9f675fdaae6606e902179d715ba3875b95
