### 2021-01-29 03c25d1
JOSS 12 (Issue): Generalized functions for any number of spectra; removed redundancies
