+++ R/find_max.R
@@ -165,6 +165,10 @@ find_max_set <- function(dat, mass_dat, spectra_cols){
     stop("Only one spectrum input. Please enter two spectra.")
   }
   
+  if(length(spectra_cols) < 2){
+    stop("Only one spectrum input. Please enter two spectra.")
+  }
+  
   
   mz <- dat[[mass_dat]]
   
