### 2021-01-29 ac65560
Updated documentation to reflect recent changes.
Former-commit-id: f30cfc505b27455cad906ce08f4f91135dd83c4b
Former-commit-id: 52a06f7716b444dbe13f6965b1810b3b362d1fa4
