### 2021-02-02 e50771b
Rendered new documentation using package 'roxygen2'.
Former-commit-id: c0c130f46ff1b487f2e931f6ec53cc5979d8d013
Former-commit-id: d3537f9d3b0577330631df55c4639dfecf698ae3
