+++ R/smoothSpectrum.R
@@ -1,4 +1,7 @@
 <<<<<<< HEAD
+<<<<<<< HEAD
+=======
+>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 ##' Smooth Noise in Spectrum
 ##' 
 ##' Offers two different methods for smoothing noise in raw spectral data: a
@@ -48,6 +51,7 @@
 ##' 
 
 
+<<<<<<< HEAD
 # ----------------------------------------------------------------------------
 # Last Updated: July 10, 2020
 # Author: Kristen Yeh
@@ -58,6 +62,12 @@
 # Author: Kristen Yeh
 # Title: subMALDI - Smoothing Parent Function
 >>>>>>> e43371a... Added smoothing
+=======
+# ----------------------------------------------------------------------------
+# Last Updated: July 10, 2020
+# Author: Kristen Yeh
+# Title: subMALDI: Smooth Noise in Spectrum
+>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 # ----------------------------------------------------------------------------
 
 
