+++ R/normSpectra.R
@@ -1,4 +1,7 @@
 <<<<<<< HEAD
+<<<<<<< HEAD
+=======
+>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 ##' Normalize Spectral Data
 ##' 
 ##' Normalize spectral data to a common scale using several different methods.
@@ -81,6 +84,7 @@
 ##' 
 ##' 
 
+<<<<<<< HEAD
 # -----------------------------------------------------------------------
 # Last Updated: January 28, 2021
 # Author: Kristen Yeh, Sophie Castel
@@ -146,17 +150,13 @@ normSpectra <- function(dat, mass_dat, method = NULL, norm_mz = NULL, upper = NU
   
   return(r)
 =======
+=======
+>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 # -----------------------------------------------------------------------
 # Last Updated: January 28, 2021
 # Author: Kristen Yeh, Sophie Castel
-# Title: subMALDI: Umbrella Normalization Function
+# Title: subMALDI: Normalize Spectral Data
 # -----------------------------------------------------------------------
-
-
-# ------------
-# normSpectra
-# ------------
-
 
 normSpectra <- function(dat, mass_dat, method = NULL, norm_mz = NULL, upper = NULL, lower = NULL,
                         spectra_cols = NULL, showHI = FALSE){
@@ -220,7 +220,11 @@ normSpectra <- function(dat, mass_dat, method = NULL, norm_mz = NULL, upper = NU
 
 
 <<<<<<< HEAD
+<<<<<<< HEAD
 # -----------------------------------------------------------------------
 =======
 >>>>>>> f70720d... TIC norm added
+=======
+# -----------------------------------------------------------------------
+>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 
