+++ man/mapSpectrum.Rd
@@ -3,6 +3,7 @@
 \name{mapSpectrum}
 \alias{mapSpectrum}
 <<<<<<< HEAD
+<<<<<<< HEAD
 \title{Map Spectrum to \emph{m/z} Vector}
 =======
 \title{
@@ -15,6 +16,10 @@ Fills in the columns of the empty data frame created using \code{createSpecDF}. 
 \usage{
 <<<<<<< HEAD
 <<<<<<< HEAD
+=======
+\title{Map Spectrum to \emph{m/z} Vector}
+\usage{
+>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 mapSpectrum(
   dat,
   massCol,
@@ -24,6 +29,7 @@ mapSpectrum(
   spec_df,
   colName
 )
+<<<<<<< HEAD
 =======
 mapSpectrum(dat, massCol, intenseCol, dig = 4, thresh = 5e-5 spec_df, colName)
 >>>>>>> 270485a... Updated documentation
@@ -33,6 +39,10 @@ mapSpectrum(dat, massCol, intenseCol, dig = 4, thresh = 1e-04, spec_df, colName)
 }
 \arguments{
 <<<<<<< HEAD
+=======
+}
+\arguments{
+>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 \item{dat}{A pairwise data frame containing your spectral data. Should
 contain two columns: one for \emph{m/z} and one for intensity.}
 
@@ -49,6 +59,7 @@ to fill. Default = 4.}
 \item{thresh}{Single numeric value; all \emph{m/z} values within
 \code{thresh} of each other are binned under that with the maximum
 intensity. Default = 5e-5.}
+<<<<<<< HEAD
 
 \item{spec_df}{An empty data frame created using \code{createSpecDF}.}
 
@@ -67,10 +78,19 @@ of the mapped spectral data frame.
   \item{spec_df}{An empty data frame created using \code{createSpecDF}.}
   \item{colName}{A character string; the name of the column that should be filled with the spectral data.}
 }
+=======
+>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 
-\value{
-Returns a vector that is used to fill in the \code{colName} column of the mapped spectral data frame.
+\item{spec_df}{An empty data frame created using \code{createSpecDF}.}
+
+\item{colName}{A character string; the name of the column that should be
+filled with the spectral data.}
 }
+\value{
+Returns a vector that is used to fill in the \code{colName} column
+of the mapped spectral data frame.
+}
+<<<<<<< HEAD
 \references{
 https://github.com/wesleyburr/subMaldi
 >>>>>>> 270485a... Updated documentation
@@ -81,6 +101,19 @@ Fills in the columns of the empty data frame created using
 compared to the \code{full_mz} vector. All peaks within the \code{thresh}
 of one another are binned, and only the maximum intensity of that bin is
 filled into the mapped spectrum.
+=======
+\description{
+Fills in the columns of the empty data frame created using
+\code{createSpecDF}. Mass to charge values from pairwise spectral data are
+compared to the \code{full_mz} vector. All peaks within the \code{thresh}
+of one another are binned, and only the maximum intensity of that bin is
+filled into the mapped spectrum.
+}
+\section{Warning}{
+ It is important that the values for \code{thresh} and
+\code{dig} are equal to that of \code{res} and \code{dig} used in
[…]

[…]
