### 2021-02-02 191e6b7
Rendered new documentation using package 'roxygen2'.
