### 2021-02-02 ddf7d36
Rendered new documentation using package 'roxygen2'.
