+++ R/avgSpectra.R
@@ -3,10 +3,13 @@
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 =======
 >>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 =======
 >>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
+=======
+>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 ##' Average Spectral Replicates
 ##' 
 ##' Combines spectral replicates either by averaging (method = "mean") or
@@ -39,6 +42,7 @@
 ##' 
 
 
+<<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
 # --------------------------------------------------------------------------------------------
@@ -75,6 +79,12 @@
 # Author: Kristen Yeh, Sophie Castel
 # Title: subMALDI Average Spectra
 >>>>>>> 3f68890... Updated avgSpectra
+=======
+# --------------------------------------------------------------------------------------------
+# Date: January 27, 2021
+# Author: Kristen Yeh, Sophie Castel
+# Title: subMALDI: Average Spectral Replicates
+>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 # --------------------------------------------------------------------------------------------
 
 # Average intensities across rows per sample
