+++ R/normSpectra.R
@@ -3,10 +3,13 @@
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 =======
 >>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 =======
 >>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
+=======
+>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 ##' Normalize Spectral Data
 ##' 
 ##' Normalize spectral data to a common scale using several different methods.
@@ -89,6 +92,7 @@
 ##' 
 ##' 
 
+<<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
 # -----------------------------------------------------------------------
@@ -298,17 +302,13 @@ normSpectra <- function(dat, mass_dat, method = NULL, norm_mz = NULL, upper = NU
   return(r)
 >>>>>>> 0c7b870... Fixed object return bug
 =======
+=======
+>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 # -----------------------------------------------------------------------
 # Last Updated: January 28, 2021
 # Author: Kristen Yeh, Sophie Castel
-# Title: subMALDI: Umbrella Normalization Function
+# Title: subMALDI: Normalize Spectral Data
 # -----------------------------------------------------------------------
-
-
-# ------------
-# normSpectra
-# ------------
-
 
 normSpectra <- function(dat, mass_dat, method = NULL, norm_mz = NULL, upper = NULL, lower = NULL,
                         spectra_cols = NULL, showHI = FALSE){
@@ -376,6 +376,7 @@ normSpectra <- function(dat, mass_dat, method = NULL, norm_mz = NULL, upper = NU
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 # -----------------------------------------------------------------------
 =======
 >>>>>>> f70720d... TIC norm added
@@ -389,4 +390,7 @@ normSpectra <- function(dat, mass_dat, method = NULL, norm_mz = NULL, upper = NU
 >>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
 =======
 >>>>>>> f70720d... TIC norm added
+=======
+# -----------------------------------------------------------------------
+>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 
