+++ R/find_max.R
@@ -1,10 +1,13 @@
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 =======
 >>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 =======
 >>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
+=======
+>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 ##' Find Peak Maxima
 ##' 
 ##' Analyzes spectral data and returns a list of the most intense peak in each
@@ -33,6 +36,7 @@
 ##' ## Find maxima of four spectra
 ##' find_max(dat = Master, mass_dat = "full_mz", spectra_cols = c("Blank1", "Before1", "After1", "After2")
 ##'
+<<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
 =======
@@ -129,11 +133,15 @@ find_max <- function (dat, mass_dat, spectra_cols){
 
 =======
 # Title: subMALDI Peak Maximum Functions
+=======
+
+# -----------------------------------------------------------------------
+# Last Updated: January 27, 2021
+# Author: Kristen Yeh, Sophie Castel
+# Title: subMALDI: Find Peak Maxima
+>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 # -----------------------------------------------------------------------
 
-# -----------------
-# FIND PEAK MAXIMA
-# -----------------
 
 # Find peak maximum and associated m/z of spectra
 find_max <- function (dat, mass_dat, spectra_cols){
@@ -175,6 +183,7 @@ find_max <- function (dat, mass_dat, spectra_cols){
 # Author: Kristen Yeh, Sophie Castel
 # Title: subMALDI: Find Peak Maxima
 # -----------------------------------------------------------------------
+<<<<<<< HEAD
 
 
 <<<<<<< HEAD
@@ -294,3 +303,5 @@ find_max_set <- function(dat, mass_dat, spectra_cols){
 >>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 =======
 >>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
+=======
+>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.

[…]
