### 2021-02-02 5dfbec6
Rendered new documentation using package 'roxygen2'.
Former-commit-id: 8fda89a27654486d9f30fc6cb1edef979d6313c1
Former-commit-id: 8d235e5ef38205a705ff937652d30f87c02af3ff
