### 2021-02-02 e6b0cdb
Rendered new documentation using package 'roxygen2'.
Former-commit-id: ddf7d36a5a7eca3a54cd02ec15aa41b5dc4a3d90
Former-commit-id: bfed3b3570edc5bfc015e558d4bbee33b8dabe0f
