+++ man/baselineCorr.Rd
@@ -3,6 +3,7 @@
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/baselineCorr.R
 \name{baselineCorr}
@@ -62,6 +63,13 @@ Offers three different methods for baseline correction of raw spectral data. Met
 >>>>>>> 5831098... Added pre-processing vignette
 }
 >>>>>>> 4bbf2ad... Baseline correction added
+=======
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/baselineCorr.R
+\name{baselineCorr}
+\alias{baselineCorr}
+\title{Baseline Correction}
+>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 \usage{
 baselineCorr(dat, mass_dat, intensity_dat, method = NULL, n = NULL)
 }
@@ -71,10 +79,13 @@ baselineCorr(dat, mass_dat, intensity_dat, method = NULL, n = NULL)
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 =======
 >>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 =======
 >>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
+=======
+>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 \item{dat}{The name of the spectral data frame, containing \code{m/z} data
 in the first column and spectral intensity data in subsequent columns.}
 
@@ -92,6 +103,7 @@ LOESS curve fitting.}
 \item{n}{Single odd numeric value. If \code{method = "linear"}, the size of
 the window that should be used for linear interpolation of the spectral
 baseline.}
+<<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
 }
@@ -238,33 +250,48 @@ raw spectral signal.  } }
   \item{n}{ Single odd numeric value. If \code{method = "linear"}, the size of the window that should be used for linear interpolation of the spectral baseline. }
 }
 \value{ Returns a new data frame containing the baseline corrected spectral data.
+=======
+>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 }
-\references{
-https://github.com/wesleyburr/subMaldi
-(1) Yang, C., He, Z. & Yu, W. Comparison of public peak detection algorithms for MALDI mass spectrometry data analysis. BMC Bioinformatics 10, 4 (2009). https://doi.org/10.1186/1471-2105-10-4
+\value{
+Returns a new data frame containing the baseline corrected spectral
+data.
 }
-\author{
-Kristen Yeh <kristenyeh@trentu.ca>
-Wesley Burr <wburr@trentu.ca>
+\description{
+Offers three different methods for baseline correction of raw spectral
+data. Methods include monotone minimum, linear interpolation, and LOESS
+curve fitting.
 }
-
 \section{Methods}{
-  \describe{
-    \item{monotone_min}{ Identifies valleys in the spectrum by evaluating the slopes between adjacent intensity values. Moving from left to right, if slope of point A is greater than 0, the nearest point B with slope of less than 0 is located. Points in between A and B are recorded as potential peaks. If slope of A < 0, nearest point B with slope > 0 is located. Points between A and B are recorded as valleys. Each intensity value in the spectrum is then subtracted by the intensity of the nearest valley, correcting any irregularities in the baseline of the data. }
[…]
