### 2021-02-02 51a20e2
Rendered new documentation using package 'roxygen2'.
Former-commit-id: 191e6b7efb3003a6a2f87074357185f61b530041
Former-commit-id: 00fbc4ff787eed02569050c4abe59b6ebe9bd5b7
