+++ man/createSpecDF.Rd
@@ -1,23 +1,45 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/createSpecDF.R
 \name{createSpecDF}
 \alias{createSpecDF}
-\title{
-Create Empty Spectral Data Frame
-}
-\description{
-Creates an empty data frame to be used for mapping spectral data to a vector of \emph{m/z} values. The vector of \emph{m/z} values is stored in the first column and can be called by "\code{full_mz}".
-}
+\title{Create Empty Spectral Data Frame}
 \usage{
-createSpecDF(min_mz = 53.76, max_mz = 1100.00, res = 0.0001, dig = 4)
+createSpecDF(min_mz = 53.76, max_mz = 1100, res = 1e-04, dig = 4)
 }
 \arguments{
-  \item{min_mz}{Single numeric value; minimum \emph{m/z} value of the observed range. Default = 53.76.}
-  \item{max_mz}{Single numeric value; upper end of \emph{m/z} range observed in spectra. Default = 1100.}
-  \item{res}{The resolution of peaks; a single numeric value indicating the step size from the minimum to maximum \emph{m/z} bin. Also known as bin width. Default = 0.0001.}
-  \item{dig}{Number of decimal places to round the \emph{m/z} vector to; required for \emph{m/z} values to match full \emph{m/z} vector. Default = 4.}
-}
+\item{min_mz}{Single numeric value; minimum \emph{m/z} value of the
+observed range. Default = 53.76.}
 
+\item{max_mz}{Single numeric value; upper end of \emph{m/z} range observed
+in spectra. Default = 1100.}
+
+\item{res}{The resolution of peaks; a single numeric value indicating the
+step size from the minimum to maximum \emph{m/z} bin. Also known as bin
+width. Default = 0.0001.}
+
+\item{dig}{Number of decimal places to round the \emph{m/z} vector to;
+required for \emph{m/z} values to match full \emph{m/z} vector. Default =
+4.}
+}
 \value{
-Returns a data frame which can be used to map irregularly spaced spectral data to a set range of \emph{m/z} values, contained in the first column of the frame.
+Returns a data frame which can be used to map irregularly spaced
+spectral data to a set range of \emph{m/z} values, contained in the first
+column of the frame.
+}
+\description{
+Creates an empty data frame to be used for mapping spectral data to a
+vector of \emph{m/z} values. The vector of \emph{m/z} values is stored in
+the first column and can be called by "\code{full_mz}".
+}
+\examples{
+
+## Creating an empty spectrum with an m/z range of 500 to 2000 m/z, with a step size of 0.001
+spec_df <- createSpecDF(min_mz = 500, max_mz = 2000, res = 0.001, dig = 3)
+
+## Creating an empty spectrum with an m/z range of 100 to 1000, 
+# with a step size of 0.0001 and sample names
+spec_df <- createSpecDF(min_mz = 100, max_mz = 1000, res = 0.0001, dig = 4)
+
 }
 \references{
 https://github.com/wesleyburr/subMaldi
@@ -25,16 +47,5 @@ https://github.com/wesleyburr/subMaldi
 \author{
 Kristen Yeh <kristenyeh@trentu.ca>
 }
-
-
-\examples{
-## Creating an empty spectrum with an m/z range of 500 to 2000 m/z, with a step size of 0.001
-spec_df <- createSpecDF(min_mz = 500, max_mz = 2000, res = 0.001, dig = 3)
-
-## Creating an empty spectrum with an m/z range of 100 to 1000, 
-# with a step size of 0.0001 and sample names
-spec_df <- createSpecDF(min_mz = 100, max_mz = 1000, res = 0.0001, dig = 4)
-}
-
-\keyword{ methods }
-\keyword{ array }
+\keyword{array}
+\keyword{methods}

[…]
