### 2021-02-02 7de72bf
Rendered new documentation using package 'roxygen2'.
Former-commit-id: ae03a50244c11fd3aa24f96f6a4ad4ed44eaf4d8
Former-commit-id: 5c947ad01bb1b7d1fdad992b0ef6ef87ef50e7d9
