### 2021-02-02 8fda89a
Rendered new documentation using package 'roxygen2'.
