+++ R/avgSpectra.R
@@ -1,7 +1,39 @@
+##' Average Spectral Replicates
+##' 
+##' Combines spectral replicates either by averaging (method = "mean") or
+##' summing (method = "sum") the intensity values across each row representing a
+##' mass-to-charge value in \code{full_mz}.
+##' 
+##' 
+##' @param dat The mapped spectral data frame, containing \code{full_mz} in the
+##' first column.
+##' @param method A character string; the method used to combine the spectra.
+##' Methods include "sum" and "mean". Default = "mean."
+##' @param spectra_cols A character vector; the names of the column in
+##' \code{dat} containing the intensity data for the spectra-of-interest.
+##' @return Returns a new column in the input data frame containing the
+##' averaged intensity data.
+##' @author Kristen Yeh <kristenyeh@@trentu.ca> Wesley Burr <wburr@@trentu.ca>
+##' Sophie Castel <sophie.castel@@ontariotechu.net>
+##' @references https://github.com/wesleyburr/subMaldi
+##' @keywords manip methods
+##' @examples
+##' 
+##' ## Load sample dataset "Master.rda"
+##' data("Master")
+##' 
+##' ## Average blank spectrum 1 and 2 using the method "mean"
+##' ex <- avgSpectra(Master, method = "mean", spectra_cols = c("Blank1", "Blank2"))
+##' 
+##' ## Average blank spectrum 1 and 2 using the method "sum"
+##' ex <- avgSpectra(Master, method = "sum", spectra_cols = c("Blank1", "Blank2"))
+##' 
+
+
 # --------------------------------------------------------------------------------------------
 # Date: January 27, 2021
 # Author: Kristen Yeh, Sophie Castel
-# Title: subMALDI Average Spectra
+# Title: subMALDI: Average Spectral Replicates
 # --------------------------------------------------------------------------------------------
 
 # Average intensities across rows per sample
