+++ man/mapSpectrum.Rd
@@ -1,42 +1,61 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/mapSpectrum.R
 \name{mapSpectrum}
 \alias{mapSpectrum}
-\title{
-Map Spectrum to \emph{m/z} Vector
-}
-\description{
-Fills in the columns of the empty data frame created using \code{createSpecDF}. Mass to charge values from pairwise spectral data are compared to the \code{full_mz} vector. All peaks within the \code{thresh} of one another are binned, and only the maximum intensity of that bin is filled into the mapped spectrum.
-}
+\title{Map Spectrum to \emph{m/z} Vector}
 \usage{
-mapSpectrum(dat, massCol, intenseCol, dig = 4, thresh = 1e-04, spec_df, colName)
+mapSpectrum(
+  dat,
+  massCol,
+  intenseCol,
+  dig = 4,
+  thresh = 1e-04,
+  spec_df,
+  colName
+)
 }
 \arguments{
-  \item{dat}{A pairwise data frame containing your spectral data. Should contain two columns: one for \emph{m/z} and one for intensity.}
-  \item{massCol}{A character string; the name of the \emph{m/z} column in \code{dat}.}
-  \item{intenseCol}{A character string; the name of the intensity column in \code{dat}.}
-  \item{dig}{Number of decimal places to round the \emph{m/z} data to; must match the same value used in \code{createSpecDF} in order for the columns to fill. Default = 4.}
-  \item{thresh}{Single numeric value; all \emph{m/z} values within \code{thresh} of each other are binned under that with the maximum intensity. Default = 5e-5.}
-  \item{spec_df}{An empty data frame created using \code{createSpecDF}.}
-  \item{colName}{A character string; the name of the column that should be filled with the spectral data.}
-}
+\item{dat}{A pairwise data frame containing your spectral data. Should
+contain two columns: one for \emph{m/z} and one for intensity.}
 
+\item{massCol}{A character string; the name of the \emph{m/z} column in
+\code{dat}.}
+
+\item{intenseCol}{A character string; the name of the intensity column in
+\code{dat}.}
+
+\item{dig}{Number of decimal places to round the \emph{m/z} data to; must
+match the same value used in \code{createSpecDF} in order for the columns
+to fill. Default = 4.}
+
+\item{thresh}{Single numeric value; all \emph{m/z} values within
+\code{thresh} of each other are binned under that with the maximum
+intensity. Default = 5e-5.}
+
+\item{spec_df}{An empty data frame created using \code{createSpecDF}.}
+
+\item{colName}{A character string; the name of the column that should be
+filled with the spectral data.}
+}
 \value{
-Returns a vector that is used to fill in the \code{colName} column of the mapped spectral data frame.
+Returns a vector that is used to fill in the \code{colName} column
+of the mapped spectral data frame.
 }
-\references{
-https://github.com/wesleyburr/subMaldi
+\description{
+Fills in the columns of the empty data frame created using
+\code{createSpecDF}. Mass to charge values from pairwise spectral data are
+compared to the \code{full_mz} vector. All peaks within the \code{thresh}
+of one another are binned, and only the maximum intensity of that bin is
+filled into the mapped spectrum.
 }
-\author{
-Kristen Yeh <kristenyeh@trentu.ca>
-Wesley Burr <wburr@trentu.ca>
+\section{Warning}{
+ It is important that the values for \code{thresh} and
[…]

[…]
