### 2021-02-02 77159a9
Rendered new documentation using package 'roxygen2'.
