+++ R/baselineCorr.R
@@ -1,8 +1,11 @@
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 =======
 >>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
+=======
+>>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
 ##' Baseline Correction
 ##' 
 ##' Offers three different methods for baseline correction of raw spectral
@@ -70,12 +73,15 @@
 
 
 <<<<<<< HEAD
+<<<<<<< HEAD
 # ----------------------------------------------------------------------------
 # Last Updated: February 3, 2021
 # Author: Kristen Yeh, Sophie Castel
 =======
 =======
 >>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
+=======
+>>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
 # ----------------------------------------------------------------------------
 <<<<<<< HEAD
 # Last Updated: July 29, 2020

[…]
