### 2021-02-02 c0c130f
Rendered new documentation using package 'roxygen2'.
