+++ R/baselineCorr.R
@@ -1,4 +1,7 @@
 <<<<<<< HEAD
+<<<<<<< HEAD
+=======
+>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 ##' Baseline Correction
 ##' 
 ##' Offers three different methods for baseline correction of raw spectral
@@ -65,10 +68,13 @@
 ##' 
 
 
+<<<<<<< HEAD
 # ----------------------------------------------------------------------------
 # Last Updated: February 3, 2021
 # Author: Kristen Yeh, Sophie Castel
 =======
+=======
+>>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 # ----------------------------------------------------------------------------
 # Last Updated: July 29, 2020
 # Author: Kristen Yeh
