### 2021-02-02 4bf3435
Rendered new documentation using package 'roxygen2'.
Former-commit-id: 77159a9dd18631a77e3ae8a05c19f0616cdd07c4
Former-commit-id: f7b0a2caef81669a5e56a9b45a4669c088a9a394
