+++ man/normSpectra.Rd
@@ -1,6 +1,7 @@
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/normSpectra.R
 \name{normSpectra}
@@ -189,57 +190,102 @@ rearranged in their original order. } }
 
 >>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 =======
+=======
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/normSpectra.R
+>>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
 \name{normSpectra}
 \alias{normSpectra}
-\title{
-Normalize Spectral Data
+\title{Normalize Spectral Data}
+\usage{
+normSpectra(
+  dat,
+  mass_dat,
+  method = NULL,
+  norm_mz = NULL,
+  upper = NULL,
+  lower = NULL,
+  spectra_cols = NULL,
+  showHI = FALSE
+)
+}
+\arguments{
+\item{dat}{The name of the spectral data frame, containing \code{m/z} data
+in the first column and spectral intensity data in subsequent columns.}
+
+\item{mass_dat}{A character string; the name of the column in \code{dat}
+containing the \emph{m/z} data for the spectrum.}
+
+\item{method}{A character string; the normalization method that should be
+used to process the data. See 'Methods' below for list of methods. Default
+= NULL.}
+
+\item{norm_mz}{Numeric. If \code{method = "custom"}, the \emph{m/z} peak to
+which the spectral intensity should be normalized to. Value should have the
+same number of decimal places as the \emph{m/z} data in \code{dat}. If
+\code{method = "custom_imprecise", this value must be given as a character
+string of numbers.}}
+
+\item{upper}{Numeric. If \code{method = "stdev"}, the upper \emph{m/z}
+bound of the noise region of the spectrum.}
+
+\item{lower}{Numeric. If \code{method = "stdev"}, the lower \emph{m/z}
+bound of the noise region of the spectrum.}
+
+\item{spectra_cols}{A character vector; the names of the column in
+\code{dat} containing the intensity data for the spectra-of-interest.}
+
+\item{showHI}{Logical. To be used with \code{method = "custom"}. If
+\code{TRUE}, intensity values greater than \code{norm_mz} will be kept. If
+\code{FALSE}, any peaks with intensity greater than \code{norm_mz} will be
+truncated to the intensity of \code{norm_mz}. Default = FALSE.}
+}
+\value{
+Returns a new data frame including the original \emph{m/z} data and
+normalized intensity data.
 }
 \description{
 Normalize spectral data to a common scale using several different methods.
 }
-\usage{
-normSpectra(dat, mass_dat, method = NULL, norm_mz = NULL, 
-            upper = NULL, lower = NULL, spectra_cols,
-            showHI = FALSE)
-}
-\arguments{
-  \item{dat}{The name of the spectral data frame, containing \code{m/z} data in the first column and spectral intensity data in subsequent columns.}
-  \item{mass_dat}{A character string; the name of the column in \code{dat} containing the \emph{m/z} data for the spectrum.}
-  \item{method}{A character string; the normalization method that should be used to process the data. See 'Methods' below for list of methods. Default = NULL.}
-  \item{norm_mz}{Numeric. If \code{method = "custom"}, the \emph{m/z} peak to which the spectral intensity should be normalized to. Value should have the same number of decimal places as the \emph{m/z} data in \code{dat}. If \code{method = "custom_imprecise", this value must be given as a character string of numbers.}}
-  \item{lower}{Numeric. If \code{method = "stdev"}, the lower \emph{m/z} bound of the noise region of the spectrum.}
[…]
