+++ man/createSpecDF.Rd
@@ -6,6 +6,7 @@
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 \title{Create Empty Spectral Data Frame}
 =======
 %- Also NEED an '\alias' for EACH other topic documented here.
@@ -23,10 +24,14 @@ Creates an empty data frame to be used for mapping spectral data to a vector of 
 =======
 \title{Create Empty Spectral Data Frame}
 >>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
+=======
+\title{Create Empty Spectral Data Frame}
+>>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
 \usage{
 createSpecDF(min_mz = 53.76, max_mz = 1100, res = 1e-04, dig = 4)
 }
 \arguments{
+<<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
@@ -99,10 +104,32 @@ column of the frame.
 }
 <<<<<<< HEAD
 <<<<<<< HEAD
+=======
+\item{min_mz}{Single numeric value; minimum \emph{m/z} value of the
+observed range. Default = 53.76.}
+
+\item{max_mz}{Single numeric value; upper end of \emph{m/z} range observed
+in spectra. Default = 1100.}
+
+\item{res}{The resolution of peaks; a single numeric value indicating the
+step size from the minimum to maximum \emph{m/z} bin. Also known as bin
+width. Default = 0.0001.}
+
+\item{dig}{Number of decimal places to round the \emph{m/z} vector to;
+required for \emph{m/z} values to match full \emph{m/z} vector. Default =
+4.}
+}
+\value{
+Returns a data frame which can be used to map irregularly spaced
+spectral data to a set range of \emph{m/z} values, contained in the first
+column of the frame.
+}
+>>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
 \description{
 Creates an empty data frame to be used for mapping spectral data to a
 vector of \emph{m/z} values. The vector of \emph{m/z} values is stored in
 the first column and can be called by "\code{full_mz}".
+<<<<<<< HEAD
 =======
 \note{
 }
@@ -130,6 +157,9 @@ the first column and can be called by "\code{full_mz}".
 
 
 >>>>>>> 2931ed6... Added pre-processing vignette
+=======
+}
+>>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
 \examples{
 
 ## Creating an empty spectrum with an m/z range of 500 to 2000 m/z, with a step size of 0.001
@@ -138,6 +168,7 @@ spec_df <- createSpecDF(min_mz = 500, max_mz = 2000, res = 0.001, dig = 3)
 ## Creating an empty spectrum with an m/z range of 100 to 1000, 
 # with a step size of 0.0001 and sample names
 spec_df <- createSpecDF(min_mz = 100, max_mz = 1000, res = 0.0001, dig = 4)
+<<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
@@ -179,3 +210,15 @@ Kristen Yeh <kristenyeh@trentu.ca>
 \keyword{ methods }
 \keyword{ array }
 >>>>>>> 14c6370... Updated documentation
+=======
+
+}
+\references{
+https://github.com/wesleyburr/subMaldi
+}
+\author{
+Kristen Yeh <kristenyeh@trentu.ca>
+}
+\keyword{array}
+\keyword{methods}
+>>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.

[…]
