### 2021-02-02 ae03a50
Rendered new documentation using package 'roxygen2'.
