### 2021-02-02 02177be
Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
Former-commit-id: 17562b6c502777952e10f6439919f59939455a6d
Former-commit-id: 888028dbd0f2e7cafab222b6b4f289fe10af1a91
