+++ R/plotSpectra.R
@@ -2,17 +2,21 @@
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 ##' Plot and Compare Spectra
 ##' 
 ##' Plot single or multiple spectra. For multiple spectra, a grid layout can be called using 'nrows'.
 =======
 =======
 >>>>>>> 046132a... Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
+=======
+>>>>>>> 6f77461... Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
 ##' Plot and Compare Multiple Spectra
 ##' 
 ##' Plots multiple spectra on the same \emph{x}-axis scale for easier
 ##' comparison. Allows for 2 to 6 spectra per plot, stacked on top of one
 ##' another.
+<<<<<<< HEAD
 <<<<<<< HEAD
 >>>>>>> 6f77461... Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
 =======
@@ -27,12 +31,15 @@
 ##' 
 ##' Plot single or multiple spectra. For multiple spectra, a grid layout can be called using 'nrows'.
 >>>>>>> 3eeaddf... Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
+=======
+>>>>>>> 6f77461... Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
 ##' 
 ##' 
 ##' @param dat The name of the spectral data frame, containing the \emph{m/z}
 ##' data in the first column.
 ##' @param mass_dat A character string; the name of the column in \code{dat}
 ##' containing the \emph{m/z} data.
+<<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
@@ -48,6 +55,8 @@
 =======
 =======
 >>>>>>> 046132a... Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
+=======
+>>>>>>> 6f77461... Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
 ##' @param spec1 A character string; the name of the column in \code{dat}
 ##' containing the intensity data for the first spectrum that is to be plotted.
 ##' @param spec2 A character string; the name of the column in \code{dat}
@@ -77,6 +86,7 @@
 ##' @param thresh Single numeric value (0-100); the threshold of peak
 ##' itensities which should be labelled.
 <<<<<<< HEAD
+<<<<<<< HEAD
 >>>>>>> 6f77461... Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
 =======
 ##' @param spectra_cols A character string; the name(s) of the column(s) in \code{dat}
@@ -100,6 +110,8 @@
 ##' @param thresh Single numeric value (0-100); the threshold of peak
 ##' intensities which should be labeled.
 >>>>>>> 3eeaddf... Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
+=======
+>>>>>>> 6f77461... Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
 ##' @param lbls Logical. If \code{lbls = TRUE}, labels indicating the
 ##' \emph{m/z} value of each peak maxima within the indicated \code{span} will
 ##' be included in the output plot. If \code{lbls = FALSE}, labels are not
@@ -120,6 +132,7 @@
[…]
