### 2021-02-02 f1e228f
Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
Former-commit-id: d7e10a6b70b86a4f3899296f1bd7f79089e6b24a
Former-commit-id: 9d1737fa84350ea828207c382be271bfaee0fea0
