+++ R/plotSpectra.R
@@ -1,12 +1,21 @@
+<<<<<<< HEAD
 ##' Plot and Compare Spectra
 ##' 
 ##' Plot single or multiple spectra. For multiple spectra, a grid layout can be called using 'nrows'.
+=======
+##' Plot and Compare Multiple Spectra
+##' 
+##' Plots multiple spectra on the same \emph{x}-axis scale for easier
+##' comparison. Allows for 2 to 6 spectra per plot, stacked on top of one
+##' another.
+>>>>>>> 6f77461... Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
 ##' 
 ##' 
 ##' @param dat The name of the spectral data frame, containing the \emph{m/z}
 ##' data in the first column.
 ##' @param mass_dat A character string; the name of the column in \code{dat}
 ##' containing the \emph{m/z} data.
+<<<<<<< HEAD
 ##' @param spectra_cols A character string; the name(s) of the column(s) in \code{dat}
 ##' containing the intensity data for the spectra-of-interest.
 ##' @param palette A character element; the RColorBrewer palette to use. See below for available palettes. 
@@ -15,6 +24,36 @@
 ##' label. Default = 5 (ignores two peak maxima on either side of each label).
 ##' @param thresh Single numeric value (0-100); the threshold of peak
 ##' intensities which should be labeled.
+=======
+##' @param spec1 A character string; the name of the column in \code{dat}
+##' containing the intensity data for the first spectrum that is to be plotted.
+##' @param spec2 A character string; the name of the column in \code{dat}
+##' containing the intensity data for the second spectrum to be plotted.
+##' @param spec3 A character string; the name of the column in \code{dat}
+##' containing the intensity data for the third spectrum to be plotted.
+##' @param spec4 A character string; the name of the column in \code{dat}
+##' containing the intensity data for the fourth spectrum to be plotted.
+##' @param spec5 A character string; the name of the column in \code{dat}
+##' containing the intensity data for the fifth spectrum to be plotted.
+##' @param spec6 A character string; the name of the column in \code{dat}
+##' containing the intensity data for the sixth spectrum to be plotted.
+##' @param colour1 A character string indicating the colour the first spectrum
+##' should be plotted in.
+##' @param colour2 A charater string indicating the colour of the second
+##' spectrum in the plot.
+##' @param colour3 A charater string indicating the colour of the third
+##' spectrum in the plot.
+##' @param colour4 A charater string indicating the colour of the fourth
+##' spectrum in the plot.
+##' @param colour5 A charater string indicating the colour of the fifth
+##' spectrum in the plot.
+##' @param colour6 A charater string indicating the colour of the sixth
+##' spectrum in the plot.
+##' @param span Single numeric value; the span of peak maximas in between each
+##' label. Default = 5 (ignores two peak maxima on either side of each label).
+##' @param thresh Single numeric value (0-100); the threshold of peak
+##' itensities which should be labelled.
+>>>>>>> 6f77461... Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
 ##' @param lbls Logical. If \code{lbls = TRUE}, labels indicating the
 ##' \emph{m/z} value of each peak maxima within the indicated \code{span} will
 ##' be included in the output plot. If \code{lbls = FALSE}, labels are not
@@ -31,13 +70,18 @@
 ##' range.
 ##' @param max_I Single numeric value; upper end of the intensity range
 ##' observed in spectra.
+<<<<<<< HEAD
[…]
