### 2021-02-02 80ee810
Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
