+++ R/subMALDI.R
@@ -1,170 +1,0 @@
-# -----------------------------------------------------------------------
-# Last Updated: May 25, 2020
-# Author: Kristen Yeh
-# Title: subMALDI Base Functions
-# -----------------------------------------------------------------------
-
-
-# --------------------------
-# IMPORT DATA TO DATA FRAME
-# --------------------------
-
-# Import .csv mass spectra files from a directory and export as .rda's for use in data frame
-# Basically just a quick and easy way to turn .csv files into individual data frames per spectrum
-readcsvSpec <- function(spec_file, massCol, intenseCol){
-  temp <- read.csv(file = spec_file)
-  mass <- select(temp, massCol)
-  Intensity <- select(temp, intenseCol)
-  mass <- unlist(cbind(mass))
-  Intensity <- unlist(cbind(Intensity))
-  spec <- data.frame(mass, Intensity)
-  return(spec)
-}
-
-readcsvDir <- function(direct, massCol, intenseCol, output){
-  all_files <- list.files(path = direct, pattern = ".csv")
-  for(j in 1:length(all_files)) {
-    temp <- read.csv(  paste0(direct, all_files[ j ]) )
-    mass <- select(temp, massCol)
-    Intensity <- select(temp, intenseCol)
-    mass <- unlist(cbind(mass))
-    Intensity <- unlist(cbind(Intensity))
-    spec <- data.frame(mass, Intensity)
-
-    new_name <- strsplit(all_files[ j ], "\\.")[[1]][1]
-    assign(x = new_name, spec)
-    save(file = paste0(output, new_name, ".rda"), list = new_name)
-  }
-}
-
-
-# -----------------------------------------------------------------------
-
-
-# ------------------------
-# MAKE EMPTY DATA FRAME
-# ------------------------
-
-
-createSpecDF <- function(min_mz = 53.76, max_mz = 1100, res = 0.0001, dig = 4) {
-  full_mz <- seq(min_mz, max_mz, by = res)
-  spec_df <- data.frame(full_mz,
-                        Sample = 0)
-  spec_df$full_mz <- round(spec_df$full_mz, digits = dig)
-  return(spec_df)
-}
-
-
-# -----------------------------------------------------------------------
-
-
-# ---------------------------------
-# MAP SPECTRUM TO EMPTY DATA FRAME
-# ---------------------------------
-
-mapSpectrum <- function(dat, massCol, intenseCol, dig = 4, thresh = 1e-4, spec_df, colName) {
-  # Sanity checks
-  stopifnot( is.character(massCol), is.character(intenseCol),
-             massCol %in% names(dat), intenseCol %in% names(dat) )
-  stopifnot(colName %in% names(spec_df)) 
-
-  # Round m/z data from dat to match full_mz
-  dat <- round(dat, digits = dig)
-
-  # Clean code by subsetting to vectors
-  mass <- dat[[massCol]]
-  intense <- dat[[intenseCol]]
-
-  # Quantify the spaces between m/z values, make vector of all differences greater than threshold
-  diffs <- mass[-1] - mass[-length(mass)]
-  too_close <- which(diffs < thresh)
-  throw_away <- c()
-
-  if(length(too_close) > 0) { # not only isolated peaks
-    # fix the 'too close together' problem: identify local maximums
-    dupes <- which(duplicated(mass))  # identifies the duplicates, but not
-    # the points they're duplicates _of_
-    if(length(dupes) > 0) { # found some ...
-      for(j in 1:length(dupes)) { # cycle through
-        mz.val <- mass[dupes[j]]  # grab the mz that's duplicated
-        all_dupes_tmp <- which(mass == mz.val)  # find all mz's that match
[…]
