### 2021-02-02 fb468dd
Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
Former-commit-id: 97f50ca8a4266da7044b98391af64b9f6b826350
Former-commit-id: bac315d7332704a2ad136ced23eabd1fbcc201e1
