### 2021-02-02 33cf971
Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
Former-commit-id: 80ee810993e5fa45af7ae58e3fdf000e762ea41a
Former-commit-id: 897f9ac2233c264f937024ebb40a3669f91aebbc
