+++ R/plotSpectra.R
@@ -1,26 +1,33 @@
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 ##' Plot and Compare Spectra
 ##' 
 ##' Plot single or multiple spectra. For multiple spectra, a grid layout can be called using 'nrows'.
 =======
+=======
+>>>>>>> 046132a... Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
 ##' Plot and Compare Multiple Spectra
 ##' 
 ##' Plots multiple spectra on the same \emph{x}-axis scale for easier
 ##' comparison. Allows for 2 to 6 spectra per plot, stacked on top of one
 ##' another.
+<<<<<<< HEAD
 >>>>>>> 6f77461... Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
 =======
 ##' Plot and Compare Spectra
 ##' 
 ##' Plot single or multiple spectra. For multiple spectra, a grid layout can be called using 'nrows'.
 >>>>>>> cd09eca... Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
+=======
+>>>>>>> 046132a... Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
 ##' 
 ##' 
 ##' @param dat The name of the spectral data frame, containing the \emph{m/z}
 ##' data in the first column.
 ##' @param mass_dat A character string; the name of the column in \code{dat}
 ##' containing the \emph{m/z} data.
+<<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
 ##' @param spectra_cols A character string; the name(s) of the column(s) in \code{dat}
@@ -32,6 +39,8 @@
 ##' @param thresh Single numeric value (0-100); the threshold of peak
 ##' intensities which should be labeled.
 =======
+=======
+>>>>>>> 046132a... Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
 ##' @param spec1 A character string; the name of the column in \code{dat}
 ##' containing the intensity data for the first spectrum that is to be plotted.
 ##' @param spec2 A character string; the name of the column in \code{dat}
@@ -60,6 +69,7 @@
 ##' label. Default = 5 (ignores two peak maxima on either side of each label).
 ##' @param thresh Single numeric value (0-100); the threshold of peak
 ##' itensities which should be labelled.
+<<<<<<< HEAD
 >>>>>>> 6f77461... Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
 =======
 ##' @param spectra_cols A character string; the name(s) of the column(s) in \code{dat}
@@ -71,6 +81,8 @@
 ##' @param thresh Single numeric value (0-100); the threshold of peak
 ##' intensities which should be labeled.
 >>>>>>> cd09eca... Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
+=======
+>>>>>>> 046132a... Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
 ##' @param lbls Logical. If \code{lbls = TRUE}, labels indicating the
 ##' \emph{m/z} value of each peak maxima within the indicated \code{span} will
 ##' be included in the output plot. If \code{lbls = FALSE}, labels are not
@@ -89,6 +101,7 @@
 ##' observed in spectra.
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
[…]
