### 2021-02-02 c33300c
Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
Former-commit-id: 6f77461c8e6a9de9b4de197f318a689c005c45f6
Former-commit-id: 2d197690f3431401d851619c1b7c272cfce08b63
