### 2021-02-02 97f50ca
Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
