### 2021-02-03 30ac7d5
Deleted legacy versions of plotting functions. Re-combined into plotSpectra().
