+++ man/plotSpectrum.Rd
@@ -1,89 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/plotSpectrum.R
-\name{plotSpectrum}
-\alias{plotSpectrum}
-\title{Plot a Single Spectrum}
-\usage{
-plotSpectrum(
-  dat,
-  mass_dat,
-  intensity_dat,
-  colour = "black",
-  span = 5,
-  thresh = 0.1,
-  min_mz = min(mass_dat),
-  max_mz = max(mass_dat),
-  min_I = 0,
-  max_I = max(intensity_dat),
-  lbls = FALSE,
-  lbl.fmt = "\%3.4f",
-  x_ticks = 500
-)
-}
-\arguments{
-\item{dat}{The mapped spectral data frame, containing \code{full_mz} in the
-first column.}
-
-\item{mass_dat}{The name of the column in \code{dat} containing the
-\emph{m/z}data for the spectrum that is to be plotted. Should be input as
-dat$mass_column. By default, the mass column in a mapped data frame is
-"full_mz". If our spectral data frame is called my_spec, then our input for
-mass_dat = my_spec$full_mz.}
-
-\item{intensity_dat}{The name of the column in \code{dat} containing the
-intensity data for the spectrum that is to be plotted. Should be input as
-dat$intensity_column, as above for mass_dat.}
-
-\item{colour}{A character string; the name of the colour the spectrum
-should be plotted in.}
-
-\item{span}{Single numeric value; the distance between peak maxima that are
-labelled if lbls = TRUE. Default = 5.}
-
-\item{thresh}{Single numeric value between 0 and 1; the intensity threshold
-of peaks in the plotted spectrum which should be labelled. Default = 0.1.}
-
-\item{min_mz}{Single numeric value; minimum \emph{m/z} value of the
-observed range.}
-
-\item{max_mz}{Single numeric value; upper end of \emph{m/z} range to be
-observed in the plotted spectra.}
-
-\item{min_I}{Single numeric value; minimum intensity value of the spectrum.}
-
-\item{max_I}{Single numeric value; upper end of the intensity range
-observed in spectra.}
-
-\item{lbls}{Logical; should labels indicating the \emph{m/z} of the most
-intense peaks be included in the plot?}
-
-\item{lbl.fmt}{Character string in the format "\%a.bf", where \emph{a} is
-the number of figures displayed to the left of decimal places in labels,
-and \emph{b} is the number of figures displayed to the right of decimal
-places in labels.}
-
-\item{x_ticks}{Single numeric value; the increment between ticks on the
-x-axis.}
-}
-\value{
-Returns a line plot of the input spectrum.
-}
-\description{
-Plots a single spectrum from a mapped spectral data frame.
-}
-\examples{
-
-## Plotting using the sample dataset "Before1.rda"
-data("Before1")
-
-plotSpectrum(dat = Before1, mass_dat = Before1$mass, 
-              intensity_dat = Before1$Intensity)
-}
-\references{
-https://github.com/wesleyburr/subMaldi
-}
-\author{
-Kristen Yeh <kristenyeh@trentu.ca> Wesley Burr <wburr@trentu.ca>
-}
-\keyword{aplot}
-\keyword{methods}
