### 2021-02-03 cd89aeb
Deleted legacy versions of plotting functions. Re-combined into plotSpectra().
