### 2021-02-03 dd8dfdc
Deleted legacy versions of plotting functions. Re-combined into plotSpectra().
Former-commit-id: f540d85a91bb05276ea8dcfe0db9d7616a7ca3fd
Former-commit-id: 43c6fd312ddf1513d814f68d6f0286cbe06cce73
