### 2021-02-03 bc84d8e
Deleted legacy versions of plotting functions. Re-combined into plotSpectra().
Former-commit-id: 30ac7d5d65f74b57bdf65cb6a850cd0d90ef1096
Former-commit-id: a78b8a9c24e409e1067cb3bdda90762fefa1773c
