### 2021-02-03 923ff8f
Deleted legacy versions of plotting functions. Re-combined into plotSpectra().
