### 2021-02-03 e344677
Deleted legacy versions of plotting functions. Re-combined into plotSpectra().
Former-commit-id: 31cca8076a5361d41ce698056f7dd65fb10be02c
Former-commit-id: 5164fed839bf093ac2f5d4f79360c1ecc1ae5dfd
