+++ man/plotgridSpectra.Rd
@@ -1,143 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/plotgridSpectra.R
-\name{plotgridSpectra}
-\alias{plotgridSpectra}
-\title{Plot and Compare Multiple Spectra in a Grid}
-\usage{
-plotgridSpectra(
-  dat,
-  mass_dat,
-  spec1,
-  spec2,
-  spec3,
-  spec4,
-  spec5 = NULL,
-  spec6 = NULL,
-  colour1 = "steelblue1",
-  colour2 = "lightpink1",
-  colour3 = "coral",
-  colour4 = "goldenrod1",
-  colour5 = "gold",
-  colour6 = "seagreen3",
-  span = 5,
-  thresh = 0.1,
-  lbls = FALSE,
-  lbl.fmt = "\%3.4f",
-  columns = 2,
-  min_mz = 0,
-  max_mz = 10000,
-  min_I = 0,
-  max_I = NULL,
-  x_ticks = 500,
-  intensity_scale = "free_y"
-)
-}
-\arguments{
-\item{dat}{The name of the spectral data frame, containing the \emph{m/z}
-data in the first column.}
-
-\item{mass_dat}{A character string; the name of the column in \code{dat}
-containing the \emph{m/z} data.}
-
-\item{spec1}{A character string; the name of the column in \code{dat}
-containing the intensity data for the first spectrum that is to be plotted.}
-
-\item{spec2}{A character string; the name of the column in \code{dat}
-containing the intensity data for the second spectrum to be plotted.}
-
-\item{spec3}{A character string; the name of the column in \code{dat}
-containing the intensity data for the third spectrum to be plotted.}
-
-\item{spec4}{A character string; the name of the column in \code{dat}
-containing the intensity data for the fourth spectrum to be plotted.}
-
-\item{spec5}{A character string; the name of the column in \code{dat}
-containing the intensity data for the fifth spectrum to be plotted. Default
-= NULL.}
-
-\item{spec6}{A character string; the name of the column in \code{dat}
-containing the intensity data for the sixth spectrum to be plotted. Default
-= NULL.}
-
-\item{colour1}{A character string indicating the colour the first spectrum
-should be plotted in.}
-
-\item{colour2}{A charater string indicating the colour of the second
-spectrum in the plot.}
-
-\item{colour3}{A charater string indicating the colour of the third
-spectrum in the plot.}
-
-\item{colour4}{A charater string indicating the colour of the fourth
-spectrum in the plot.}
-
-\item{colour5}{A character string indicating the colour of the fifth
-spectrum in the plot.}
-
-\item{colour6}{A character string indicating the colour of the sixth
-spectrum in the plot.}
-
-\item{span}{Single numeric value; the span of peak maximas in between each
-label. Default = 5 (ignores two peak maxima on either side of each label).}
-
-\item{thresh}{Single numeric value (0-100); the threshold of peak
-itensities which should be labelled. Default = 0.1.}
-
-\item{lbls}{Logical. If \code{lbls = TRUE}, labels indicating the
-\emph{m/z} value of each peak maxima within the indicated \code{span} will
-be included in the output plot. If \code{lbls = FALSE}, labels are not
-shown. Default = FALSE.}
-
-\item{columns}{Single numeric value; either 2 or 3. The number of columns
-spectral plots should be displayed in.}
-
-\item{min_mz}{Single numeric value; minimum \emph{m/z} value of the
-observed range.}
-
-\item{max_mz}{Single numeric value; upper end of \emph{m/z} range observed
-in spectra.}
-
[…]
