### 2021-02-03 5286868
Deleted legacy versions of plotting functions. Re-combined into plotSpectra().
Former-commit-id: ae39e57c66ae8931b0d081ce379743cc430df878
Former-commit-id: e6f8fa1a95fc76cc4b2e4a6164fed749894384c8
