+++ R/plotgridSpectra.R
@@ -1,364 +1,0 @@
-##' Plot and Compare Multiple Spectra in a Grid
-##' 
-##' Plots 4 or 6 spectra in a grid format.
-##' 
-##' 
-##' @param dat The name of the spectral data frame, containing the \emph{m/z}
-##' data in the first column.
-##' @param mass_dat A character string; the name of the column in \code{dat}
-##' containing the \emph{m/z} data.
-##' @param spec1 A character string; the name of the column in \code{dat}
-##' containing the intensity data for the first spectrum that is to be plotted.
-##' @param spec2 A character string; the name of the column in \code{dat}
-##' containing the intensity data for the second spectrum to be plotted.
-##' @param spec3 A character string; the name of the column in \code{dat}
-##' containing the intensity data for the third spectrum to be plotted.
-##' @param spec4 A character string; the name of the column in \code{dat}
-##' containing the intensity data for the fourth spectrum to be plotted.
-##' @param spec5 A character string; the name of the column in \code{dat}
-##' containing the intensity data for the fifth spectrum to be plotted. Default
-##' = NULL.
-##' @param spec6 A character string; the name of the column in \code{dat}
-##' containing the intensity data for the sixth spectrum to be plotted. Default
-##' = NULL.
-##' @param colour1 A character string indicating the colour the first spectrum
-##' should be plotted in.
-##' @param colour2 A charater string indicating the colour of the second
-##' spectrum in the plot.
-##' @param colour3 A charater string indicating the colour of the third
-##' spectrum in the plot.
-##' @param colour4 A charater string indicating the colour of the fourth
-##' spectrum in the plot.
-##' @param colour5 A character string indicating the colour of the fifth
-##' spectrum in the plot.
-##' @param colour6 A character string indicating the colour of the sixth
-##' spectrum in the plot.
-##' @param span Single numeric value; the span of peak maximas in between each
-##' label. Default = 5 (ignores two peak maxima on either side of each label).
-##' @param thresh Single numeric value (0-100); the threshold of peak
-##' itensities which should be labelled. Default = 0.1.
-##' @param lbls Logical. If \code{lbls = TRUE}, labels indicating the
-##' \emph{m/z} value of each peak maxima within the indicated \code{span} will
-##' be included in the output plot. If \code{lbls = FALSE}, labels are not
-##' shown. Default = FALSE.
-##' @param lbl.fmt Character string in the format \code{"%a.bf"}, where
-##' \emph{a} is the number of figures displayed to the left of decimal places
-##' in labels, and \emph{b} is the number of figures displayed to the right of
-##' decimal places in labels.
-##' @param columns Single numeric value; either 2 or 3. The number of columns
-##' spectral plots should be displayed in.
-##' @param min_mz Single numeric value; minimum \emph{m/z} value of the
-##' observed range.
-##' @param max_mz Single numeric value; upper end of \emph{m/z} range observed
-##' in spectra.
-##' @param min_I Single numeric value; minimum intensity value to be plotted.
-##' @param max_I Single numeric value; upper end of the intensity range to be
-##' plotted.
-##' @param x_ticks Single numeric value; the space between ticks on x-axis.
-##' @param intensity_scale A character string; the method that should be used
-##' for determining the y-axis scales for each spectrum. If \code{method =
-##' "free_y"}, each spectrum will be plotted with its own intensity scale. If
-##' \code{method = "fixed"}, each spectrum will be plotted with the y-axis of
-##' the most intense spectrum in the set.
-##' @return Returns a line plot of the input spectra arranged in two columns.
[…]
