### 2021-02-03 6b1075e
Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
Former-commit-id: 0f9a70eff043b111538cff822271b9a61648c390
Former-commit-id: 53a719d5ff3cef347e80b27a50b3e63841e1bebd
