+++ R/plotSpectra.R
@@ -1,6 +1,7 @@
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 ##' Plot and Compare Spectra
 ##' 
 ##' Plot single or multiple spectra. For multiple spectra, a grid layout can be called using 'nrows'.
@@ -21,12 +22,18 @@
 >>>>>>> cd09eca... Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
 =======
 >>>>>>> 046132a... Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
+=======
+##' Plot and Compare Spectra
+##' 
+##' Plot single or multiple spectra. For multiple spectra, a grid layout can be called using 'nrows'.
+>>>>>>> 3eeaddf... Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
 ##' 
 ##' 
 ##' @param dat The name of the spectral data frame, containing the \emph{m/z}
 ##' data in the first column.
 ##' @param mass_dat A character string; the name of the column in \code{dat}
 ##' containing the \emph{m/z} data.
+<<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
@@ -83,6 +90,15 @@
 >>>>>>> cd09eca... Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
 =======
 >>>>>>> 046132a... Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
+=======
+##' @param spectra_cols A character string; the name(s) of the column(s) in \code{dat}
+##' containing the intensity data for the spectra-of-interest.
+##' @param colours A character string indicating the desired colour(s)
+##' @param span Single numeric value; the span of peak maxima in between each
+##' label. Default = 5 (ignores two peak maxima on either side of each label).
+##' @param thresh Single numeric value (0-100); the threshold of peak
+##' intensities which should be labeled.
+>>>>>>> 3eeaddf... Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
 ##' @param lbls Logical. If \code{lbls = TRUE}, labels indicating the
 ##' \emph{m/z} value of each peak maxima within the indicated \code{span} will
 ##' be included in the output plot. If \code{lbls = FALSE}, labels are not
@@ -102,6 +118,7 @@
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 ##' @param x_ticks Single numeric value; the number of ticks on the x-axis.
 ##' @param nrows Single integer value; the number of rows in plot arrangement.
 =======
@@ -114,6 +131,10 @@
 =======
 ##' @param x_ticks Single numeric value; the space between ticks on x-axis.
 >>>>>>> 046132a... Each function in subMALDI now has its own .R file and documentation (if provided previously, except daughter functions).
+=======
+##' @param x_ticks Single numeric value; the number of ticks on the x-axis.
+##' @param nrows Single integer value; the number of rows in plot arrangement.
[…]
