### 2021-02-03 c588c0a
Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
Former-commit-id: 3eeaddfe53a0c247cd057df83610fa07116293f7
Former-commit-id: bccab8510113119068dbc34b55b3c7a45d8efc4c
