+++ man/plotSpectra.Rd
@@ -2,32 +2,23 @@
 % Please edit documentation in R/plotSpectra.R
 \name{plotSpectra}
 \alias{plotSpectra}
-\title{Plot and Compare Multiple Spectra}
+\title{Plot and Compare Spectra}
 \usage{
 plotSpectra(
   dat,
   mass_dat,
-  spec1,
-  spec2 = NULL,
-  spec3 = NULL,
-  spec4 = NULL,
-  spec5 = NULL,
-  spec6 = NULL,
-  colour1 = "steelblue1",
-  colour2 = "lightpink1",
-  colour3 = "coral",
-  colour4 = "goldenrod1",
-  colour5 = "gold",
-  colour6 = "seagreen3",
+  spectra_cols,
+  colours = brewer.pal(length(spectra_cols), "Set2"),
   span = 5,
   thresh = 0.1,
   lbls = FALSE,
   lbl.fmt = "\%3.4f",
   min_mz = 0,
-  max_mz = 10000,
+  max_mz = max(dat[[mass_dat]]),
   min_I = 0,
-  max_I = NULL,
+  max_I = max(dat[spectra_cols]),
   x_ticks = 100,
+  nrows = round_any(length(spectra_cols), 2, f = ceiling)/2,
   intensity_scale = "free_y"
 )
 }
@@ -38,47 +29,16 @@ data in the first column.}
 \item{mass_dat}{A character string; the name of the column in \code{dat}
 containing the \emph{m/z} data.}
 
-\item{spec1}{A character string; the name of the column in \code{dat}
-containing the intensity data for the first spectrum that is to be plotted.}
+\item{spectra_cols}{A character string; the name(s) of the column(s) in \code{dat}
+containing the intensity data for the spectra-of-interest.}
 
-\item{spec2}{A character string; the name of the column in \code{dat}
-containing the intensity data for the second spectrum to be plotted.}
+\item{colours}{A character string indicating the desired colour(s)}
 
-\item{spec3}{A character string; the name of the column in \code{dat}
-containing the intensity data for the third spectrum to be plotted.}
-
-\item{spec4}{A character string; the name of the column in \code{dat}
-containing the intensity data for the fourth spectrum to be plotted.}
-
-\item{spec5}{A character string; the name of the column in \code{dat}
-containing the intensity data for the fifth spectrum to be plotted.}
-
-\item{spec6}{A character string; the name of the column in \code{dat}
-containing the intensity data for the sixth spectrum to be plotted.}
-
-\item{colour1}{A character string indicating the colour the first spectrum
-should be plotted in.}
-
-\item{colour2}{A charater string indicating the colour of the second
-spectrum in the plot.}
-
-\item{colour3}{A charater string indicating the colour of the third
-spectrum in the plot.}
-
-\item{colour4}{A charater string indicating the colour of the fourth
-spectrum in the plot.}
-
-\item{colour5}{A charater string indicating the colour of the fifth
-spectrum in the plot.}
-
-\item{colour6}{A charater string indicating the colour of the sixth
-spectrum in the plot.}
-
-\item{span}{Single numeric value; the span of peak maximas in between each
+\item{span}{Single numeric value; the span of peak maxima in between each
 label. Default = 5 (ignores two peak maxima on either side of each label).}
 
 \item{thresh}{Single numeric value (0-100); the threshold of peak
-itensities which should be labelled.}
+intensities which should be labeled.}
 
 \item{lbls}{Logical. If \code{lbls = TRUE}, labels indicating the
 \emph{m/z} value of each peak maxima within the indicated \code{span} will
@@ -102,7 +62,9 @@ range.}
[…]
