### 2021-02-03 beef2df
Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
Former-commit-id: cbb155912c94a4b7875ad0ea639ec3beb1d9db90
Former-commit-id: ea2c4cd7cdf0c70912385d5a43efcfac87949940
