+++ man/plotSpectra.Rd
@@ -7,6 +7,7 @@
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 \title{Plot and Compare Spectra}
 =======
 =======
@@ -46,10 +47,14 @@ plotSpectra(
 =======
 \title{Plot and Compare Multiple Spectra}
 >>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
+=======
+\title{Plot and Compare Spectra}
+>>>>>>> 3eeaddf... Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
 \usage{
 plotSpectra(
   dat,
   mass_dat,
+<<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
 =======
@@ -79,11 +84,16 @@ plotSpectra(
 >>>>>>> 3488b06... Added RColorBrewer palettes to plotSpectra() function
 =======
 >>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
+=======
+  spectra_cols,
+  colours = brewer.pal(length(spectra_cols), "Set2"),
+>>>>>>> 3eeaddf... Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
   span = 5,
   thresh = 0.1,
   lbls = FALSE,
   lbl.fmt = "\%3.4f",
   min_mz = 0,
+<<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
@@ -367,9 +377,13 @@ label. Default = 5 (ignores two peak maxima on either side of each label).}
 intensities which should be labeled.}
 =======
   max_mz = 10000,
+=======
+  max_mz = max(dat[[mass_dat]]),
+>>>>>>> 3eeaddf... Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
   min_I = 0,
-  max_I = NULL,
+  max_I = max(dat[spectra_cols]),
   x_ticks = 100,
+  nrows = round_any(length(spectra_cols), 2, f = ceiling)/2,
   intensity_scale = "free_y"
 )
 }
@@ -380,48 +394,21 @@ data in the first column.}
 \item{mass_dat}{A character string; the name of the column in \code{dat}
 containing the \emph{m/z} data.}
 
-\item{spec1}{A character string; the name of the column in \code{dat}
-containing the intensity data for the first spectrum that is to be plotted.}
+\item{spectra_cols}{A character string; the name(s) of the column(s) in \code{dat}
+containing the intensity data for the spectra-of-interest.}
 
-\item{spec2}{A character string; the name of the column in \code{dat}
-containing the intensity data for the second spectrum to be plotted.}
+\item{colours}{A character string indicating the desired colour(s)}
 
-\item{spec3}{A character string; the name of the column in \code{dat}
-containing the intensity data for the third spectrum to be plotted.}
-
-\item{spec4}{A character string; the name of the column in \code{dat}
-containing the intensity data for the fourth spectrum to be plotted.}
-
-\item{spec5}{A character string; the name of the column in \code{dat}
-containing the intensity data for the fifth spectrum to be plotted.}
-
-\item{spec6}{A character string; the name of the column in \code{dat}
-containing the intensity data for the sixth spectrum to be plotted.}
-
[…]
