+++ man/plotSpectra.Rd
@@ -10,6 +10,7 @@
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 \title{Plot and Compare Spectra}
 =======
 =======
@@ -71,10 +72,14 @@ plotSpectra(
 >>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
 =======
 \title{Plot and Compare Multiple Spectra}
+=======
+\title{Plot and Compare Spectra}
+>>>>>>> cd09eca... Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
 \usage{
 plotSpectra(
   dat,
   mass_dat,
+<<<<<<< HEAD
 >>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
   spec1,
   spec2 = NULL,
@@ -113,11 +118,16 @@ plotSpectra(
 >>>>>>> d58b4f6... Added RColorBrewer palettes to plotSpectra() function
 =======
 >>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
+=======
+  spectra_cols,
+  colours = brewer.pal(length(spectra_cols), "Set2"),
+>>>>>>> cd09eca... Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
   span = 5,
   thresh = 0.1,
   lbls = FALSE,
   lbl.fmt = "\%3.4f",
   min_mz = 0,
+<<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
@@ -659,9 +669,13 @@ Plot single or multiple spectra. For multiple spectra, a grid layout can be call
 >>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
 =======
   max_mz = 10000,
+=======
+  max_mz = max(dat[[mass_dat]]),
+>>>>>>> cd09eca... Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
   min_I = 0,
-  max_I = NULL,
+  max_I = max(dat[spectra_cols]),
   x_ticks = 100,
+  nrows = round_any(length(spectra_cols), 2, f = ceiling)/2,
   intensity_scale = "free_y"
 )
 }
@@ -672,47 +686,16 @@ data in the first column.}
 \item{mass_dat}{A character string; the name of the column in \code{dat}
 containing the \emph{m/z} data.}
 
-\item{spec1}{A character string; the name of the column in \code{dat}
-containing the intensity data for the first spectrum that is to be plotted.}
+\item{spectra_cols}{A character string; the name(s) of the column(s) in \code{dat}
+containing the intensity data for the spectra-of-interest.}
 
-\item{spec2}{A character string; the name of the column in \code{dat}
-containing the intensity data for the second spectrum to be plotted.}
+\item{colours}{A character string indicating the desired colour(s)}
 
-\item{spec3}{A character string; the name of the column in \code{dat}
-containing the intensity data for the third spectrum to be plotted.}
-
-\item{spec4}{A character string; the name of the column in \code{dat}
-containing the intensity data for the fourth spectrum to be plotted.}
-
-\item{spec5}{A character string; the name of the column in \code{dat}
-containing the intensity data for the fifth spectrum to be plotted.}
-
[…]
