### 2021-02-03 0dbd108
Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
Former-commit-id: d13d4832a28862f8bbc1dd93467a08291e7d7e9e
Former-commit-id: 5b21523bb6fcd074188716412fbfa4aedc1a9078
