### 2021-02-03 e9e1b91
Added RColorBrewer dependency
