### 2021-02-03 e5ddeb5
Added RColorBrewer dependency
