+++ DESCRIPTION
@@ -31,6 +31,7 @@ Description: subMALDI is an open framework tool that permits organization,
     sets are compatible with functions from a wide variety of other R packages, and 
     user-defined functions are easier to implement and test.
 <<<<<<< HEAD
+<<<<<<< HEAD
 Depends: R (>= 3.5.0), dplyr (>= 0.8.0), tidyr (>= 1.1.0), ggplot2 (>= 3.3.0), ggpmisc (>= 0.3.3), reshape2 (>= 1.4.4), RColorBrewer (>= 1.1-2)
 Imports: signal (>= 0.7-6), MassSpecWavelet (>= 1.48.0)
 License: GPL (>= 3) + file LICENSE
@@ -59,6 +60,9 @@ Imports: signal (>= 0.7-6), MassSpecWavelet (>= 1.54.0)
 =======
 Depends: R (>= 3.5.0), dplyr (>= 0.8.0), tidyr (>= 1.1.0), ggplot2 (>= 3.3.0), ggpmisc (>= 0.3.3), reshape2 (>= 1.4.4)
 >>>>>>> 5914622... added reshape2 dependency
+=======
+Depends: R (>= 3.5.0), dplyr (>= 0.8.0), tidyr (>= 1.1.0), ggplot2 (>= 3.3.0), ggpmisc (>= 0.3.3), reshape2 (>= 1.4.4), RColorBrewer (>= 1.1-2)
+>>>>>>> e8c3d8c... Added RColorBrewer dependency
 Imports: signal (>= 0.7-6), MassSpecWavelet (>= 1.48.0)
 >>>>>>> c823b44... Updated docs
 License: GPL (>= 3) + file LICENSE
