### 2021-02-03 0dc3d12
Added RColorBrewer palettes to plotSpectra() function
