### 2021-02-03 e88009d
Added RColorBrewer palettes to plotSpectra() function
