### 2021-02-03 76b1bb2
Added RColorBrewer palettes to plotSpectra() function
