### 2021-02-03 c5325f8
Added RColorBrewer palettes to plotSpectra() function
Former-commit-id: 0b27e789341b826e5967d9c7ed7376bfac2d97da
Former-commit-id: db9ad4f3fae499aa56d3f6f0b9c69ca66a79959a
