### 2021-02-03 2a8efb5
Added RColorBrewer palettes to plotSpectra() function
Former-commit-id: 3488b0680874079885ece05fb14d42c89f68d62d
Former-commit-id: f9e4b6570b55133bb5f9e395ffb71e3bb3ea5e04
