+++ R/plotSpectra.R
@@ -9,6 +9,7 @@
 ##' containing the \emph{m/z} data.
 ##' @param spectra_cols A character string; the name(s) of the column(s) in \code{dat}
 ##' containing the intensity data for the spectra-of-interest.
+##' @param palette A character element; the RColorBrewer palette to use. See below for available palettes. 
 ##' @param colours A character string indicating the desired colour(s)
 ##' @param span Single numeric value; the span of peak maxima in between each
 ##' label. Default = 5 (ignores two peak maxima on either side of each label).
@@ -37,7 +38,22 @@
 ##' "free_y"}, each spectrum will be plotted with its own intensity scale. If
 ##' \code{method = "fixed"}, each spectrum will be plotted with the y-axis of
 ##' the most intense spectrum in the set.
+##' 
 ##' @return Returns a line plot of the input spectra.
+##' 
+##' @section RColorBrewer Palettes: 
+##' \describe{\itemize{\item{\code{"Accent"}}
+##'                    \item{\code{"Dark2"}} 
+##'                    \item{\code{"Paired"}}  
+##'                    \item{\code{"Pastel1"}}  
+##'                    \item{\code{"Pastel2"}}  
+##'                    \item{\code{"Set1"}}  
+##'                    \item{\code{"Set2"} (default)} 
+##'                    \item{\code{"Set3"}}
+##'                    } 
+##'           }
+##' 
+##' 
 ##' @author Kristen Yeh <kristenyeh@@trentu.ca> Wesley Burr <wburr@@trentu.ca> Sophie Castel <sophie.castel@@ontariotechu.net>
 ##' @references https://github.com/wesleyburr/subMaldi
 ##' @keywords methods aplot
@@ -56,8 +72,9 @@
 # -----------------------------------------------------------------------
 
 plotSpectra <- function(dat, mass_dat,
-                        spectra_cols, 
-                        colours = brewer.pal(length(spectra_cols), "Set2"),
+                        spectra_cols,
+                        palette = "Set2",
+                        colours = brewer.pal(length(spectra_cols), palette),
                         span = 5,
                         thresh = 0.1,
                         lbls = FALSE, 
@@ -115,7 +132,9 @@ plotSpectra <- function(dat, mass_dat,
                 is.numeric(max_I), 
                 is.numeric(x_ticks),
                 is.numeric(span),
-                is.numeric(thresh))
+                is.numeric(thresh)),
+            palette %in% c("Accent", "Dark2", "Paired", "Pastel1", "Pastel2", "Set1", "Set2", "Set3")
+            
             )
   
   
