### 2021-02-03 b95a329
Added RColorBrewer palettes to plotSpectra() function
Former-commit-id: 0dc3d120470d0c2b19f412ae471497197d039ea8
Former-commit-id: 29d99b6fd7ac1077e34b0de95031d87ead0aaf0d
