+++ man/plotSpectra.Rd
@@ -48,8 +48,13 @@ plotSpectra(
 >>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 =======
   spectra_cols,
+<<<<<<< HEAD
   colours = brewer.pal(length(spectra_cols), "Set2"),
 >>>>>>> cd09eca... Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
+=======
+  palette = "Set2",
+  colours = brewer.pal(length(spectra_cols), palette),
+>>>>>>> 3488b06... Added RColorBrewer palettes to plotSpectra() function
   span = 5,
   thresh = 0.1,
   lbls = FALSE,
@@ -239,6 +244,8 @@ containing the \emph{m/z} data.}
 \item{spectra_cols}{A character string; the name(s) of the column(s) in \code{dat}
 containing the intensity data for the spectra-of-interest.}
 
+\item{palette}{A character string; the RColorBrewer palette to use. See below for available palettes.}
+
 \item{colours}{A character string indicating the desired colour(s)}
 
 \item{span}{Single numeric value; the span of peak maxima in between each
@@ -285,7 +292,24 @@ Returns a line plot of the input spectra.
 \description{
 Plot single or multiple spectra. For multiple spectra, a grid layout can be called using 'nrows'.
 }
+<<<<<<< HEAD
 >>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
+=======
+\section{RColorBrewer Palettes}{
+ 
+\describe{\itemize{\item{\code{"Accent"}}
+                   \item{\code{"Dark2"}} 
+                   \item{\code{"Paired"}}  
+                   \item{\code{"Pastel1"}}  
+                   \item{\code{"Pastel2"}}  
+                   \item{\code{"Set1"}}  
+                   \item{\code{"Set2"}} 
+                   \item{\code{"Set3"}}
+                   } 
+          }
+}
+
+>>>>>>> 3488b06... Added RColorBrewer palettes to plotSpectra() function
 \examples{
 
 ## Plotting using the sample dataset "Master.rda"
