### 2021-02-03 0b27e78
Added RColorBrewer palettes to plotSpectra() function
