+++ man/plotSpectra.Rd
@@ -76,6 +76,7 @@ plotSpectra(
 =======
   spectra_cols,
 <<<<<<< HEAD
+<<<<<<< HEAD
   colours = brewer.pal(length(spectra_cols), "Set2"),
 >>>>>>> cd09eca... Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
 =======
@@ -88,6 +89,10 @@ plotSpectra(
   spectra_cols,
   colours = brewer.pal(length(spectra_cols), "Set2"),
 >>>>>>> 3eeaddf... Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
+=======
+  palette = "Set2",
+  colours = brewer.pal(length(spectra_cols), palette),
+>>>>>>> d58b4f6... Added RColorBrewer palettes to plotSpectra() function
   span = 5,
   thresh = 0.1,
   lbls = FALSE,
@@ -397,6 +402,8 @@ containing the \emph{m/z} data.}
 \item{spectra_cols}{A character string; the name(s) of the column(s) in \code{dat}
 containing the intensity data for the spectra-of-interest.}
 
+\item{palette}{A character string; the RColorBrewer palette to use. See below for available palettes.}
+
 \item{colours}{A character string indicating the desired colour(s)}
 
 \item{span}{Single numeric value; the span of peak maxima in between each
@@ -461,8 +468,11 @@ Returns a line plot of the input spectra.
 Plot single or multiple spectra. For multiple spectra, a grid layout can be called using 'nrows'.
 }
 <<<<<<< HEAD
+<<<<<<< HEAD
 >>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 =======
+=======
+>>>>>>> d58b4f6... Added RColorBrewer palettes to plotSpectra() function
 \section{RColorBrewer Palettes}{
  
 \describe{\itemize{\item{\code{"Accent"}}
@@ -477,10 +487,13 @@ Plot single or multiple spectra. For multiple spectra, a grid layout can be call
           }
 }
 
+<<<<<<< HEAD
 >>>>>>> 3488b06... Added RColorBrewer palettes to plotSpectra() function
 =======
 
 >>>>>>> 2931ed6... Added pre-processing vignette
+=======
+>>>>>>> d58b4f6... Added RColorBrewer palettes to plotSpectra() function
 \examples{
 <<<<<<< HEAD
 
