### 2021-02-03 7c0c024
.normMethod_ function format deprecated ----> norm_ . Scripts and functions renamed accordingly.
