### 2021-02-03 5f1bb92
.normMethod_ function format deprecated ----> norm_ . Scripts and functions renamed accordingly.
