### 2021-02-03 fe730fc
.normMethod_ function format deprecated ----> norm_ . Scripts and functions renamed accordingly.
