### 2021-02-03 49efe10
.normMethod_ function format deprecated ----> norm_ . Scripts and functions renamed accordingly.
