+++ R/normSpectra.R
@@ -176,43 +176,43 @@ normSpectra <- function(dat, mass_dat, method = NULL, norm_mz = NULL, upper = NU
   
   
   if(method == "max"){ 
-    .normMethod_max(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) 
+    norm_max(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) 
   }
   
   if(method == "custom"){ 
-    .normMethod_custom(dat = dat, mass_dat = mass_dat, norm_mz = norm_mz, spectra_cols = spectra_cols, showHI = showHI) 
+    norm_custom(dat = dat, mass_dat = mass_dat, norm_mz = norm_mz, spectra_cols = spectra_cols, showHI = showHI) 
   } 
     
   if(method == "custom_imprecise"){ 
-    .normMethod_custimp(dat = dat, mass_dat = mass_dat, norm_mz = norm_mz, spectra_cols = spectra_cols, showHI = showHI) 
+    norm_custimp(dat = dat, mass_dat = mass_dat, norm_mz = norm_mz, spectra_cols = spectra_cols, showHI = showHI) 
   } 
     
   if(method == "max_set"){ 
-    .normMethod_max_set(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) 
+    norm_max_set(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) 
   }
     
   if( method == "TIC"){
-    .normMethod_TIC(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) 
+    norm_TIC(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) 
   } 
     
   if(method == "RMS"){
-    .normMethod_RMS(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) 
+    norm_RMS(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) 
   }
     
   if(method == "rel_TIC"){
-    .normMethod_rel_TIC(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) 
+    norm_rel_TIC(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) 
   }
     
   if(method == "median"){
-    .normMethod_median(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols)
+    norm_median(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols)
   }
       
   if(method == "stdev"){
-    .normMethod_stdev(dat = dat, mass_dat = mass_dat, lower = lower, upper = upper, spectra_cols = spectra_cols) 
+    norm_stdev(dat = dat, mass_dat = mass_dat, lower = lower, upper = upper, spectra_cols = spectra_cols) 
   }
     
   if(method == "quantile"){
-    .normMethod_quantile(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) 
+    norm_quantile(dat = dat, mass_dat = mass_dat, spectra_cols = spectra_cols) 
   }
 >>>>>>> f70720d... TIC norm added
 }  
