### 2021-02-03 9e86e3e
.normMethod_ function format deprecated ----> norm_ . Scripts and functions renamed accordingly.
Former-commit-id: b2e89cf70da4fa0606360256bd6df684a2485e9d
Former-commit-id: 51d8cb84bb609f0da2146a3ea11232e4d5f1699c
