### 2021-02-03 e97cf7f
.normMethod_ function format deprecated ----> norm_ . Scripts and functions renamed accordingly.
Former-commit-id: d4299269178fd052ef82abe2ab3b058bda0148dc
Former-commit-id: 1bdef147d66b09489787fdf28984e83d5fea1603
