### 2021-02-03 0cf68b3
.normMethod_ function format deprecated ----> norm_ . Scripts and functions renamed accordingly.
Former-commit-id: 7c0c024b4b03de01330f3a6e0215ff217c8d457a
Former-commit-id: 08fcab729c99858dac8c8af9b1e7ce561661e8df
