### 2021-02-03 b2e89cf
.normMethod_ function format deprecated ----> norm_ . Scripts and functions renamed accordingly.
