### 2021-02-03 d429926
.normMethod_ function format deprecated ----> norm_ . Scripts and functions renamed accordingly.
