### 2021-02-03 b99315a
.normMethod_ function format deprecated ----> norm_ . Scripts and functions renamed accordingly.
Former-commit-id: 49efe1050fe1d7caae126726d94b18a4172c47e7
Former-commit-id: be78793606a322dd19d16f532204c27b61a31f63
