### 2021-02-03 0560914
Encapsulated label-checking functions in if()
Former-commit-id: a9541765623661b8a96380d9db3774e638109e65
Former-commit-id: fa024761b69f90a05ecba84f8344125ed7d0f32e
