### 2021-02-03 8316aa0
Encapsulated label-checking functions in if()
Former-commit-id: 3da6910163edadd92e5bce4b51751fb89a5ce6b3
Former-commit-id: b9ed90fe3acd018b5082ddbb9133318aba85ffea
