+++ R/plotSpectra.R
@@ -631,34 +631,6 @@ plotSpectra <- function(dat, mass_dat,
                         nrows = round_any(length(spectra_cols), 2, f = ceiling)/2,
                         intensity_scale = "free_y"){   
   
-  
-  # Functions to check that labels have less than or equal to decimal places as data
-  
-  .deci <- function(x) {
-    if ((x %% 1) != 0) {
-      nchar(strsplit(sub('0+$', '', as.character(x)), ".", fixed=TRUE)[[1]][[2]])
-    } else {
-      return(0)
-    }
-  }
-  
-  .test_lbl <- function(mass_dat, lbl.fmt){
-    x <- mass_dat
-    dp <- c()
-    out <- c()
-    
-    lbl.dec <- strsplit(lbl.fmt, "[.]")[[1]][2]
-    lbl.dec <- as.numeric(gsub("[a-zA-Z ]", "", lbl.dec))
-    
-    for(i in 1:length(x)){
-      dp[i] <- .deci(x[i])
-    }
-    
-    dp <- max(dp)
-    out <- isTRUE(dp < lbl.dec) 
-    return(out)
-  }
-  
   # ----------------------------------
   # LOGICAL CHECKS
   # ----------------------------------
@@ -837,8 +809,11 @@ plotSpectra <- function(dat, mass_dat,
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 =======
 >>>>>>> 48ec512... Encapsulated label-checking functions in if()
+=======
+>>>>>>> 00fd778... Encapsulated label-checking functions in if()
     test_lbl <- function(mass_dat, lbl.fmt){
       x <- mass_dat
       dp <- c()
@@ -866,6 +841,7 @@ plotSpectra <- function(dat, mass_dat,
     
     
 <<<<<<< HEAD
+<<<<<<< HEAD
     # Check that labels have <= decimal places as data
     tst <- test_lbl(mass_dat = mass, lbl.fmt = lbl.fmt)
 =======
@@ -880,6 +856,10 @@ plotSpectra <- function(dat, mass_dat,
     # Check that labels have <= decimal places as data
     tst <- .test_lbl(mass_dat = mass, lbl.fmt = lbl.fmt)
 >>>>>>> cd09eca... Combined legacy plotting functions (plotSpectrum(), plotSpectra(), plotgridSpectra()) into a single function; plotSpectra(). Included parameter 'nrows' to allow user to construct a grid layout.
+=======
+    # Check that labels have <= decimal places as data
+    tst <- test_lbl(mass_dat = mass, lbl.fmt = lbl.fmt)
+>>>>>>> 00fd778... Encapsulated label-checking functions in if()
     
     if(tst){
       stop("Label format indicates more decimal places 
