### 2021-02-03 4c6b1f3
Encapsulated label-checking functions in if()
Former-commit-id: 455fe1677738cb0d6e860a8eeadef51e27b08a79
Former-commit-id: 0a3561a9eb98aa97d877019e7afe23a1d62fd873
