### 2021-02-03 e313b80
Encapsulated label-checking functions in if()
Former-commit-id: 5e86d1b75dce20a2e3a3b8a72ae14e2423856d6c
Former-commit-id: efb1d3273cbba3b7fd9b369cc12e6ad9f45a3822
