### 2021-02-03 de49d72
Created unique .R files for each baseline correction method
Former-commit-id: ca64545fe7135b740a40783c3e63498e6ca0a55b
Former-commit-id: 75c8f349d67ad666429014c6c7e07b601d501563
