+++ R/baselineCorr.R
@@ -76,9 +76,14 @@
 =======
 >>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 # ----------------------------------------------------------------------------
+<<<<<<< HEAD
 # Last Updated: July 29, 2020
 # Author: Kristen Yeh
 >>>>>>> 4bbf2ad... Baseline correction added
+=======
+# Last Updated: February 3, 2021
+# Author: Kristen Yeh, Sophie Castel
+>>>>>>> c919861... Created unique .R files for each baseline correction method
 # Title: subMALDI - Baseline Correction Parent Function
 # ----------------------------------------------------------------------------
 
@@ -87,6 +92,9 @@
 baselineCorr <- function(dat, mass_dat, intensity_dat, method = NULL, 
                            n = NULL){
 <<<<<<< HEAD
+<<<<<<< HEAD
+=======
+>>>>>>> c919861... Created unique .R files for each baseline correction method
   
   if(is.null(method)){
     stop('Please select a valid baseline correction method. See ?baselineCorr for list of methods.')
@@ -107,6 +115,7 @@ baselineCorr <- function(dat, mass_dat, intensity_dat, method = NULL,
   else if(method == "loess"){ 
     
     r <- base_loess(dat = dat, mass_dat = mass_dat, intensity_dat = intensity_dat) 
+<<<<<<< HEAD
     
   }
   
@@ -117,14 +126,16 @@ baselineCorr <- function(dat, mass_dat, intensity_dat, method = NULL,
   else { 
     if(method == "monotone_min"){ .base_mono(dat = dat, mass_dat = mass_dat, 
                                        intensity_dat = intensity_dat) }
+=======
+>>>>>>> c919861... Created unique .R files for each baseline correction method
     
-    else if(method == "linear"){ .base_linear(dat = dat, mass_dat = mass_dat, 
-                                             intensity_dat = intensity_dat,
-                                             n = n) }
-    else if(method == "loess"){.base_loess(dat = dat, mass_dat = mass_dat,
-                                           intensity_dat = intensity_dat) }
   }
+<<<<<<< HEAD
 >>>>>>> 4bbf2ad... Baseline correction added
+=======
+  
+  return(r)
+>>>>>>> c919861... Created unique .R files for each baseline correction method
 }
 
 
