+++ R/baselineCorr.R
@@ -65,8 +65,8 @@
 
 
 # ----------------------------------------------------------------------------
-# Last Updated: July 29, 2020
-# Author: Kristen Yeh
+# Last Updated: February 3, 2021
+# Author: Kristen Yeh, Sophie Castel
 # Title: subMALDI - Baseline Correction Parent Function
 # ----------------------------------------------------------------------------
 
@@ -74,18 +74,30 @@
 
 baselineCorr <- function(dat, mass_dat, intensity_dat, method = NULL, 
                            n = NULL){
-  if(is.null(method)){ stop('Please select a valid baseline correction method. 
-                            See ?baselineCorr for list of methods.') } 
-  else { 
-    if(method == "monotone_min"){ .base_mono(dat = dat, mass_dat = mass_dat, 
-                                       intensity_dat = intensity_dat) }
-    
-    else if(method == "linear"){ .base_linear(dat = dat, mass_dat = mass_dat, 
-                                             intensity_dat = intensity_dat,
-                                             n = n) }
-    else if(method == "loess"){.base_loess(dat = dat, mass_dat = mass_dat,
-                                           intensity_dat = intensity_dat) }
+  
+  if(is.null(method)){
+    stop('Please select a valid baseline correction method. See ?baselineCorr for list of methods.')
   }
+  
+  if(method == "monotone_min"){
+      
+    r <- base_mono(dat = dat, mass_dat = mass_dat, intensity_dat = intensity_dat) 
+    
+    }
+    
+  else if(method == "linear"){
+    
+    r <- base_linear(dat = dat, mass_dat = mass_dat, intensity_dat = intensity_dat, n = n) 
+    
+  }
+  
+  else if(method == "loess"){ 
+    
+    r <- base_loess(dat = dat, mass_dat = mass_dat, intensity_dat = intensity_dat) 
+    
+  }
+  
+  return(r)
 }
 
 
