### 2021-02-03 5b2b592
Created unique .R files for each baseline correction method
Former-commit-id: fed460b220886130ab171b4c48ecfdcc951ed45f
Former-commit-id: c9af62ad267c7695a65b6a2f8e1d107e5a65c564
