### 2021-02-03 cc04ee4
Created unique .R files for each baseline correction method
Former-commit-id: 3361762cd5af72979d3a39402f99b6b6dbd86709
Former-commit-id: f53776028eec32c62f2ed0c388f92cfe4975c758
