+++ R/baselineCorr.R
@@ -110,9 +110,14 @@
 >>>>>>> 7f2152d... Created unique .R files for each baseline correction method
 =======
 # ----------------------------------------------------------------------------
+<<<<<<< HEAD
 # Last Updated: July 29, 2020
 # Author: Kristen Yeh
 >>>>>>> 4bbf2ad... Baseline correction added
+=======
+# Last Updated: February 3, 2021
+# Author: Kristen Yeh, Sophie Castel
+>>>>>>> c919861... Created unique .R files for each baseline correction method
 # Title: subMALDI - Baseline Correction Parent Function
 # ----------------------------------------------------------------------------
 
@@ -125,10 +130,13 @@ baselineCorr <- function(dat, mass_dat, intensity_dat, method = NULL,
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 =======
 >>>>>>> c919861... Created unique .R files for each baseline correction method
 =======
 >>>>>>> 7f2152d... Created unique .R files for each baseline correction method
+=======
+>>>>>>> c919861... Created unique .R files for each baseline correction method
   
   if(is.null(method)){
     stop('Please select a valid baseline correction method. See ?baselineCorr for list of methods.')
@@ -149,6 +157,7 @@ baselineCorr <- function(dat, mass_dat, intensity_dat, method = NULL,
   else if(method == "loess"){ 
     
     r <- base_loess(dat = dat, mass_dat = mass_dat, intensity_dat = intensity_dat) 
+<<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
     
@@ -189,14 +198,16 @@ baselineCorr <- function(dat, mass_dat, intensity_dat, method = NULL,
   return(r)
 >>>>>>> 7f2152d... Created unique .R files for each baseline correction method
 =======
+=======
+>>>>>>> c919861... Created unique .R files for each baseline correction method
     
-    else if(method == "linear"){ .base_linear(dat = dat, mass_dat = mass_dat, 
-                                             intensity_dat = intensity_dat,
-                                             n = n) }
-    else if(method == "loess"){.base_loess(dat = dat, mass_dat = mass_dat,
-                                           intensity_dat = intensity_dat) }
   }
+<<<<<<< HEAD
 >>>>>>> 4bbf2ad... Baseline correction added
+=======
+  
+  return(r)
+>>>>>>> c919861... Created unique .R files for each baseline correction method
 }
 
 
