### 2021-02-03 38d8b0b
Removed lower/upper from stopifnot()
Former-commit-id: f052ebeda1279a476dde91a7c0893c54f8b2f46e
Former-commit-id: 761cc17adf7394556d07059838c1ed6b6a75441a
