### 2021-02-03 446ed26
Removed lower/upper from stopifnot()
Former-commit-id: a19317b69c15ac1c61ac7c93ab7475be4ead2d43
Former-commit-id: f07eb00f68954dae5cdddf657001b8e511880b65
