+++ R/norm_TIC.R
@@ -20,12 +20,16 @@ norm_TIC <- function(dat, mass_dat, spectra_cols){
     is.character(spectra_cols),
     mass_dat %in% colnames(dat),
 <<<<<<< HEAD
+<<<<<<< HEAD
     all(spectra_cols %in% colnames(dat))
 =======
     all(spectra_cols %in% colnames(dat)),
     is.numeric(lower),
     is.numeric(upper)
 >>>>>>> b2e89cf... .normMethod_ function format deprecated ----> norm_ . Scripts and functions renamed accordingly.
+=======
+    all(spectra_cols %in% colnames(dat))
+>>>>>>> 7b46ea1... Removed lower/upper from stopifnot()
   )    
   
   if(length(spectra_cols) < 2){
