### 2021-02-03 f0b96d3
Removed lower/upper from stopifnot()
