### 2021-02-03 8d03442
Removed lower/upper from stopifnot()
Former-commit-id: 981bd0f4b5a42195c14f6a2e66a83583d6e99fe1
Former-commit-id: 52fb6ed304cd0ca6dc48be18781bb41c921694d2
