+++ tests/testthat/test-baselineCorr.R
@@ -1,9 +1,13 @@
 <<<<<<< HEAD
+<<<<<<< HEAD
+=======
+>>>>>>> 0aa2619... JOSS 10 (Issue): Fixed formatting of test files
 # -----------------------------------------------------------------------
 # Last Updated: February 3, 2021
 # Author: Kristen Yeh, Sophie Castel
 # Title: testthat: baselineCorr()
 # -----------------------------------------------------------------------
+<<<<<<< HEAD
 
 data("bsline")
 
@@ -23,30 +27,31 @@ test_that("output minimum intensity is 0", {
 test_that("method = NULL yields error", {
   expect_error(baselineCorr(bsline, mass_dat = "mass", intensity_dat = "raw", method = NULL))
 =======
+=======
+>>>>>>> 0aa2619... JOSS 10 (Issue): Fixed formatting of test files
 
-
-context("Baseline Correction")
-library(subMALDI)
 data("bsline")
 
-loess <- baselineCorr(bsline, "mass", "raw","loess")
-loess <- select(loess, "baseline")
+# Generating baseline correction for all three methods 
 
-linear <- baselineCorr(bsline, "mass", "raw", "linear", 7)
-linear <- select(linear, "baseline")
+loess <- baselineCorr(bsline, mass_dat = "mass", intensity_dat = "raw", method = "loess")
+linear <- baselineCorr(bsline, mass_dat = "mass", intensity_dat = "raw", method = "linear", n = 7)
+mono <- baselineCorr(bsline, mass_dat = "mass", intensity_dat = "raw", method = "monotone_min")
 
-mono <- baselineCorr(bsline, "mass", "raw", "monotone_min")
-mono <- select(mono, "baseline")
 
 test_that("output minimum intensity is 0", {
-  expect_equal(min(loess), 0)
-  expect_equal(min(linear), 0)
-  expect_equal(min(mono), 0)
+  expect_equal(min(loess$baseline), 0)
+  expect_equal(min(linear$baseline), 0)
+  expect_equal(min(mono$baseline), 0)
 })
 
 test_that("method = NULL yields error", {
+<<<<<<< HEAD
   expect_error(baselineCorr(bsline, "mass", "raw"))
 >>>>>>> 40a45a4... Added tests
+=======
+  expect_error(baselineCorr(bsline, mass_dat = "mass", intensity_dat = "raw", method = NULL))
+>>>>>>> 0aa2619... JOSS 10 (Issue): Fixed formatting of test files
 })
 
 
@@ -55,9 +60,12 @@ test_that("method = NULL yields error", {
 # -----------------
 
 <<<<<<< HEAD
+<<<<<<< HEAD
 =======
 
 >>>>>>> 40a45a4... Added tests
+=======
+>>>>>>> 0aa2619... JOSS 10 (Issue): Fixed formatting of test files
 mz_test <- rbind(bsline, bsline[1,])
 mz_test <- mz_test[order("mass"),]
 
@@ -66,6 +74,9 @@ test_that("duplicated m/z stops fmethod = monotone_min", {
 })
 
 <<<<<<< HEAD
+<<<<<<< HEAD
+=======
+>>>>>>> 0aa2619... JOSS 10 (Issue): Fixed formatting of test files
 # -----------------
 # Linear
 # -----------------
@@ -76,5 +87,8 @@ test_that("duplicated m/z stops fmethod = monotone_min", {
 # -----------------
 # LOESS
 # -----------------
+<<<<<<< HEAD
 =======
 >>>>>>> 40a45a4... Added tests
+=======
+>>>>>>> 0aa2619... JOSS 10 (Issue): Fixed formatting of test files
