+++ dev.Rmd
@@ -71,6 +71,7 @@ save(Master2, file = "/home/sophie/subMaldi/data/compressed/Master2.rda", compre
 
 
 <<<<<<< HEAD
+<<<<<<< HEAD
 
 ## JOSS 5 (Closed)
 Wrote biocViews: into DESCRIPTION install_github() can search BioConductor for MassSpecWavelet
@@ -89,6 +90,9 @@ Wrote biocViews: into DESCRIPTION install_github() can search BioConductor for M
 =======
 ## JOSS 5 
 =======
+=======
+
+>>>>>>> 0aa2619... JOSS 10 (Issue): Fixed formatting of test files
 ## JOSS 5 (Closed)
 >>>>>>> 191e6b7... Rendered new documentation using package 'roxygen2'.
 Wrote biocViews: into DESCRIPTION install_github() can search BioConductor for MassSpecWavelet
@@ -100,14 +104,19 @@ Wrote biocViews: into DESCRIPTION install_github() can search BioConductor for M
 
 
 
+
 ## JOSS 12 - generalize and fix redundancies in code (Closed)
 
+<<<<<<< HEAD
 <<<<<<< HEAD
 # avgSpectra.R
 >>>>>>> f435d1f... Updated issue log
 =======
 # avgSpectra.R (done)
 >>>>>>> d7c9654... Updated issue log
+=======
+## avgSpectra.R (done)
+>>>>>>> 0aa2619... JOSS 10 (Issue): Fixed formatting of test files
 
 ```{r}
 avgSpectra <- function(dat, method = "mean", spectra_cols){
@@ -167,6 +176,7 @@ avgSpectra <- function(dat, method = "mean", spectra_cols){
 
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 ## find_max.R (done)
 =======
 # find_max.R
@@ -174,6 +184,9 @@ avgSpectra <- function(dat, method = "mean", spectra_cols){
 =======
 # find_max.R (done)
 >>>>>>> d7c9654... Updated issue log
+=======
+## find_max.R (done)
+>>>>>>> 0aa2619... JOSS 10 (Issue): Fixed formatting of test files
 
 ```{r}
 # -----------------
@@ -237,6 +250,7 @@ find_max_set <- function(dat, mass_dat, spectra_cols){
 }
 ```
 
+<<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
 ## norm_custimp.R (done)
@@ -373,13 +387,16 @@ norm_custimp <- function(dat, mass_dat, norm_mz, spectra_cols, showHI = FALSE){
 =======
 # normMethod_custimp.R (done)
 >>>>>>> 6db06a2... Updating issue log
+=======
+## norm_custimp.R (done)
+>>>>>>> 0aa2619... JOSS 10 (Issue): Fixed formatting of test files
 
 ```{r}
 # --------------------------------------
 # METHOD: CUSTOM M/Z VALUE, NOT PRECISE
 # --------------------------------------
 
-.normMethod_custimp <- function(dat, mass_dat, norm_mz, spectra_cols, showHI = FALSE){
+norm_custimp <- function(dat, mass_dat, norm_mz, spectra_cols, showHI = FALSE){
   
   # ---------------------
   # LOGICAL CHECKS
@@ -436,7 +453,7 @@ norm_custimp <- function(dat, mass_dat, norm_mz, spectra_cols, showHI = FALSE){
 
 ```
 
-# normMethod_custom.R (done)
[…]
