### 2021-02-03 188c2a2
JOSS 10 (Issue): Fixed formatting of test files
Former-commit-id: 3753a1b5139156bad51c99685917f4c2b853e2f1
Former-commit-id: 2d2126c4e3aa9687a3b1460626fdeb11b3c69f46
