+++ dev.Rmd
@@ -56,6 +56,7 @@ save(Master2, file = "/home/sophie/subMaldi/data/compressed/Master2.rda", compre
 
 
 
+
 ## JOSS 5 (Closed)
 Wrote biocViews: into DESCRIPTION install_github() can search BioConductor for MassSpecWavelet
 
@@ -66,9 +67,10 @@ Wrote biocViews: into DESCRIPTION install_github() can search BioConductor for M
 
 
 
+
 ## JOSS 12 - generalize and fix redundancies in code (Closed)
 
-# avgSpectra.R (done)
+## avgSpectra.R (done)
 
 ```{r}
 avgSpectra <- function(dat, method = "mean", spectra_cols){
@@ -126,7 +128,7 @@ avgSpectra <- function(dat, method = "mean", spectra_cols){
 
 ```
 
-# find_max.R (done)
+## find_max.R (done)
 
 ```{r}
 # -----------------
@@ -190,14 +192,14 @@ find_max_set <- function(dat, mass_dat, spectra_cols){
 }
 ```
 
-# normMethod_custimp.R (done)
+## norm_custimp.R (done)
 
 ```{r}
 # --------------------------------------
 # METHOD: CUSTOM M/Z VALUE, NOT PRECISE
 # --------------------------------------
 
-.normMethod_custimp <- function(dat, mass_dat, norm_mz, spectra_cols, showHI = FALSE){
+norm_custimp <- function(dat, mass_dat, norm_mz, spectra_cols, showHI = FALSE){
   
   # ---------------------
   # LOGICAL CHECKS
@@ -254,7 +256,7 @@ find_max_set <- function(dat, mass_dat, spectra_cols){
 
 ```
 
-# normMethod_custom.R (done)
+## norm_custom.R (done)
 
 ```{r}
 
@@ -318,7 +320,7 @@ find_max_set <- function(dat, mass_dat, spectra_cols){
 
 ```
 
-# normMethod_max.R (done)
+## norm_max.R (done)
 
 ```{r}
 # -----------------------
@@ -334,7 +336,7 @@ find_max_set <- function(dat, mass_dat, spectra_cols){
 # METHOD: MAX. OF EACH SPECTRUM
 # ------------------------------
 
-.normMethod_max <- function(dat, mass_dat, spectra_cols){
+norm_max <- function(dat, mass_dat, spectra_cols){
 
   # ---------------------
   # LOGICAL CHECKS
@@ -405,7 +407,7 @@ find_max_set <- function(dat, mass_dat, spectra_cols){
 
 ```
 
-# normMethod_median.R (done)
+## norm_median.R (done)
 
 ```{r}
 
@@ -413,7 +415,7 @@ find_max_set <- function(dat, mass_dat, spectra_cols){
 # METHOD: MEDIAN
 # ---------------
 
-.normMethod_med<- function(dat, mass_dat, spectra_cols){
+norm_med<- function(dat, mass_dat, spectra_cols){
 
   # ---------------------
   # LOGICAL CHECKS
@@ -469,7 +471,7 @@ find_max_set <- function(dat, mass_dat, spectra_cols){
 }
 ```
 
-# normMethod_quantile.R (done)
+## norm_quantile.R (done)
 
 ```{r}
 # --------------------------------------------------------------------------------------------
@@ -492,7 +494,7 @@ find_max_set <- function(dat, mass_dat, spectra_cols){
[…]
