+++ tests/testthat/test-baselineCorr.R
@@ -3,15 +3,19 @@
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 =======
 >>>>>>> 0aa2619... JOSS 10 (Issue): Fixed formatting of test files
 =======
 >>>>>>> 3753a1b... JOSS 10 (Issue): Fixed formatting of test files
+=======
+>>>>>>> 0aa2619... JOSS 10 (Issue): Fixed formatting of test files
 # -----------------------------------------------------------------------
 # Last Updated: February 3, 2021
 # Author: Kristen Yeh, Sophie Castel
 # Title: testthat: baselineCorr()
 # -----------------------------------------------------------------------
+<<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
 
@@ -85,30 +89,31 @@ test_that("method = NULL yields error", {
   expect_error(baselineCorr(bsline, mass_dat = "mass", intensity_dat = "raw", method = NULL))
 >>>>>>> 3753a1b... JOSS 10 (Issue): Fixed formatting of test files
 =======
+=======
+>>>>>>> 0aa2619... JOSS 10 (Issue): Fixed formatting of test files
 
-
-context("Baseline Correction")
-library(subMALDI)
 data("bsline")
 
-loess <- baselineCorr(bsline, "mass", "raw","loess")
-loess <- select(loess, "baseline")
+# Generating baseline correction for all three methods 
 
-linear <- baselineCorr(bsline, "mass", "raw", "linear", 7)
-linear <- select(linear, "baseline")
+loess <- baselineCorr(bsline, mass_dat = "mass", intensity_dat = "raw", method = "loess")
+linear <- baselineCorr(bsline, mass_dat = "mass", intensity_dat = "raw", method = "linear", n = 7)
+mono <- baselineCorr(bsline, mass_dat = "mass", intensity_dat = "raw", method = "monotone_min")
 
-mono <- baselineCorr(bsline, "mass", "raw", "monotone_min")
-mono <- select(mono, "baseline")
 
 test_that("output minimum intensity is 0", {
-  expect_equal(min(loess), 0)
-  expect_equal(min(linear), 0)
-  expect_equal(min(mono), 0)
+  expect_equal(min(loess$baseline), 0)
+  expect_equal(min(linear$baseline), 0)
+  expect_equal(min(mono$baseline), 0)
 })
 
 test_that("method = NULL yields error", {
+<<<<<<< HEAD
   expect_error(baselineCorr(bsline, "mass", "raw"))
 >>>>>>> 40a45a4... Added tests
+=======
+  expect_error(baselineCorr(bsline, mass_dat = "mass", intensity_dat = "raw", method = NULL))
+>>>>>>> 0aa2619... JOSS 10 (Issue): Fixed formatting of test files
 })
 
 
@@ -121,6 +126,7 @@ test_that("method = NULL yields error", {
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 =======
 
 >>>>>>> 40a45a4... Added tests
@@ -134,6 +140,8 @@ test_that("method = NULL yields error", {
 =======
 
 >>>>>>> 40a45a4... Added tests
+=======
+>>>>>>> 0aa2619... JOSS 10 (Issue): Fixed formatting of test files
 mz_test <- rbind(bsline, bsline[1,])
 mz_test <- mz_test[order("mass"),]
 
[…]
