### 2021-02-03 a401043
JOSS 10 (Issue): Fixed formatting of test files
Former-commit-id: 9c0c0f690a1ef18355cc2df7f38b3d1b16cc9085
Former-commit-id: cdd61714d66a76b4c046030de673ff1d077c6499
