+++ dev.Rmd
@@ -87,6 +87,7 @@ save(Master2, file = "/home/sophie/subMaldi/data/compressed/Master2.rda", compre
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 
 ## JOSS 5 (Closed)
 Wrote biocViews: into DESCRIPTION install_github() can search BioConductor for MassSpecWavelet
@@ -135,6 +136,9 @@ Wrote biocViews: into DESCRIPTION install_github() can search BioConductor for M
 =======
 ## JOSS 5 
 =======
+=======
+
+>>>>>>> 3753a1b... JOSS 10 (Issue): Fixed formatting of test files
 ## JOSS 5 (Closed)
 >>>>>>> ae03a50... Rendered new documentation using package 'roxygen2'.
 Wrote biocViews: into DESCRIPTION install_github() can search BioConductor for MassSpecWavelet
@@ -146,14 +150,19 @@ Wrote biocViews: into DESCRIPTION install_github() can search BioConductor for M
 
 
 
+
 ## JOSS 12 - generalize and fix redundancies in code (Closed)
 
+<<<<<<< HEAD
 <<<<<<< HEAD
 # avgSpectra.R
 >>>>>>> 46dcb76... Updated issue log
 =======
 # avgSpectra.R (done)
 >>>>>>> 1b35d86... Updated issue log
+=======
+## avgSpectra.R (done)
+>>>>>>> 3753a1b... JOSS 10 (Issue): Fixed formatting of test files
 
 ```{r}
 avgSpectra <- function(dat, method = "mean", spectra_cols){
@@ -216,6 +225,7 @@ avgSpectra <- function(dat, method = "mean", spectra_cols){
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 ## find_max.R (done)
 =======
 # find_max.R
@@ -232,6 +242,9 @@ avgSpectra <- function(dat, method = "mean", spectra_cols){
 =======
 # find_max.R (done)
 >>>>>>> 1b35d86... Updated issue log
+=======
+## find_max.R (done)
+>>>>>>> 3753a1b... JOSS 10 (Issue): Fixed formatting of test files
 
 ```{r}
 # -----------------
@@ -300,6 +313,9 @@ find_max_set <- function(dat, mass_dat, spectra_cols){
 <<<<<<< HEAD
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
+=======
+>>>>>>> 3753a1b... JOSS 10 (Issue): Fixed formatting of test files
 ## norm_custimp.R (done)
 
 ```{r}
@@ -429,6 +445,7 @@ norm_custimp <- function(dat, mass_dat, norm_mz, spectra_cols, showHI = FALSE){
 ```
 
 ## norm_max.R (done)
+<<<<<<< HEAD
 =======
 # normMethod_custimp.R
 =======
@@ -442,6 +459,24 @@ norm_custimp <- function(dat, mass_dat, norm_mz, spectra_cols, showHI = FALSE){
 # --------------------------------------
 # METHOD: CUSTOM M/Z VALUE, NOT PRECISE
 # --------------------------------------
+=======
+
+```{r}
+# -----------------------
+# NORMALIZATION FUNCTION
+# -----------------------
+
+# Function rescales intensity to 0,1
+.normalize <- function(x) {
+  return((x - min(x)) / (max(x) - min(x)))
+}
+
[…]
