### 2021-02-03 fe6b953
Fixed brewer.pal() bug; palette can now have less than three colours
Former-commit-id: 9a4af3299d2c8dc2724f29a0f5ace241975efa10
Former-commit-id: 18696ca23faa6b0c96fc654a472f629cdbd68bb8
