### 2021-02-03 9a7126e
Fixed brewer.pal() bug; palette can now have less than three colours
