+++ R/plotSpectra.R
@@ -91,6 +91,8 @@ plotSpectra <- function(dat, mass_dat,
   # LOGICAL CHECKS
   # ----------------------------------
   
+  colours <- colours[1:length(spectra_cols)]
+  
   
   stopifnot(spectra_cols %in% colnames(dat),
             mass_dat %in% colnames(dat),
