### 2021-02-03 ecbefe1
Fixed brewer.pal() bug; palette can now have less than three colours
