### 2021-02-15 #2694 @pibion
@wesleyburr @sigven @jspaezp apologies for my long silence - I didn't realize how much time I needed off over the holidays.

I'm finally back online and wanted to check in to see what the current status of the paper is.  @wesleyburr are you currently working on responding to comments?
