### 2021-03-06 eef4c76
Added unit tests and improved error-catching.
