+++ tests/testthat/test-peak_sub.R
@@ -12,28 +12,34 @@ library(subMALDI)
   # SNR
   ######
 
-  test_that("Spectrum Names must be specified", {
+  test_that("Dat input is dataframe", {
+    expect_error(
+      .peak_snr(dat = "apple")
+    )
+  })
+
+  test_that("Spectrum Names are specified", {
     data(Master2)
     expect_error(
       .peak_snr(dat = Master2, mass_dat = "full_mz")
     )
   })
   
-  test_that("Mass column must be contained in data frame", {
+  test_that("Mass column contained in data frame", {
     data(Master2)
     expect_error(
       .peak_snr(dat = Master2, mass_dat = "apple", intensity_dat = "Blank2")
     )
   })
   
-  test_that("Window covers entire spectrum", {
+  test_that("Window does not cover entire spectrum", {
     data(Master2)
     expect_error(
       .peak_snr(dat = Master2, mass_dat = "full_mz", intensity_dat = "Blank2", n = length(Master2$Blank2))
     )
   })
   
-  test_that("No signals detected below specified SNR", {
+  test_that("Signals detected below specified SNR", {
     data(Master2)
     expect_error(
       .peak_snr(dat = Master2, mass_dat = "full_mz", intensity_dat = "Blank2", n = 500, SNR_thresh = 0)
@@ -44,21 +50,27 @@ library(subMALDI)
   # SLOPE
   #######
   
-  test_that("Spectrum Names must be specified", {
+  test_that("Dat input is dataframe", {
+    expect_error(
+      .peak_snr(dat = "apple")
+    )
+  })
+  
+  test_that("Spectrum Names are specified", {
     data(Master2)
     expect_error(
       .peak_slope(dat = Master2, mass_dat = "full_mz")
     )
   })
   
-  test_that("Mass column must be contained in data frame", {
+  test_that("Mass column contained in data frame", {
     data(Master2)
     expect_error(
       .peak_slope(dat = Master2, mass_dat = "apple", intensity_dat = "Blank2")
     )
   })
   
-  test_that("Window covers entire spectrum", {
+  test_that("Window does not cover entire spectrum", {
     data(Master2)
     expect_error(
       .peak_slope(dat = Master2, mass_dat = "full_mz", intensity_dat = "Blank2", n = length(Master2$Blank2))
