### 2021-03-17 wesleyburr/subMaldi#5 @jspaezp
Hello @wesleyburr (and @castels), first of all, really nice job on the code-base! It definitely exceeded my expectations.

On the other hand I feel like I missed a couple of important points on the paper (which are still easily fixable and after which the paper should be good to go).

1. The second line of the paper resembles a little too closely the second paragraph of the introduction of "Mass Spectrometry, A textbook 2nd Edition" (Gross, 2011). I would appreciate a re-write of this section.
2. In several places of the paper there are references to the transparency of the data in the structure that subMaldi uses, I think another key point is that the data-frame structure allows interoperability with other packages, which I think should be highlighted. Since the tidy-dataframe has become the de-facto data structure in R (see https://vita.had.co.nz/papers/tidy-data.pdf)
3. There are several mentions of the fact that "MS generates irregularly spaced data", but no mention of the cause of this irregular spacing (spectra centroiding/mass accuracy/coalescence in the mass analyzer).
4. At the end of paragraph 1 , I believe a single citation is not enough to make the point that "an extensive catalog of software for analysis of MS data is available to researchers, at varying levels of accessibility", please elaborate on this matter (or move that point to the next paragraph, line 4 where a similar point is made).

Kindest wishes,
Sebastian
