### 2021-03-17 wesleyburr/subMaldi#5 @wesleyburr
Thanks for the feedback, we'll take care of these things ASAP. Your first review was very focused on the code, and we appreciate your kind words on our work. We were a bit ashamed, actually ... taking a good look at the code after all the time between made us realize what had been left in, and how poor some of it was! Glad you like the changes.
