### 2021-05-13 #2694 @tystan
Hi @pibion, @whedon and @wesleyburr. Thanks for having me on board! I can confirm I can edit the checklist. I'm on leave until mid next week so I'll get back to this then.
