### 2021-06-24 wesleyburr/subMaldi#22 @tystan
[JOSS REVIEW]: subMALDI - tystan
Just trying to create a checklist like @Jspaezp did. Please bear with me, I will delete this if it doesn't go to plan!

[Issue #2694](https://github.com/openjournals/joss-reviews/issues/2694)

---

    + yes, [here](https://github.com/wesleyburr/subMaldi)
    + yes, GNU GENERAL PUBLIC LICENSE Version 3
    + yes, @wesleyburr, @castels and Kristen Yeh have made most of the commits
    + yes, "JOSS’ minimum allowable contribution should represent not less than three months of work for an individual" would seem to have been met, plus some

    + Yes, no issues
   + [sample workflow online](https://wesleyburr.github.io/subMaldi/articles/subMALDI_workflow.html), [sample workflow vignette](https://github.com/wesleyburr/subMaldi/blob/main/vignettes/subMALDI_workflow.Rmd), [Processing with subMALDI online](https://wesleyburr.github.io/subMaldi/articles/subMALDIprocessing.html) and [Processing with subMALDI vignette](https://github.com/wesleyburr/subMaldi/blob/main/vignettes/subMALDIprocessing.Rmd) now checked - all OK for version 1.0.1.
    + I would suggest changing the default argument `min_mz = 0` in `plotSpectra()` to be changed to `min_mz = min(dat[spectra_cols])` to avoid the warning message, pasted below, whenever plotting is performed on spectra with m/z values greater than 0. (or alternatively, remove the warning message for `min_mz = 0`. Either way, it is not ideal getting warnings for default usage)
``Warning message:
In plotSpectra(norm_spec, mass_dat = "full_mz", spectra_cols = c("Before1",  :
  Specified value of 'min_mz' is beyond the mass range in 'dat'. Defaulting to the minimum mass.``
    + [line 377 of the sample workflow](https://github.com/wesleyburr/subMaldi/blob/bb01030bf161008465441cf3da4f87a27dda09fa/vignettes/subMALDI_workflow.Rmd#L377) produces the warning `2: Removed 116926 row(s) containing missing values (geom_path).` from ggplot2. Either no warning should be produced (trim the dataset before plotting?) or produce a more helpful warning from subMALDI
    + `testthat` unit tests exist for confirming the functional claims

    + Yes well presented in the README.md landing page
    + Yes, clear instructions. The only thing I would add is a hyperlink to the Releases page ([https://github.com/wesleyburr/subMaldi/releases](https://github.com/wesleyburr/subMaldi/releases)) to the sentence "To install this package in R, begin by downloading the source copy of the package from the releases folder"
    + Yes, the 2 vignettes cover that off nicely.
    + There are help (.Rd) files for the functions in the man/ directory
    + they contain argument, output and example code - looks great
    + Yes units tests in the `testthat/` directory. However the files `test-plotSpectra.R` and `test-readcsvDir.R` are empty. If they are methods that don't need/can't be unit tested, commented out notes explaining why seems sensible. Additionally, `test-baselineCorr.R` seems incomplete but I think the unit testing has been done in the other `test_base_*.R` tests.
    + However, running `R CMD check subMALDI_1.0-1.tar.gz` on the package produced warnings/errors such as: (1) `Packages in Depends field not imported from: 'RColorBrewer' 'reshape2' These packages need to be imported from (in the NAMESPACE file  for when this namespace is loaded but not attached.` (2) `Undocumented code objects: 'norm_RMS' 'norm_TIC' 'norm_custimp' 'norm_custom' 'norm_max' 'norm_max_set' 'norm_median' 'norm_quantile' 'norm_rel_TIC' 'norm_stdev'` (3) `* checking for code/documentation mismatches ... WARNING Functions or methods with usage in documentation object '.normMethod_RMS' but not in code: '.normMethod_RMS'` (4) `* checking examples ... ERROR Running examples in 'subMALDI-Ex.R' failed[...] + cleanEx() Error: unexpected symbol in:"cleanEx" Execution halted` etc
    + `1)` yes, link in last line of README.md `2)` I assume github issues is enough `3)` The help files contain the authors and emails

     + Yes explains what mass spec data are. There are proprietary software packages but this is open source, has functionality for the vast array of pre-processing steps and allows interrogation of the spectra at the (non-expert) user level
    + similar to above. The other R packages in MSnbase and MALDIquant are very different, and subMALDI definitely provides a simpler interface that non-R users would appreciate
    + Yes a good list of other proprietary (Bruker Daltonics) and open source packages in various languages (Python, R, Matlab, etc)
    + No issues, well written
    + References look great, only exception is the Gibb, S (2016) reference purely because the URL goes off the side of the page but that's a formatting thing not a content thing.
