### 2021-07-01 #2694 @tystan
Hi @wesleyburr. Making my way through the reviewer checklist. I have found issues with the [sample workflow](https://wesleyburr.github.io/subMaldi/articles/subMALDI_workflow.html) and the [Processing with subMALDI vignette](https://wesleyburr.github.io/subMaldi/articles/subMALDIprocessing.html) that I have mentioned under "Functionality" of the review [here](https://github.com/wesleyburr/subMaldi/issues/22#issue-929174183). If these pages can be fixed up, I will be able to assess the package's functionality better.

I couldn't add the screenshot I mentioned in the review note, so I am pasting it here to show you the resulting graph:
![graph-screenshot](https://user-images.githubusercontent.com/12855431/124116306-f9cb9c00-daad-11eb-8afb-1fe365464b31.png)
