### 2021-07-06 #2694 @wesleyburr
@tystan - sorry for delay, caught me over the long weekend. The issue appears to be that the pkgdown-generated website didn't update with all our changes post-review by J. Paez. So the vignettes built into the **package** appear to be fine (we can't see what you see), but the **pkgdown** versions on the github.wesleyburr.com/subMaldi/ are referencing a year+ old code. I've re-generated the pkgdown, and uploaded it, and you should hopefully be not seeing the old code (which won't work, because we changed the syntax of the function calls a lot, to clean up some poor design). Let me know if this doesn't clear it up, or I'm misunderstanding something about your message, please. :)

(I've pinned a new release, 1.0-1, which has an updated .tar.gz if you'd like to install locally and then try things. The vignettes compile and run fine against this version, and the pkgdown is updated)
