### 2021-07-08 #2694 @tystan
Great, thanks @wesleyburr. v1.0-1 installs and the online vignettes all work now (and I'm now using the in-package vignettes too, cheers for the heads up).
