### 2021-07-28 #2694 @tystan
Hi @wesleyburr . I've got some leave and am away camping at the moment sorry but will be back next week. In the meantime, there are some things that are listed to be addressed in my reviewing notes [here](https://github.com/wesleyburr/subMaldi/issues/22#issue-929174183).
