### 2021-08-16 #2694 @wesleyburr
Thanks, @tystan! Sophie and I will bang out your findings this week, along with the one outstanding thing from @jspaezp's review, and then hand it off to @pibion to setup final determination.
