### 2021-08-27 wesleyburr/subMaldi#22 @wesleyburr
Found all kinds of things that had slipped through - hadn't been religious about R CMD check as the refactor was happening. Thanks for flagging this - found some functionality issues and documentation issues that would have been missed otherwise. R CMD check passes with no warnings or errors now.

I think the fixes for you are done!
