### 2021-08-27 wesleyburr/subMaldi#5 @wesleyburr
Edits made - sorry for the extensive delay, I just waited until the second reviewer was fully done so I was working on this all at once. Please review the latest copy of the paper to see if it fixes your flagged requests.
