### 2021-08-27 wesleyburr/subMaldi#5 @jspaezp
Comparisson for reference:

https://github.com/wesleyburr/subMaldi/compare/eef4c76e43106ef9ba7275866bf7dc042941c8c1..97144d3aec6190ac6a3d384eac2ed5a175ba03eb

Looks good to me, marking as done.

Note on non-blocking changes:
some of the plots in the documentation dont show up/dont show whats intended

- https://wesleyburr.github.io/subMaldi/articles/subMALDI_workflow.html#plotting-multiple-spectra . missing images

- https://wesleyburr.github.io/subMaldi/articles/subMALDI_workflow.html#standard-deviation-of-noise-normalization
for instance (NAs get generated because of the range used in the example) . Not a big deal ... code is there and is reproducible ...
