### 2021-09-09 #2694 @tystan
R CMD check was a success except for the error pasted below which is minor (is `subMALDI::select` being loaded in the `NAMESPACE` or it it legacy code?).

No warnings when plotting in the vignette - now just prints a note to user (once in the `return()` statement and once to console using `cat()`?)

Have signed off and well done.

The only thing is the Gibb, S (2016) reference still has URL going off the side of the page

```
* checking examples ... ERROR
Running examples in 'subMALDI-Ex.R' failed
The error most likely occurred in:

> ### Name: rmveEmpty
> ### Title: Remove Empty Rows
> ### Aliases: rmveEmpty
> ### Keywords: manip methods
>
> ### ** Examples
>
>
> ## Load sample dataset "Mastre.rda"
> data("Master")
>
> ## Select only the spectra "Before1" and "Before2"
> ex <- select(Master, "full_mz", "Before1", "Before2")
Error in select(Master, "full_mz", "Before1", "Before2") :
  could not find function "select"
Execution halted
```
