### 2021-09-09 wesleyburr/subMaldi#22 @tystan
Hi @wesleyburr . Thanks for this. Have replied on the main review comment chain: [https://github.com/openjournals/joss-reviews/issues/2694#issuecomment-916103675](https://github.com/openjournals/joss-reviews/issues/2694#issuecomment-916103675)
