### 2021-09-09 #2694 @wesleyburr
Oh, I know what that error is - dplyr is only being loaded selectively now, to avoid the issue with signal::select and dplyr::select running into one another in the namespace. I'll fix that up, thanks for seeing it! I'm sure I ran the CMD check and had nothing.....

And I'll look at the URL and see if there's any way to fix it. Auto-formats are tricky.

Thanks, @tystan!
