### 2021-09-09 wesleyburr/subMaldi#5 @wesleyburr
I regenerated the pkgdown and pushed, and it ... fixed the missing images? NFC what was happening there ...

Thanks for your great attention to detail!
