### 2021-09-09 #2694 @wesleyburr
@pibion - I think we're done, at long last. Both issues with the reviews are finished, 20/20, and closed out. I've pushed a version (v1.0-4) which has all of the changes incorporated, and the final edited paper copy.
