### 2021-09-09 #2694 @pibion
@wesleyburr does the above error make sense to you?  @arfon the odd thing is this reference renders fine in the paper itself.
