### 2021-09-09 #2694 @pibion
@wesleyburr ideally I'd like to see the `@whedon check references` command pass and then ask you to tag the version and archive it (through e.g. zenodo).  I'll need the archive DOI to recommend acceptance.
