### 2021-09-09 #2694 @wesleyburr
@pibion - so looks like it was a malformed field in one reference, not at all the one that was cited. Plus all of the page **ranges** weren't in {} or "", so were propagating issues. Fixed them all up.
