### 2021-09-09 #2694 @wesleyburr
10.5281/zenodo.5498242 points to v1.0-5, which is the version with all of this fixed up.
